+++ paper/paper.bib
@@ -179,7 +179,7 @@ pages = "50 - 54",
 year = "2008",
 note = "Analytical biotechnology",
 issn = "0958-1669",
-doi = "https://doi.org/10.1016/j.copbio.2007.11.005",
+doi = "10.1016/j.copbio.2007.11.005",
 url = "http://www.sciencedirect.com/science/article/pii/S0958166907001474",
 author = "Thomas Werner"
 }
@@ -1287,7 +1287,7 @@ pages = "441 - 448",
 year = "1975",
 note = "",
 issn = "0022-2836",
-doi = "http://dx.doi.org/10.1016/0022-2836(75)90213-2",
+doi = "10.1016/0022-2836(75)90213-2",
 url = "http://www.sciencedirect.com/science/article/pii/0022283675902132",
 author = "F. Sanger and A.R. Coulson",
 
@@ -1451,7 +1451,7 @@ pages = "55 - 71",
 year = "2005",
 note = "",
 issn = "0378-8733",
-doi = "http://dx.doi.org/10.1016/j.socnet.2004.11.008",
+doi = "10.1016/j.socnet.2004.11.008",
 url = "http://www.sciencedirect.com/science/article/pii/S0378873304000693",
 author = "Stephen P. Borgatti",
 keywords = "Centrality",
@@ -1597,7 +1597,7 @@ pages = "861 - 874",
 year = "2006",
 note = "\{ROC\} Analysis in Pattern Recognition ",
 issn = "0167-8655",
-doi = "http://dx.doi.org/10.1016/j.patrec.2005.10.010",
+doi = "10.1016/j.patrec.2005.10.010",
 url = "http://www.sciencedirect.com/science/article/pii/S016786550500303X",
 author = "Tom Fawcett",
 keywords = "\{ROC\} analysis",
@@ -4579,7 +4579,7 @@ pages = "107 - 117",
 year = "1998",
 note = "",
 issn = "0169-7552",
-doi = "http://dx.doi.org/10.1016/S0169-7552(98)00110-X",
+doi = "10.1016/S0169-7552(98)00110-X",
 url = "http://www.sciencedirect.com/science/article/pii/S016975529800110X",
 author = "Sergey Brin and Lawrence Page",
 keywords = "World Wide Web",
@@ -4716,7 +4716,7 @@ keywords = "Google",
   year = {1999},
   note = {},
   issn = {0968-0004},
-  doi = {http://dx.doi.org/10.1016/S0968-0004(98)01343-7},
+  doi = {10.1016/S0968-0004(98)01343-7},
   url = {http://www.sciencedirect.com/science/article/pii/S0968000498013437},
   author = {Gerhard Christofori and Henrik Semb},
   keywords = {tumour suppressor},
@@ -4732,7 +4732,7 @@ keywords = "Google",
   year = {1999},
   note = {},
   issn = {1357-4310},
-  doi = {http://dx.doi.org/10.1016/S1357-4310(99)01461-6},
+  doi = {10.1016/S1357-4310(99)01461-6},
   url = {http://www.sciencedirect.com/science/article/pii/S1357431099014616},
   author = {Parry Guilford},
   keywords = {tumorigenesis},
@@ -4748,7 +4748,7 @@ keywords = "Google",
   year = {2000},
   note = {},
   issn = {0959-8049},
-  doi = {http://dx.doi.org/10.1016/S0959-8049(00)00158-1},
+  doi = {10.1016/S0959-8049(00)00158-1},
   url = {http://www.sciencedirect.com/science/article/pii/S0959804900001581},
   author = {I.R.G Beavon},
   keywords = {Epidermal growth factor receptor},
@@ -4764,7 +4764,7 @@ keywords = "Google",
[…]
