### 2020-03-04 #2137 @Kevin-Mattheus-Moerman
@TomKellyGenetics I think you need to use only the command message text and no other text.
