### 2020-03-04 #2137 @TomKellyGenetics
@Kevin-Mattheus-Moerman Thanks, I was thinking to try that. Wasn't sure how long it was meant to take. That's fixed the bibliography.
