### 2020-03-09 8a9accb
update documentation to meet CRAN requirements
