+++ tests/testthat/test-cpp.R
@@ -1,4 +1,0 @@
-context("C++")
-test_that("Catch unit tests pass", {
-    expect_cpp_tests_pass("graphsim")
-})

[…]
