+++ R/distance.R
@@ -12,6 +12,8 @@
 ##' @keywords graph network igraph adjacency
 ##' @import igraph
 ##' @examples 
+##' 
+##' library("igraph")
 ##' graph_test_edges <- rbind(c("A", "B"), c("B", "C"), c("B", "D"))
 ##' graph_test <- graph.edgelist(graph_test_edges, directed = TRUE)
 ##' adjacency_matrix <- make_adjmatrix_graph(graph_test)
