+++ man/make_sigma.Rd
@@ -41,6 +41,7 @@ or pre-computed matrix. These are compatible with \code{\link[mvtnorm]{rmvnorm}}
 }
 \examples{
 
+library("igraph")
 graph_test_edges <- rbind(c("A", "B"), c("B", "C"), c("B", "D"))
 graph_test <- graph.edgelist(graph_test_edges, directed = TRUE)
 adjacency_matrix <- make_adjmatrix_graph(graph_test)
