### 2020-03-11 #2137 @majensen
@corybrunson would you be able to review [this submission](#2137) to [JOSS](https://joss.theoj.org)? Thanks for considering!
