### 2020-03-11 #2137 @arfon
@TomKellyGenetics - you might want to truncate your bibtex entries to only those that are cited in the JOSS paper to prevent Whedon responses like this one: https://github.com/openjournals/joss-reviews/issues/2137#issuecomment-594413402
