### 2020-03-12 #2137 @majensen
> @majensen not this week, but i would be able to next week! If that's OK, then i'm glad to be assigned.

That would be excellent @corybrunson -- very much appreciated!
