### 2020-03-12 #2161 @whedon
[REVIEW]: graphsim: An R package for simulating gene expression data from graph structures of biological pathways
**Submitting author:** @TomKellyGenetics (<a href="http://orcid.org/0000-0003-3904-6690">S. Thomas Kelly</a>)
**Repository:** <a href="https://github.com/TomKellyGenetics/graphsim" target ="_blank">https://github.com/TomKellyGenetics/graphsim</a>
**Version:** v1.0.0-joss
