### 2020-03-12 #2161 @majensen
Replicating @arfon comment here regarding trimming bibtex entries:
> @TomKellyGenetics - you might want to truncate your bibtex entries to only those that are cited in the JOSS paper to prevent Whedon responses like this one: [#2137 (comment)](https://github.com/openjournals/joss-reviews/issues/2137#issuecomment-594413402)

https://github.com/openjournals/joss-reviews/issues/2137#issuecomment-597923101
