### 2020-03-12 #2161 @majensen
@TomKellyGenetics - in the first paragraph of *Methodology and software*, there is a sentence:
> ...of the popular package, which is often used...

Seems like the package name is missing here, or I'm missing something.
