### 2020-03-12 #2161 @TomKellyGenetics
This should build with updates from minor (formatting) corrections from errors found while submitting to CRAN and bioRxiv. I’ll look into removing unneeded references from the bibtex file.
