### 2020-03-16 #2161 @TomKellyGenetics
Dear Arfon,

Thank you for informing us on your policy. I'm currently based in Japan (which is considering to declare a state of emergency) so we're acutely aware of the developing situation here. I understand completely if reviews will take longer than usual during this time. I am safely able to work from home so can still respond to reviewers if they are able to.

Thank you,

Tom Kelly, Postdoctoral researcher at RIKEN IMS, Yokohama
