### 2020-03-20 #2161 @corybrunson
@TomKellyGenetics the submission looks pretty cool, as i'm just starting my review. I have some comments on the paper before i begin exploring the software. Sorry if you've already caught some of these but not yet updated the PDF.

* I think "Present address:" should be removed from the second affiliation (RIKEN).
* On my reading, the abstract-style Statement of Need encapsulates what graphsim does but doesn't really justify it. This is taken care of in the Summary, so could the Statement of Need be excised? I know JOSS guidelines require a statement of need, but in my experience this does not need to be a separate section. (@majensen this is something i've wondered before. Does JOSS prefer the SoN to be a section with header?)
* The paper could use a careful proofreading. For example, there is a stray curly brace after "R" in the Summary, and "released as the package" would read more clearly as "released as a package".
* In the first paragraph of "Methodology and software", it's not clear to me whether the simulation procedure you've implemented is of your own design or is already widely used but not yet conveniently packaged. If it is the same procedure as that of Wang &al (2018), then it would be good to say this at the start.
* Package names are not showing up in the paper.
* The section cross-reference at the end of "Generating a Simulated Expression Dataset" is broken.
