### 2020-03-20 TomKellyGenetics/graphsim#3 @corybrunson
edges not rendered using character argument to state parameter
I'm just running through the examples in the JOSS paper, and i've encountered a problem with the `plot_directed()` function. As illustrated below, while the function accepts the correctly-spelled character string (contrast to when the string is misspelled), it fails to render the edges; they are successfully rendered when using a numeric sign (`1` or `-1`) instead of a character.

I have igraph installed from CRAN, and i installed the current version of graphsim from GitHub.

``` r
library(igraph)
#>
#> Attaching package: 'igraph'
#> The following objects are masked from 'package:stats':
#>
#>     decompose, spectrum
#> The following object is masked from 'package:base':
#>
#>     union
library(graphsim)
graph_edges <- rbind(c("A", "C"), c("B", "C"), c("C", "D"), c("D", "E"),
                     c("D", "F"), c("F", "G"), c("F", "I"), c("H", "I"))
graph <- graph.edgelist(graph_edges, directed = TRUE)
# using the character string
plot_directed(graph, state = "activating", layout = layout.kamada.kawai,
              cex.node = 3, cex.arrow = 5, arrow_clip = 0.2)
```

![](https://i.imgur.com/36k01Ft.png)

``` r
# using a single sign
plot_directed(graph, state = 1, layout = layout.kamada.kawai,
              cex.node = 3, cex.arrow = 5, arrow_clip = 0.2)
```

![](https://i.imgur.com/vvvhseX.png)

``` r
# using a different character string
plot_directed(graph, state = "activate", layout = layout.kamada.kawai,
              cex.node = 3, cex.arrow = 5, arrow_clip = 0.2)
#> Warning in plot_directed(graph, state = "activate", layout =
#> layout.kamada.kawai, : please give state as a scalar or vector of
#> length(E(graph))
#> Error in plot_directed(graph, state = "activate", layout = layout.kamada.kawai, :
```

![](https://i.imgur.com/j95A32F.png)

_Part of [this JOSS review](https://github.com/openjournals/joss-reviews/issues/2161)._
