### 2020-03-20 #2161 @corybrunson
@TomKellyGenetics a couple of other comments, specifically on the code in the paper:

* The gplots package is used but not explicitly attached to the library.
* There's at least one typo: `trace = "none\"` toward the bottom of page 6.
* I'm unable to exactly reproduce the heatmaps displayed in the paper. They're based on the graph with inhibitions, and i'm also not clear on how to incorporate the vector `state` into these calculations, after looking into the code only a slight bit. (For issues running examples, etc. in the package itself, i'll open issues on the GitHub repo.)
