### 2020-03-20 #2161 @corybrunson
@TomKellyGenetics one last point on the paper:

* The second author has not directly contributed to the package, as reflected on GitHub. Could you briefly describe their contributions to the project in the paper?
