### 2020-03-23 3236c0d
correct arrow colours for character states https://github.com/TomKellyGenetics/graphsim/issues/3
