+++ DESCRIPTION
@@ -1,18 +1,13 @@
 Package: graphsim
 Type: Package
-Title: Simulate Expression Data from 'igraph' Networks 
-Version: 0.1.2
-Date: 2020-03-02
+Title: Simulate Expression Data from iGraph Networks 
+Version: 0.1.1.9001
+Date: 2019-07-15
 Authors@R: c(person("S. Thomas", "Kelly",email = "tom.kelly@postgrad.otago.ac.nz", role = c("ctb")),
              person("S. Thomas", "Kelly",email = "tom.kelly@riken.jp", role = c("aut", "cre")),
              person("Michael A.", "Black", email = "mik.black@otago.ac.nz", role = c("aut", "ths")))
-Description: Functions to develop simulated continuous data (e.g., gene expression) from a sigma covariance matrix derived from a graph structure in 'igraph' objects. Intended to extend 'mvtnorm' to take 'igraph' structures rather than sigma matrices as input.
-Language: en-GB
+Description: Functions to develop simulated continuous data (e.g., gene expression) from a sigma covariance matrix derived from a graph structure in iGraph. Intended to extend mvtnorm to take iGraph structures rather than sigma matrices as input.
 License: GPL-3
-URL: https://github.com/TomKellyGenetics/graphsim
-BugReports: https://github.com/TomKellyGenetics/graphsim/issues
-Depends:
-    R (>= 2.10.0)
 Imports:
     utils,
     igraph,
@@ -27,8 +22,7 @@ Suggests:
     testthat,
     gplots,
     scales,
-    devtools,
-    vdiffr
+    devtools
 LazyData: TRUE
 RoxygenNote: 7.0.2
 Encoding: UTF-8
