### 2020-03-23 TomKellyGenetics/graphsim#3 @TomKellyGenetics
I've managed to identify the problem. The state parameter is applied to all edges when give as a scalar (`length(state) == 1`). The issue was that default colours (`col.arrow`) were not defined in the case of text inputs (so they were rendered as white).

This should be fixed in the development version (activating edges default to `par("fg")` and inhibiting edges default to `"red"`). See commit: 3236c0d399796e09ee3583214ddc6753ec870285

```
devtools::install_github("TomKellyGenetics/graphsim", ref = "dev")
```

Working example with test input:

``` r
library("graphsim")
library("igraph")
#>
#> Attaching package: 'igraph'
#> The following objects are masked from 'package:stats':
#>
#>     decompose, spectrum
#> The following object is masked from 'package:base':
#>
#>     union
graph_edges <- rbind(c("A", "C"), c("B", "C"), c("C", "D"), c("D", "E"),
                     c("D", "F"), c("F", "G"), c("F", "I"), c("H", "I"))
graph <- graph.edgelist(graph_edges, directed = TRUE)
# using the character string
plot_directed(graph, state = "activating", layout = layout.kamada.kawai,
              cex.node = 3, cex.arrow = 5, arrow_clip = 0.2)
```

![](https://i.imgur.com/S1BnSd8.png)

``` r
library("graphsim")
library("igraph")
#>
#> Attaching package: 'igraph'
#> The following objects are masked from 'package:stats':
#>
#>     decompose, spectrum
#> The following object is masked from 'package:base':
#>
#>     union
graph_edges <- rbind(c("A", "C"), c("B", "C"), c("C", "D"), c("D", "E"),
                     c("D", "F"), c("F", "G"), c("F", "I"), c("H", "I"))
graph <- graph.edgelist(graph_edges, directed = TRUE)
# using the character string
plot_directed(graph, state = "inhibiting", layout = layout.kamada.kawai,
              cex.node = 3, cex.arrow = 5, arrow_clip = 0.2)
```

![](https://i.imgur.com/CuKFkKV.png)

Error handling when incorrect string given.

``` r
library("graphsim")
library("igraph")
#>
#> Attaching package: 'igraph'
#> The following objects are masked from 'package:stats':
#>
#>     decompose, spectrum
#> The following object is masked from 'package:base':
#>
#>     union
graph_edges <- rbind(c("A", "C"), c("B", "C"), c("C", "D"), c("D", "E"),
                     c("D", "F"), c("F", "G"), c("F", "I"), c("H", "I"))
graph <- graph.edgelist(graph_edges, directed = TRUE)
# using the character string
plot_directed(graph, state = "activate", layout = layout.kamada.kawai,
              cex.node = 3, cex.arrow = 5, arrow_clip = 0.2)
#> Warning in plot_directed(graph, state = "activate", layout =
#> layout.kamada.kawai, : please give state as a scalar or vector of
#> length(E(graph)): input must be 'activating', 'inhibiting' or an integer
#> Error in plot_directed(graph, state = "activate", layout = layout.kamada.kawai, :
```
I'll consider adding common aliases (e.g., "activate", "activation").

Vector input is also supported either directly calling the function with a value for each edge in `E(graph)` or as an edge "attribute" in the igraph object (using `E(g)$state <- states` as described [here](https://stackoverflow.com/questions/33813835/assigning-edge-weight-to-an-igraph-object-in-r)). Since this seems to be unclear, I'll review the documentations and consider adding more examples to demonstrate this.
