### 2020-03-23 #2161 @TomKellyGenetics
Thanks @corybrunson, I'll take a look at Issues as they come into the package repo (I think the [problem](https://github.com/TomKellyGenetics/graphsim/issues/3) with `plot_directed` is resolved).

I can reformat the Statement of Need if this is required explicitly by JOSS. I made it a separate section (as done by other recent submissions) so it would be clear for the editorial process.

Sorry it seems that some text formatting for language and package names weren't handled well when exporting LaTeX to whedon-compatible markdown. I've fixed these on the biorXiv version and I'll update the JOSS submission accordingly.

Wang _et al_ (2018) also simulate gene expression counts as a log-normal distribution. This package was originally developed before this publication. The approach to simulate based on graph structures is our original design. To my knowledge, there are no pre-existing implementations for this procedure in R.

The second author is my PhD supervisor. He was involved in planning, overseeing the project and preparing the manuscript. He has tested the software and given feedback on it.

I'll update the PDF once I've corrected the issues mentioned with the markdown document.
