### 2020-03-23 TomKellyGenetics/graphsim#3 @corybrunson
Great—i'll update to the development version, and it would be good to have these changes merged into `master` in order to finish the review.

By the way, i did not mean to suggest that aliases _should_ be allowed! I only meant to illustrate what happens when an unrecognized character string is passed, which was different than what happens when a recognized string was passed. I leave it to you to decide how flexible the interface should be. Thank you!
