### 2020-03-23 dea922e
fix small typos
