+++ paper/paper.md
@@ -21,7 +21,7 @@ affiliations:
    index: 1
  - name: "Present address:RIKEN Center for Integrative Medical Sciences, Suehiro-cho-1-7-22, Tsurumi Ward, Yokohama, Kanagawa 230-0045, Japan"
    index: 2
-date: 30 Janunary 2019
+date: 30 January 2019
 bibliography: paper.bib
 ---
 
@@ -82,7 +82,7 @@ genomic analysis. A number of methods have already been developed to
 utilise timecourse gene expression data [@Arner2015; @Yamaguchi2007]
 using gene regulatory modules in state-space models and recursive vector
 autoregressive models [@Hirose2008; @Shimamura2009]. Various approaches
-to gene regulation and networks at the genome-wide scale have lead to
+to gene regulation and networks at the genome-wide scale have led to
 novel biological insights [@Arner2015; @Komatsu2013]. However, inference
 of regulatory networks has thus far relied on experimental validation or
 resampling-based approaches to estimate the likelihood of specific
@@ -90,7 +90,7 @@ network modules being predicted [@Markowetz2007; @Hawe2019].
 
 There is a need, therefore, for a systematic framework for statistical
 modelling and simulation of gene expression data derived from
-hypothetical, inferred or known gene networks. Here we present an
+hypothetical, inferred or known gene networks. Here we present a
 package to achieve this, where samples from a multivariate normal
 distribution are used to generate normally-distributed log-expression
 data, with correlations between genes derived from the structure of the
@@ -183,7 +183,7 @@ parameter in multivariate normal simulations, particularly when negative
 correlations are included for inhibitions (as shown below). Matrices
 that could not be inverted occurred rarely with biologically plausible
 graph structures but this approach allows for the computation of a
-plausible correlation matrix when the graph structure given is
+plausible correlation matrix when the given graph structure is
 incomplete or contains loops. When required, the nearest positive
 definite matrix is computed using the `nearPD` function of the package
 [@Matrix] to perform Higham's algorithm [@Higham2002] on
@@ -267,7 +267,7 @@ are coloured white to red from $0$ to $1$. This $\Sigma$ matrix has been
 used to generate a simulated expression dataset of 100 samples (coloured
 blue to red from low to high) via sampling from the multivariate
 normal distribution. Here genes with closer relationships in the pathway
-structure show higher correlation between simulated values.](figure2.pdf){ width=95% #fig:simulation_activating label="simulation_activating"}
+structure show a higher correlation between simulated values.](figure2.pdf){ width=95% #fig:simulation_activating label="simulation_activating"}
 
 Generating a Simulated Expression Dataset {#sec:graphsim_demo}
 -----------------------------------------
