### 2020-03-23 560bac2
enable syntax highlighting for R
