### 2020-03-24 48271f6
add common aliases for state
