### 2020-03-24 5a75ca7
update documentation and enable edge attributes for make_state_matrix
