+++ R/plot_directed.R
@@ -15,19 +15,37 @@
 ##' @description Functions to plot_directed or graph structures including customised colours, layout, states, arrows. Uses graphs functions as an extension of \code{\link[igraph]{igraph}}. Designed for plotting directed graphs.
 ##'
 ##' @param graph An \code{\link[igraph]{igraph}} object. Must be directed with known states.
-##' @param state character or integer. Defaults to "activating" if no "state" edge attribute found. May be applied a scalar across all edges or as a vector for each edge respectively. Accepts non-integer values for weighted edges provided that the sign indicates whether links are activating (positive) or inhibiting (negative). May also be entered as text for "activating" or "inhibiting" or as integers for activating (0,1) or inhibiting (-1,2). Compatible with inputs for make_state_matrix or generate_expression_graph in the graphsim package \url{https://github.com/TomKellyGenetics/graphsim}.
-##' @param labels character vector. For labels to plot nodes. Defaults to vertex names in graph object. Entering "" would yield unlabelled nodes.
-##' @param layout function. Layout function as selected from \code{\link[igraph]{layout_}}. Defaults to layout.fruchterman.reingold. Alternatives include layout.kamada.kawai, layout.reingold.tilford, layout.sugiyama, and layout.davidson.harel. A 2-column layout matrix giving x and y co-ordinates of each node can be given.
+##' @param state character or integer. Defaults to "activating" if no "state" edge attribute 
+##' found. May be applied a scalar across all edges or as a vector for each edge respectively. 
+##' Accepts non-integer values for weighted edges provided that the sign indicates whether links
+##'  are activating (positive) or inhibiting (negative). May also be entered as text for 
+##'  "activating" or "inhibiting" or as integers for activating (0,1) or inhibiting (-1,2). 
+##'  Compatible with inputs for make_state_matrix or generate_expression_graph in the graphsim 
+##'  package \url{https://github.com/TomKellyGenetics/graphsim}. Vector input is supported 
+##' @param labels character vector. For labels to plot nodes. Defaults to vertex names in 
+##' graph object. Entering "" would yield unlabelled nodes.
+##' @param layout function. Layout function as selected from \code{\link[igraph]{layout_}}. 
+##' Defaults to layout.fruchterman.reingold. Alternatives include layout.kamada.kawai, 
+##' layout.reingold.tilford, layout.sugiyama, and layout.davidson.harel. A 2-column 
+##' layout matrix giving x and y co-ordinates of each node can be given.
 ##' @param cex.node numeric. Defaults to 1.
 ##' @param cex.label numeric. Defaults to 0.75.
 ##' @param cex.main numeric. Defaults to 0.8.
-##' @param cex.arrow numeric Defaults to 1.25. May take a scalar applied to all edges or a vector with values for each edge respectively.
-##' @param col.label character. Specfies the colours of node labels passed to plot. Defaults to par("fg").
+##' @param cex.arrow numeric Defaults to 1.25. May take a scalar applied to all edges 
+##' or a vector with values for each edge respectively.
+##' @param col.label character. Specfies the colours of node labels passed to plot. 
+##' Defaults to par("fg").
 ##' @param arrow_clip numeric Defaults to 0.075 (7.5\%).
-##' @param pch parameter passed to plot. Defaults to 21. Recommends using selecting between 21-25 to preserve colour behaviour. Otherwise entire node will inherit border.node as it's colour, in which case a light colour is recommended to see labels.
-##' @param border.node character. Specifies the colours of node border passed to plot. Defaults to grey33. Applies to whole node shape if pch has only one colour.
-##' @param fill.node character. Specfies the colours of node fill passed to plot. Defaults to grey66.
[…]
