+++ man/plot_directed.Rd
@@ -29,31 +29,49 @@ plot_directed(
 \arguments{
 \item{graph}{An \code{\link[igraph]{igraph}} object. Must be directed with known states.}
 
-\item{state}{character or integer. Defaults to "activating" if no "state" edge attribute found. May be applied a scalar across all edges or as a vector for each edge respectively. Accepts non-integer values for weighted edges provided that the sign indicates whether links are activating (positive) or inhibiting (negative). May also be entered as text for "activating" or "inhibiting" or as integers for activating (0,1) or inhibiting (-1,2). Compatible with inputs for make_state_matrix or generate_expression_graph in the graphsim package \url{https://github.com/TomKellyGenetics/graphsim}.}
+\item{state}{character or integer. Defaults to "activating" if no "state" edge attribute 
+found. May be applied a scalar across all edges or as a vector for each edge respectively. 
+Accepts non-integer values for weighted edges provided that the sign indicates whether links
+ are activating (positive) or inhibiting (negative). May also be entered as text for 
+ "activating" or "inhibiting" or as integers for activating (0,1) or inhibiting (-1,2). 
+ Compatible with inputs for make_state_matrix or generate_expression_graph in the graphsim 
+ package \url{https://github.com/TomKellyGenetics/graphsim}. Vector input is supported}
 
-\item{labels}{character vector. For labels to plot nodes. Defaults to vertex names in graph object. Entering "" would yield unlabelled nodes.}
+\item{labels}{character vector. For labels to plot nodes. Defaults to vertex names in 
+graph object. Entering "" would yield unlabelled nodes.}
 
-\item{layout}{function. Layout function as selected from \code{\link[igraph]{layout_}}. Defaults to layout.fruchterman.reingold. Alternatives include layout.kamada.kawai, layout.reingold.tilford, layout.sugiyama, and layout.davidson.harel. A 2-column layout matrix giving x and y co-ordinates of each node can be given.}
+\item{layout}{function. Layout function as selected from \code{\link[igraph]{layout_}}. 
+Defaults to layout.fruchterman.reingold. Alternatives include layout.kamada.kawai, 
+layout.reingold.tilford, layout.sugiyama, and layout.davidson.harel. A 2-column 
+layout matrix giving x and y co-ordinates of each node can be given.}
 
 \item{cex.node}{numeric. Defaults to 1.}
 
 \item{cex.label}{numeric. Defaults to 0.75.}
 
-\item{cex.arrow}{numeric Defaults to 1.25. May take a scalar applied to all edges or a vector with values for each edge respectively.}
+\item{cex.arrow}{numeric Defaults to 1.25. May take a scalar applied to all edges 
+or a vector with values for each edge respectively.}
 
 \item{cex.main}{numeric. Defaults to 0.8.}
 
 \item{arrow_clip}{numeric Defaults to 0.075 (7.5\%).}
 
-\item{pch}{parameter passed to plot. Defaults to 21. Recommends using selecting between 21-25 to preserve colour behaviour. Otherwise entire node will inherit border.node as it's colour, in which case a light colour is recommended to see labels.}
+\item{pch}{parameter passed to plot. Defaults to 21. Recommends using selecting 
+between 21-25 to preserve colour behaviour. Otherwise entire node will inherit 
+border.node as it's colour, in which case a light colour is recommended to see labels.}
 
-\item{border.node}{character. Specifies the colours of node border passed to plot. Defaults to grey33. Applies to whole node shape if pch has only one colour.}
+\item{border.node}{character. Specifies the colours of node border passed to plot.
+Defaults to grey33. Applies to whole node shape if pch has only one colour.}
 
[…]
