+++ DESCRIPTION
@@ -1,7 +1,7 @@
 Package: graphsim
 Type: Package
 Title: Simulate Expression Data from iGraph Networks 
-Version: 0.1.1.9001
+Version: 0.1.2.9001
 Date: 2019-07-15
 Authors@R: c(person("S. Thomas", "Kelly",email = "tom.kelly@postgrad.otago.ac.nz", role = c("ctb")),
              person("S. Thomas", "Kelly",email = "tom.kelly@riken.jp", role = c("aut", "cre")),
