### 2020-03-24 ed74141
update documentation and enable edge attributes for make_state_matrix https://github.com/TomKellyGenetics/graphsim/issues/3
