### 2020-03-24 488c156
update warning messages for plot_directed and make_state_matrix
