### 2020-03-24 TomKellyGenetics/graphsim#3 @TomKellyGenetics
Documentation has been updated for `plot_directed` and `make_state_matrix` (which both take a `state` parameter). The warning messages and aliases should make this more user-friendly.

I've merged this to master branch.
