### 2020-03-24 #2161 @majensen
@corybrunson - my opinion as editor (which will be addressed by @arfon if I'm wrong) is that the paper should be sure to include the rationale for writing the software and why there is a niche for it. Not necessary to have a heading (personally, I think that would be stylistically gauche-- but that's only my opinion!).
