### 2020-03-25 #2161 @corybrunson
@TomKellyGenetics no problem, i'll come back to the PDF after the next update.

Regarding the Statement of Need, i'll be fine with the current formatting (above the Summary with its own header) if you want to keep it; my concern is that it says only what the software does, not what need it satisfies. The first three paragraphs of the Introduction do this quite well, so i'd suggest encapsulating them into 2 or 3 sentences for that purpose.

Onward to the software itself. : )
