### 2020-03-25 TomKellyGenetics/graphsim#6 @corybrunson
additional documentation
Functions like `generate_expression()` and `plot_directed()` are documented with toy examples, while the four empirical networks included with the package are very sparsely documented. I think the value of the documentation in each case would be much increased by including illustrations that apply these functions to the empirical networks. Would this make sense to do, and would it be feasible? (Examples with long runtimes are discouraged by CRAN.)

_Part of [this JOSS review](https://github.com/openjournals/joss-reviews/issues/2161)._
