### 2020-03-25 TomKellyGenetics/graphsim#7 @corybrunson
igraph dependency
Is graphsim meant to be used (a) always in tandem with igraph or (b) sometimes in a standalone way? The examples all load igraph, but many of them still work if the occasional igraph functions are accessed using `igraph::`. If the package is not meant to be used without also loading igraph, then perhaps it makes sense for igraph to be listed under Depends rather than Imports (in the DESCRIPTION); or, if the package is meant to be used, at least sometimes, on its own, then perhaps some examples could do without loading igraph? Maybe the current design is intentional, and my review won't hinge on this, but i thought it worth asking.

(Also, a formatting point: the igraph package is usually formatting in all-lowercase, rather than "iGraph", as sometimes appears in the documentation.)

_Part of [this JOSS review](https://github.com/openjournals/joss-reviews/issues/2161)._
