### 2020-03-25 TomKellyGenetics/graphsim#8 @corybrunson
cor default
The `make_sigma` documentation says that the `cor` parameter defaults to 0.8, but currently there is no default:

``` r
library(igraph, warn.conflicts = FALSE)
library(graphsim)
graph_test_edges <- rbind(c("A", "B"), c("B", "C"), c("B", "D"))
graph_test <- graph.edgelist(graph_test_edges, directed = TRUE)
adjacency_matrix <- make_adjmatrix_graph(graph_test)
make_sigma_mat_adjmat(adjacency_matrix)
#> Error in ifelse(mat > 0, cor, 0): argument "cor" is missing, with no default
make_sigma_mat_adjmat(adjacency_matrix, cor = 0.8)
#>     A   B   C   D
#> A 1.0 0.8 0.0 0.0
#> B 0.8 1.0 0.8 0.8
#> C 0.0 0.8 1.0 0.0
#> D 0.0 0.8 0.0 1.0
```

Either the behavior or the documentation should be changed. : )

_Part of [this JOSS review](https://github.com/openjournals/joss-reviews/issues/2161)._
