### 2020-03-25 TomKellyGenetics/graphsim#9 @corybrunson
common link graph returned invisibly
Is it intended for the common link graph computed by `make_commonlink_graph()` to be returned invisibly, whereas other functions use `return()` (which ensures that the returned object is printed)? I'm flagging it in case the different behavior was accidental.

_Part of [this JOSS review](https://github.com/openjournals/joss-reviews/issues/2161)._
