### 2020-03-26 #2161 @TomKellyGenetics
@majensen @corybrunson Thanks for your advice, I agree. I thought a separate Statement of Need section was required by JOSS. I'd prefer to merge it into the Summary and avoid repetition so I'll do this as no one seems to object to this.
