### 2020-04-01 #2161 @corybrunson
@TomKellyGenetics what do you think about incorporating the summary statement of need you're preparing into the DESCRIPTION and/or README as well? This would resolve the criterion for the software documentation as well as for the paper.
