### 2020-04-01 TomKellyGenetics/graphsim#10 @corybrunson
contributing guidelines
The README does not indicate how users might contribute to the package, nor is there a CONTRIBUTING file. The latter is preferable, and there are several points of reference, for example [this contributing template](https://github.com/nayafia/contributing-template/blob/master/CONTRIBUTING-template.md) and the [ggplot2 contributing guidelines](https://github.com/tidyverse/ggplot2/blob/master/CONTRIBUTING.md) (from which i adapted my own for a ggplot2 extension). You may of course want to tailor it to the specific design, use, needs, etc. of your package!

_Part of [this JOSS review](https://github.com/openjournals/joss-reviews/issues/2161)._
