### 2020-04-02 #2161 @TomKellyGenetics
Hi @corybrunson, that sounds like a good idea. I'm planning update the manuscript and documentation in the next few days to make the the changes discussed so far.
