### 2020-04-02 5d5244e
update references to R packages on CRAN and GitHub https://github.com/TomKellyGenetics/graphsim/issues/5
