+++ paper/paper.md
@@ -48,7 +48,7 @@ pathway relationships and correlations. Here we present a versatile
 statistical framework to simulate correlated gene expression data from
 biological pathways, by sampling from a multivariate normal distribution
 derived from a graph structure. This procedure has been released as the
-`graphsim` R} package (\url{https://github.com/TomKellyGenetics/graphsim})
+\texttt{graphsim} R package on CRAN and GitHub (\url{https://github.com/TomKellyGenetics/graphsim})
  and is compatible with any graph structure that can be described using
 the `igraph` package.
 
@@ -120,14 +120,14 @@ $log(X_{ij}) \sim MVN({\bf\mu}, \Sigma)$, where ${\bf\mu}$ and $\Sigma$
 are the mean vector and variance-covariance matrix respectively, for
 gene expression data derived from a biological pathway) after
 appropriate normalisation [@Law2014; @Li2015]. Log-normality of gene
-expression matches the assumptions of the popular `limma` package [@limma], which is
+expression matches the assumptions of the popular \texttt{limma} package [@limma], which is
 often used for the analysis of intensity-based data from gene expression
 microarray studies and count-based data from RNA-Seq experiments. This
 approach has also been applied for modelling UMI-based count data from
 single-cell RNA-Seq experiments in the package [@Wang2018].
 
 In order to simulate transcriptomic data, a pathway is first constructed
-as a graph structure, using the R package [@igraph], with the status of
+as a graph structure, using the \texttt{igraph} R package [@igraph], with the status of
 the edge relationships defined (i.e, whether they activate or inhibit
 downstream pathway members). [This procedure uses]{style="color: black"}
 a graph structure such as that presented in
@@ -159,7 +159,7 @@ related by $\frac{1}{2}$ compared to the next nearest, as shown in
 Figure [2b](#fig:simulation_activating:second){reference-type="ref"
 reference="fig:simulation_activating:second"}. An
 arithmetically-decreasing (absolute) distance weighting is also
-supported in the package which implements this procedure: [ $$R_{ij} = 
+supported in the \texttt{graphsim} R package which implements this procedure: [ $$R_{ij} = 
 \begin{cases}
    1                            & \mbox{if genes } i \mbox{ and } j \mbox{ are adjacent} \\
    1-\frac{d_{ij}}{diam(G)}     & \mbox{if a path can be found between genes } i \mbox{ and } j \\
@@ -185,9 +185,9 @@ that could not be inverted occurred rarely with biologically plausible
 graph structures but this approach allows for the computation of a
 plausible correlation matrix when the given graph structure is
 incomplete or contains loops. When required, the nearest positive
-definite matrix is computed using the `nearPD` function of the package
+definite matrix is computed using the `nearPD` function of the \texttt{Matrix} R package
 [@Matrix] to perform Higham's algorithm [@Higham2002] on
-variance-covariance matrices. The package gives a warning when this
+variance-covariance matrices. The \texttt{graphsim} package gives a warning when this
 occurs.
 
 Illustrations {#sec:illustrations}
@@ -204,8 +204,8 @@ using the R package [@Genz2009; @mvtnorm]. The graph structure
 visualisation in
 Figure [1](#fig:simple_graph){reference-type="ref"
 reference="fig:simple_graph"} was specifically developed for (directed)
-iGraph objects in and is available in the and packages. The
-`plot_directed` function enables customisation of plot parameters for
+\texttt{igraph} objects in and is available in the and packages. The
+\texttt{plot_directed} function enables customisation of plot parameters for
 each node or edge, and mixed (directed) edge types for indicating
 activation or inhibition. These inhibition links (which occur frequently
 in biological pathways) are demonstrated in
[…]
