### 2020-04-02 TomKellyGenetics/graphsim#6 @TomKellyGenetics
I think this is a good idea if CRAN will allow it. The TGF-Beta/Smad pathway was selected for the manuscript because should be a manageable to size to run examples (although the runtime will be longer than the test examples).

Would a vignette using the biological networks be helpful? I have tried pre-generating figures for vignettes in HTML but CRAN has a limit on package size of 5MB which makes this difficult.

I'll review the examples in the documentation and see if it can pass CRAN checks.
