### 2020-04-02 TomKellyGenetics/graphsim#7 @TomKellyGenetics
Some functions running matrix operations should work without igraph loaded. The `library(igraph)` call in the examples is used for the `igraph` object class created by `graph_from_adjacency`. All igraph functions required by the functions in the package are imported but others are not.

I'm not sure whether listing igraph in "Depends" or "Imports" is better. Does this explanation help? If I understand correctly, we can remove the `library(igraph)` calls in the examples if the DESCRIPTION contains `Depends: igraph`.
