### 2020-04-02 8dba70f
update defaults for sigma to match generate https://github.com/TomKellyGenetics/graphsim/issues/8
