### 2020-04-02 TomKellyGenetics/graphsim#8 @TomKellyGenetics
Thanks for pointing this out. I've updated behaviour of the function to match the documentation and the `generate_expression` function which has the same parameter.

https://github.com/TomKellyGenetics/graphsim/blob/8dba70f20fc28360b157f58cb45ebe2db716b070/R/generate.R#L32
