### 2020-04-02 80e35f4
update return call for commonlink graph https://github.com/TomKellyGenetics/graphsim/issues/9
