+++ R/commonlink.R
@@ -38,4 +38,5 @@ make_commonlink_adjmat <- function(adj_mat){
 make_commonlink_graph <- function(graph, directed = FALSE){
   adj_mat <- make_adjmatrix_graph(graph, directed = directed)
   comm_mat <- make_commonlink_adjmat(adj_mat)
+  return(comm_mat)
 }
