### 2020-04-02 TomKellyGenetics/graphsim#9 @TomKellyGenetics
Thanks, I wasn't aware that this resulted in different behaviour. The function has been updated for consistency.
