### 2020-04-02 8ae71d4
add guide for contributions https://github.com/TomKellyGenetics/graphsim/issues/10
