### 2020-04-02 c1d9f28
refer to github issues for BugReports https://github.com/TomKellyGenetics/graphsim/issues/10
