### 2020-04-02 TomKellyGenetics/graphsim#10 @TomKellyGenetics
I've added [contributing guidelines](https://github.com/TomKellyGenetics/graphsim/blob/master/CONTRIBUTING.md), an [issue template](https://github.com/TomKellyGenetics/graphsim/blob/master/.github/ISSUE_TEMPLATE/ISSUE_TEMPLATE.md), and a [code of conduct](https://github.com/TomKellyGenetics/graphsim/blob/master/CODE_OF_CONDUCT.md). This follows the examples given for [ggplot2](https://github.com/igraph/rigraph/blob/dev/CONTRIBUTING.md) and [r-igraph](https://github.com/tidyverse/ggplot2/blob/master/CONTRIBUTING.md) and aims at contributors who may not be familiar with GitHub. I've also added the URL for the GitHub repo to the DESCRIPTION so it can be found from the CRAN release.
