### 2020-04-02 c067fd6
remove reference to thesis chapter
