### 2020-04-02 cbe48d3
add gplots::heatmap.2 as a dependency
