### 2020-04-02 742a5c3
add gplots::heatmap.2 as a dependency
