+++ NAMESPACE
@@ -16,6 +16,7 @@ import(graphics)
 import(igraph)
 import(mvtnorm)
 importFrom(Matrix,nearPD)
+importFrom(gplots,heatmap.2)
 importFrom(igraph,graph_from_adjacency_matrix)
 importFrom(igraph,is_igraph)
 importFrom(matrixcalc,is.positive.definite)
