### 2020-04-02 #2161 @corybrunson
@TomKellyGenetics i see your recent responses and will get to the repo issues next. I see only a few remaining issues with the paper:

* The "Statement of Need" header is still there, with no text under it.
* In the coded example, `graph_test4` is not defined, though it is used on lines 311 and 336 (in `paper.md`).
* In the "Computational details" section, i wonder if this sentence could be stated more clearly: "These functions can also be installed using the library on GitHub (https://github. com/TomKellyGenetics/igraph.extensions) which includes additional tools for igraph used." I see now that this package simply installs all the packages used by graphsim, but that makes me wonder whether it is necessary, since most package installation functions take the extra step of checking and installing dependencies. (Either way, this is less of an issue than a question.)
* "Vignettes" (in the same section) does not need to be capitalized.
