### 2020-04-02 #2161 @corybrunson
@TomKellyGenetics sorry, one more:

* You've explained the role of the second author here, but could you add a comment about this to the paper itself? A reader should, i think, be able to glean from the paper what contributions each author made. (As for the package itself, the second author's role is clear from the `ths` relator in the DESCRIPTION file.)
