### 2020-04-02 TomKellyGenetics/graphsim#6 @corybrunson
It would make a lot of sense to use these examples in a vignette, especially if there's no room to use them in the examples.

By the way, the existing vignettes rely on the animation package (among others), though the package itself does not; so it would be helpful to mention, with the installation instructions, that these packages must be installed in order to build the vignettes when installing graphsim:

```r
remotes::install_github("TomKellyGenetics/graphsim", build_vignettes = TRUE)
```
