### 2020-04-02 TomKellyGenetics/graphsim#11 @corybrunson
vignette formatting
Some of the vignettes are not properly formatted when installed. Most notably, section headers do not render due to a lack of space between `#`, `##`, or `###` and the header. Some, for example `test_graph_network_inhibiting`, include very little text and therefore function more as examples than as vignettes. Would it make sense to trim and/or combine the existing vignettes into a smaller number that are more comprehensive and contain thorough exposition?

_Part of [this JOSS review](https://github.com/openjournals/joss-reviews/issues/2161)._
