### 2020-04-02 TomKellyGenetics/graphsim#7 @corybrunson
Yes, that's true, since then loading graphsim would have the effect of loading igraph (and graphsim is automatically loaded before examples are run).

It sounds to me like the package is not intended as an extension of igraph, i.e. it is intended for at least some use cases that don't require additional igraph functionality. As such, it's appropriate to leave igraph under `Imports`.
