### 2020-04-02 TomKellyGenetics/graphsim#8 @corybrunson
Seems to work! Thanks.
