### 2020-04-02 TomKellyGenetics/graphsim#10 @corybrunson
This looks good! I'd only suggest that you adapt the issue template to reflect the GitHub setting and any particular preferences for your repo.
