### 2020-04-06 a20de3c
Changes to manuscript for JOSS review https://github.com/openjournals/joss-reviews/issues/2161
