### 2020-04-06 #2161 @TomKellyGenetics
Thanks @corybrunson, you've explained what is required very clearly. I've updated the manuscript as suggested.

The "igraph.extensions" library isn't so important at this stage, I've migrated all required functions into "graphsim" to satisfy CRAN requirements (no remote dependencies). It's for convenience to install other packages similar to the initial release of "tidyverse". I've updated the text to clarify this but I can remove it if it is not relevant (i.e., if JOSS submissions should only describe software on CRAN or similar repositories).
