### 2020-04-06 TomKellyGenetics/graphsim#11 @TomKellyGenetics
As discussed [here](https://github.com/TomKellyGenetics/graphsim/issues/6#issuecomment-607684815), I had some issues submitting to CRAN due to file size limitations. I agree that vignettes are important to guide users so I will review this and see if it possible to expand them within the limitations of CRAN. Examples in documentation (help menu) will be [updated](https://github.com/TomKellyGenetics/graphsim/issues/6) to match the vignettes.
