### 2020-04-06 TomKellyGenetics/graphsim#7 @TomKellyGenetics
Thanks for clarifying. In that case, I will leave `Imports: igraph` in the DESCRIPTION.

For examples in [documentation](https://github.com/TomKellyGenetics/graphsim/issues/6) and [vignettes](https://github.com/TomKellyGenetics/graphsim/issues/11) `library("igraph")` will still be needed to be called separately to use other `igraph` functions and the "igraph" object class.
