### 2020-04-06 TomKellyGenetics/graphsim#11 @corybrunson
That sounds good—though it's quite OK for the vignettes to complement the help examples; there's no need to have them consist of the same code, in case that's what you meant. I look forward to exploring the updated vignettes. Thanks.
