+++ man/generate_expression.Rd
@@ -12,7 +12,8 @@ generate_expression(
   mean = 0,
   comm = FALSE,
   dist = FALSE,
-  absolute = FALSE
+  absolute = FALSE,
+  laplacian = FALSE
 )
 }
 \arguments{
@@ -26,7 +27,7 @@ generate_expression(
 
 \item{mean}{mean value of each simulated gene. Defaults to 0. May be entered as a scalar applying to all genes or a vector with a separate value for each.}
 
-\item{comm, absolute}{logical. Parameters for Sigma matrix generation. Passed on to \code{\link[graphsim]{make_sigma_mat_dist_graph}} or \code{\link[graphsim]{make_sigma_mat_graph}}.}
+\item{comm, absolute, laplacian}{logical. Parameters for Sigma matrix generation. Passed on to \code{\link[graphsim]{make_sigma_mat_dist_graph}} or \code{\link[graphsim]{make_sigma_mat_graph}}.}
 
 \item{dist}{logical. Whether a graph distance (\code{\link[graphsim]{make_sigma_mat_dist_graph}}) or derived matrix (\code{\link[graphsim]{make_sigma_mat_graph}}) is used to compute the sigma matrix.}
 }

[…]
