### 2020-04-08 19ca1fa
compute distance and sigma based on Laplacian matrix #2
