### 2020-04-08 5a04c37
add matrix method for generate_expression and support Laplacian #2
