+++ R/generate.R
@@ -68,3 +68,55 @@ generate_expression <- function(n, graph, state = NULL, cor = 0.8, mean = 0, com
   colnames(expr_mat) <- paste0("sample_", 1:n)
   return(expr_mat)
 }
+
+
+##' @rdname generate_expression
+##' @export
+generate_expression_mat <- function(n, mat, state = NULL, cor = 0.8, mean = 0, comm = FALSE, dist = FALSE, absolute = FALSE, laplacian = FALSE){
+if(!is.integer(n)){
+  if(is.numeric(n)){
+    if(floor(n) == n){
+      n <- as.integer(n)
+    } else{
+      n <- floor(n)
+      print(paste("rounding to sample size", n))
+    }
+  } else{
+    stop("sample size n must be an integer of length 1")
+  }
+}
+if(length(n) > 1) stop("sample size n must be an integer of length 1")
+if(!is_igraph(graph)) stop("graph must be an igraph class")
+if(is.vector(state) || length(state) == 1) state <- make_state_matrix(graph, state)
+if(!(is.vector(mean)) || length(mean) == 1 ) mean <- rep(mean,length(V(graph)))
+if(dist){
+  sig <- make_sigma_mat_dist_adjmat(mat, cor, absolute = absolute)
+} else {
+  if(comm && laplacian){
+    warning("Error: only one of commonlink or laplacian can be used")
+    stop()
+  }
+  if(!comm && !laplacian) sig <- make_sigma_mat_adjmat(mat, cor)
+  if(comm){
+    mat <- make_commonlink_adjmat(mat)
+    sig <- make_sigma_mat_comm(mat, cor)
+  }
+  if(laplacian){
+    mat <- make_laplacian_adjmat(mat)
+    sig <- make_sigma_mat_laplacian(mat, cor)
+  }
+}
+if(!(is.null(state))) sig <- state * sig
+if(is.symmetric.matrix(sig) == FALSE) {
+  warning("sigma matrix was not positive definite, nearest approximation used.")
+  sig <- as.matrix(nearPD(sig, corr=T, keepDiag = TRUE)$mat) #postive definite correction
+}
+if(is.positive.definite(sig) == FALSE) {
+  warning("sigma matrix was not positive definite, nearest approximation used.")
+  sig <- as.matrix(nearPD(sig, corr=T, keepDiag = TRUE)$mat) #postive definite correction
+}
+expr_mat <- t(rmvnorm(n,mean=mean, sigma=sig))
+rownames(expr_mat) <- names(V(graph))
+colnames(expr_mat) <- paste0("sample_", 1:n)
+return(expr_mat)
+}
