+++ man/make_laplacian_graph.Rd
@@ -1,32 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/laplacian.R
-\name{make_laplacian}
-\alias{make_laplacian}
-\alias{make_laplacian_graph}
-\title{Generate Laplacian Matrix}
-\usage{
-make_laplacian_graph(graph, directed = FALSE)
-}
-\arguments{
-\item{graph}{An \code{\link[igraph]{igraph}} object. May be directed or weighted.}
-
-\item{directed}{logical. Whether directed information is passed to the Laplacian matrix.}
-}
-\value{
-An Laplacian matrix compatible with generating an expression matrix
-}
-\description{
-Compute the Laplacian matrix of a (directed) \code{\link[igraph]{igraph}} structure, preserving node/column/row names (and direction).
-}
-\examples{
-
-library("igraph")
-graph_test_edges <- rbind(c("A", "B"), c("B", "C"), c("B", "D"))
-graph_test <- graph.edgelist(graph_test_edges, directed = TRUE)
-laplacian_matrix <- make_laplacian_graph(graph_test)
-
-}
-\keyword{Laplacian}
-\keyword{graph}
-\keyword{igraph}
-\keyword{network}
