### 2020-04-15 TomKellyGenetics/graphsim#13 @rcannood
[JOSS Review] 
Sorry for the delay in my review. Below are several concerns regarding the article and the documentation (related to openjournals/joss-reviews#2161).

- [x] **Refine [introduction](https://github.com/TomKellyGenetics/graphsim#simulate-expression-data-from-igraph-networks) in README**: Better highlight the statement of need. In the article this was clearly stated in the first paragraph. In the readme, not so much.
- [x] **[Installation instructions](https://github.com/TomKellyGenetics/graphsim#installation)**: There's no need to manually install the dependency packages; `install.packages("graphsim")` will install the dependencies for you.
- [x] **Documentation**: Some of the vignettes (e.g. test_graph_divergent.Rmd) do not contain enough documentation / description of the plots being produced to understand what is going on.
- [x] **State of the field**: Please compare against similar methods, including [dyngen](https://www.biorxiv.org/content/10.1101/2020.02.06.936971v2), [SERGIO](https://www.biorxiv.org/content/10.1101/716811v1.abstract), [GeneNetWeaver](https://www.ncbi.nlm.nih.gov/pubmed/21697125).
- [x] **Code in manuscript**: Remove installation instructions or move it to a separate section. Since graphsim is on cran, the installation instructions can just be `install.packages("graphsim")`. You don't need to install devtools or igraph.
- [x] **Code in manuscript**: Fix syntax highlighting issue at the bottom of page 6 due to `trace = "none\",`.
- [x] **Code in manuscript**: Given that there is a one-to-one relation between the figures generated and the code blocks, would it be desirable to use an Rmd instead of an md file? This part is up for discussion. Example of Rmd JOSS paper: https://github.com/tidyverse/tidyverse/blob/master/vignettes/paper.Rmd
