### 2020-04-22 a0ef4ab
correct passing character states to matrix (plot pass tests)
