+++ R/state.R
@@ -41,6 +41,7 @@ make_state_matrix <- function(graph, state = NULL){
   state[state == 2] <- -1
   state[state == 1] <- 1
   if(is.character(state)){
+    state <- as.list(state)
     state[grep("activating", state)] <- 1
     state[grep("activation", state)] <- 1
     state[grep("activate", state)] <- 1
@@ -54,7 +55,7 @@ make_state_matrix <- function(graph, state = NULL){
     if(is.character(state)){
       warning("Please give state as a scalar or vector of length(E(graph)): input must be 'activating', 'inhibiting' or an integer")
     }
-    state <- as.numeric(state)
+    state <- unlist(as.numeric(state))
   }
   if(!all(state %in% -1:2)){
     state <- sign(state) # coerce to vector or 1 and -1 if not already
