### 2020-04-22 7c67ba8
correct passing character states to matrix (plot pass tests)
