### 2020-04-22 e1120ef
remove Catch C++ testing
