### 2020-04-22 922791f
update dependencies (Appveyor WARNING)
