### 2020-04-22 8b53316
document matrix methods for Laplacian
