### 2020-04-22 TomKellyGenetics/graphsim#10 @TomKellyGenetics
GitHub the Issue Template with this notice. I'll review this and see if I can update it to be compatible the [new version](https://help.github.com/en/github/building-a-strong-community/about-issue-and-pull-request-templates) of GitHub templates.

![Screen Shot 2020-04-22 at 19 15 14](https://user-images.githubusercontent.com/5493325/79970304-ad5df700-84cd-11ea-86e3-d62d481a2d74.png)
