+++ ISSUE_TEMPLATE.md
@@ -1,19 +1,0 @@
-Please briefly describe your problem and what output you expect. If you have a question,
-please check that it has not already been addressed but a previous issue or
-asked on stackoverflow <http://stackoverflow.com>.
-
-Please include a minimal reprex. The goal of a reprex is to make it as easy as possible
-for me to recreate your problem so that I can fix it. If you've never heard of a reprex
-before, start by reading <https://www.tidyverse.org/help/#reprex>, and follow the advice
-further down the page. Do NOT include session info unless it's explicitly asked for,
-or you've used `reprex::reprex(..., si = TRUE)` to hide it away.  
-  
-  Delete these instructions once you have read them.
-
----
-  
-  Brief description of the problem
-
-```r
-# insert reprex here
-```
