### 2020-04-23 e1a3f67
add passing state parameter to sigma
