### 2020-04-23 bbcde8e
document state for sigma
