+++ man/make_sigma.Rd
@@ -10,11 +10,17 @@
 \usage{
 make_sigma_mat_adjmat(mat, cor = 0.8)
 
-make_sigma_mat_graph(graph, cor = 0.8, comm = FALSE, directed = FALSE)
+make_sigma_mat_graph(
+  graph,
+  state = NULL,
+  cor = 0.8,
+  comm = FALSE,
+  directed = FALSE
+)
 
-make_sigma_mat_dist_adjmat(mat, cor = 0.8, absolute = FALSE)
+make_sigma_mat_dist_adjmat(mat, state = NULL, cor = 0.8, absolute = FALSE)
 
-make_sigma_mat_dist_graph(graph, cor = 0.8, absolute = FALSE)
+make_sigma_mat_dist_graph(graph, state = NULL, cor = 0.8, absolute = FALSE)
 }
 \arguments{
 \item{mat}{precomputed adjacency, commonlink, or scaled distance matrix.}
@@ -22,6 +28,8 @@ make_sigma_mat_dist_graph(graph, cor = 0.8, absolute = FALSE)
 \item{cor}{numeric. Simulated maximum correlation/covariance of two adjacent nodes. Default to 0.8.}
 
 \item{graph}{An \code{\link[igraph]{igraph}} object. May be directed or weighted.}
+
+\item{state}{numeric vector. Vector of length E(graph). Sign used to calculate state matrix, may be an integer state or inferred directly from expected correlations for each edge. May be applied a scalar across all edges or as a vector for each edge respectively. May also be entered as text for "activating" or "inhibiting" or as integers for activating (0,1) or inhibiting (-1,2). Compatible with inputs for \code{\link[graphsim]{plot_directed}}. Also takes a pre-computed state matrix from \code{\link[graphsim]{make_state_matrix}} if applied to the same graph multiple times.}
 
 \item{comm}{logical whether a common link matrix is used to compute sigma. Defaults to FALSE (adjacency matrix).}
 
