+++ NAMESPACE
@@ -6,6 +6,8 @@ export(make_commonlink_adjmat)
 export(make_commonlink_graph)
 export(make_distance_adjmat)
 export(make_distance_graph)
+export(make_laplacian_adjmat)
+export(make_laplacian_graph)
 export(make_sigma_mat_adjmat)
 export(make_sigma_mat_dist_adjmat)
 export(make_sigma_mat_dist_graph)
@@ -15,9 +17,14 @@ export(plot_directed)
 import(graphics)
 import(igraph)
 import(mvtnorm)
+importClassesFrom(Matrix,dgCMatrix)
+importFrom(Matrix,Matrix)
 importFrom(Matrix,nearPD)
 importFrom(gplots,heatmap.2)
+importFrom(igraph,as.undirected)
+importFrom(igraph,graph.edgelist)
 importFrom(igraph,graph_from_adjacency_matrix)
 importFrom(igraph,is_igraph)
+importFrom(igraph,laplacian_matrix)
 importFrom(matrixcalc,is.positive.definite)
 importFrom(matrixcalc,is.symmetric.matrix)
