### 2020-04-23 502766b
check correct output from state_matrix inhibiting
