### 2020-04-23 25d0c1d
compute paths for each subgraph
