### 2020-04-23 1f9ce21
refactor and check for cycles
