### 2020-04-23 e05480c
skip resolution for activating matrix
