### 2020-04-24 e52f6dd
resolve states if not in same path
