### 2020-04-24 ec42209
correct passing undirected graphs from sigma to resolve states
