+++ R/sigma.R
@@ -65,7 +65,8 @@ make_sigma_mat_dist_adjmat <- function(mat, state = NULL, cor = 0.8, absolute = 
   if(!is.null(state)){
     #adjacency matrix from distance
     adjmat <- ifelse(mat == max(mat[mat != 1]), 1, 0)
-    graph <- graph_from_adjacency_matrix(adjmat)
+    graph <- graph_from_adjacency_matrix(adjmat, mode = "undirected")
+    graph <- as.directed(graph, mode = "arbitrary")
     #pass state parameters sign of sigma
     state_mat <- make_state_matrix(graph, state)
     sig <- sig * sign(state_mat)
