### 2020-04-24 935cf6a
correct passing states with directed graphs
