### 2020-04-24 a28af84
remove installation of dependencies from README #13
