### 2020-04-24 TomKellyGenetics/graphsim#13 @TomKellyGenetics
Thanks for the review, sorry for taking a while to get back on this. I'll start making corrections to address these.

As also mentioned by @corybrunson [here](https://github.com/openjournals/joss-reviews/issues/2161#issuecomment-601921101) there is an issue reproducing the inhibition plots. I've made some adjustments to the package so that hardcoded variables are not required for this. This is the first step to make sure the code is reproducible. I'll update paper.md consider switching to Rmarkdown format. Figure 3 will need to be updated to account for these changes.

Here is a preview of the code for the first 3 figures. This runs in the [current version](935cf6a]) pushed to master.

```
######## Section 1 ####################

#install packages required (once per machine)
install.packages("graphsim")

#load required packages (once per R instance)
library("igraph")
library("graphsim")
library("gplots")
#generate graph structure
graph_edges <- rbind(c("A", "C"), c("B", "C"), c("C", "D"), c("D", "E"),
                     c("D", "F"), c("F", "G"), c("F", "I"), c("H", "I"))
graph <- graph.edgelist(graph_edges, directed = TRUE)
#plot graph structure (Figure 1)
plot_directed(graph, state ="activating", layout = layout.kamada.kawai,
              cex.node=3, cex.arrow=5, arrow_clip = 0.2)
#generate parameters for inhibitions
state <- c(1, 1, -1, 1, 1, 1, 1, -1, 1)
#plot graph structure with inhibitions (Figure 2)
plot_directed(graph, state=state, layout = layout.kamada.kawai,
              cex.node=3, cex.arrow=5, arrow_clip = 0.2)

######## Section 2 ####################

# activating graph
state <- rep(length(E(graph)), 1)
plot_directed(graph, state=state, layout = layout.kamada.kawai,
              cex.node=3, cex.arrow=5, arrow_clip = 0.2)

#adjacency matrix
adj_mat <- make_adjmatrix_graph(graph)
#relationship matrix
dist_mat <- make_distance_graph(graph, absolute = FALSE)
#sigma matrix directly from graph
sigma_mat <- make_sigma_mat_dist_graph(graph, 0.8, absolute = FALSE)
#show shortest paths of graph
shortest_paths <- shortest.paths(graph)
#generate expression data directly from graph
expr <- generate_expression(100, graph, cor = 0.8, mean = 0, comm = FALSE,
                            dist = TRUE, absolute = FALSE, state = state)
#plot adjacency matrix
heatmap.2(make_adjmatrix_graph(graph), scale = "none", trace = "none",
          col = colorpanel(3, "grey75", "white", "blue"),
          colsep = 1:length(V(graph)), rowsep = 1:length(V(graph)))
#plot relationship matrix
heatmap.2(make_distance_graph(graph, absolute = FALSE),
          scale = "none", trace = "none", col = colorpanel(50, "white", "red"),
colsep = 1:length(V(graph)), rowsep = 1:length(V(graph)))
#plot sigma matrix
heatmap.2(make_sigma_mat_dist_graph(graph, 0.8, absolute = FALSE),
scale = "none", trace = "none", col = colorpanel(50, "white", "red"),
colsep = 1:length(V(graph)), rowsep = 1:length(V(graph)))
expr <- generate_expression(100, graph, cor = 0.8, mean = 0,
comm = FALSE, dist =TRUE, absolute = FALSE, state = state)
#plot simulated expression data
heatmap.2(expr, scale = "none", trace = "none", col = bluered(50),
colsep = 1:length(V(graph)), rowsep = 1:length(V(graph)))
#plot simulated correlations
heatmap.2(cor(t(expr)), scale = "none", trace = "none", col = colorpanel(50, "white", "red"),
colsep = 1:length(V(graph)), rowsep = 1:length(V(graph)))

######## Section 3 ####################

#generate parameters for inhibitions
state <- c(1, 1, -1, 1, 1, 1, 1, -1, 1)
plot_directed(graph, state=state, layout = layout.kamada.kawai,
              cex.node=3, cex.arrow=5, arrow_clip = 0.2)

#adjacency matrix
adj_mat <- make_adjmatrix_graph(graph)
#relationship matrix
dist_mat <- make_distance_graph(graph, absolute = FALSE)
#state matrix
state_mat <- make_state_matrix(graph, state)
#sigma matrix directly from graph
sigma_mat <- make_sigma_mat_dist_graph(graph, state = state, 0.8, absolute = FALSE)
#show shortest paths of graph
shortest_paths <- shortest.paths(graph)
#generate expression data directly from graph
expr <- generate_expression(100, graph, cor = 0.8, mean = 0, comm = FALSE,
                            dist = TRUE, absolute = FALSE, state = state)
#plot adjacency matrix
heatmap.2(make_adjmatrix_graph(graph), scale = "none", trace = "none",
          col = colorpanel(3, "grey75", "white", "blue"),
          colsep = 1:length(V(graph)), rowsep = 1:length(V(graph)))
#plot state matrix
heatmap.2(make_state_matrix(graph, state),
          scale = "none", trace = "none", col = colorpanel(50, "white", "red"),
          colsep = 1:length(V(graph)), rowsep = 1:length(V(graph)))
#plot relationship matrix
heatmap.2(make_distance_graph(graph, absolute = FALSE),
          scale = "none", trace = "none", col = colorpanel(50, "white", "red"),
          colsep = 1:length(V(graph)), rowsep = 1:length(V(graph)))
#plot sigma matrix
heatmap.2(make_sigma_mat_dist_graph(graph, state = state, 0.8, absolute = FALSE),
          scale = "none", trace = "none", col = colorpanel(50, "blue", "white", "red"),
          colsep = 1:length(V(graph)), rowsep = 1:length(V(graph)))
expr <- generate_expression(100, graph, state = state, cor = 0.8, mean = 0,
                            comm = FALSE, dist =TRUE, absolute = FALSE)
#plot simulated expression data
heatmap.2(expr, scale = "none", trace = "none", col = bluered(50),
          colsep = 1:length(V(graph)), rowsep = 1:length(V(graph)))
#plot simulated correlations
heatmap.2(cor(t(expr)), scale = "none", trace = "none", col = colorpanel(50, "blue", "white", "red"),
          colsep = 1:length(V(graph)), rowsep = 1:length(V(graph)))
```

I think these results better represent the input graph. Various updates to [`make_state_matrix`](https://github.com/TomKellyGenetics/graphsim/blob/master/R/state.R) allow for disconnected graphs and cycles. Ideally the input graph is a directed acyclic tree but  a Sigma matrix can be generated if the graph object is not so clean and tidy (as happens in biology).

## Figure 3
### Inhibiting graph structure
![Rplot](https://user-images.githubusercontent.com/5493325/80210044-05336400-866e-11ea-8ff1-aefcc480e844.png)
### Sigma
![sigma](https://user-images.githubusercontent.com/5493325/80209600-5131d900-866d-11ea-9666-72ae963c37c6.png)
### Correlations
![cor](https://user-images.githubusercontent.com/5493325/80209583-4e36e880-866d-11ea-826c-9ce3d24534ad.png)
### Simulated Expression
![expr](https://user-images.githubusercontent.com/5493325/80209597-50994280-866d-11ea-8f0a-dd1a5cd9f6a9.png)

This is a work in progress. I'll update the unit tests and vignettes to reflect these changes before revising the manuscript.
