### 2020-04-24 TomKellyGenetics/graphsim#13 @TomKellyGenetics
Issue with trace is fixed with commit cbe48d326c0cd82fdc19.
