### 2020-04-27 f836198
allow passing state and commonlinks to sigma
