### 2020-04-27 30fe76a
example and tests for generate matrix
