+++ R/generate.R
@@ -26,9 +26,11 @@
 ##' library("igraph")
 ##' graph_test_edges <- rbind(c("A", "B"), c("B", "C"), c("B", "D"))
 ##' graph_test <- graph.edgelist(graph_test_edges, directed = TRUE)
-##' adjacency_matrix <- make_adjmatrix_graph(graph_test)
 ##' n <- 100
 ##' generate_expression(n, graph_test, cor = 0.8)
+##' 
+##' adjacency_matrix <- make_adjmatrix_graph(graph_test)
+##' generate_expression_mat(n, adjacency_matrix, cor = 0.8)
 ##' 
 ##' @return numeric matrix of simulated data (log-normalised counts)
 ##' 
@@ -75,50 +77,51 @@ generate_expression <- function(n, graph, state = NULL, cor = 0.8, mean = 0, com
 ##' @rdname generate_expression
 ##' @export
 generate_expression_mat <- function(n, mat, state = NULL, cor = 0.8, mean = 0, comm = FALSE, dist = FALSE, absolute = FALSE, laplacian = FALSE){
-if(!is.integer(n)){
-  if(is.numeric(n)){
-    if(floor(n) == n){
-      n <- as.integer(n)
+  if(!is.integer(n)){
+    if(is.numeric(n)){
+      if(floor(n) == n){
+        n <- as.integer(n)
+      } else{
+        n <- floor(n)
+        print(paste("rounding to sample size", n))
+      }
     } else{
-      n <- floor(n)
-      print(paste("rounding to sample size", n))
+      stop("sample size n must be an integer of length 1")
     }
-  } else{
-    stop("sample size n must be an integer of length 1")
   }
-}
-if(length(n) > 1) stop("sample size n must be an integer of length 1")
-if(!is_igraph(graph)) stop("graph must be an igraph class")
-if(is.vector(state) || length(state) == 1) state <- make_state_matrix(graph, state)
-if(!(is.vector(mean)) || length(mean) == 1 ) mean <- rep(mean,length(V(graph)))
-if(dist){
-  sig <- make_sigma_mat_dist_adjmat(mat, cor, absolute = absolute)
-} else {
-  if(comm && laplacian){
-    warning("Error: only one of commonlink or laplacian can be used")
-    stop()
+  if(length(n) > 1) stop("sample size n must be an integer of length 1")
+  if(!is.matrix(mat)) stop("graph must be an igraph class")
+  if(is.vector(state) || length(state) == 1) state <- make_state_matrix(graph, state)
+  if(!(is.vector(mean)) || length(mean) == 1 ) mean <- rep(mean,ncol(mat))
+  if(dist){
+    distmat <- make_distance_adjmat(mat)
+    sig <- make_sigma_mat_dist_adjmat(distmat, cor, absolute = absolute)
+  } else {
+    if(comm && laplacian){
+      warning("Error: only one of commonlink or laplacian can be used")
+      stop()
+    }
+    if(!comm && !laplacian) sig <- make_sigma_mat_adjmat(mat, cor)
+    if(comm){
+      mat <- make_commonlink_adjmat(mat)
+      sig <- make_sigma_mat_comm(mat, cor)
+    }
+    if(laplacian){
+      mat <- make_laplacian_adjmat(mat)
+      sig <- make_sigma_mat_laplacian(mat, cor)
+    }
   }
-  if(!comm && !laplacian) sig <- make_sigma_mat_adjmat(mat, cor)
-  if(comm){
-    mat <- make_commonlink_adjmat(mat)
-    sig <- make_sigma_mat_comm(mat, cor)
+  if(!(is.null(state))) sig <- state * sig
[…]
