### 2020-04-27 5c12fd1
document laplacian
