### 2020-04-29 #2161 @corybrunson
@majensen thanks for checking. There are a few issues that remain open on the repo, and @TomKellyGenetics has been making some revisions to the paper. (They may wish to comment.) I expect to be done with my review once these changes are made. (I don't expect to be slowed by the pandemic.)
