### 2020-04-30 #2161 @TomKellyGenetics
@majensen Sorry for the delays. I'm balancing childcare at home and a new position not related to this work. I'm still able work during the pandemic but revisions are taking longer than I'd have liked.

I expect the most time-consuming aspect will be revising the vignettes. The reviewers made good suggestions and I will make the requested revisions. Overall, I'm impressed by the JOSS review process and feel it will genuinely improve the package and paper.

It is a holiday period here in Japan next week. I will get back onto it after that and update you when it is ready. Thanks all for your patience.
