### 2020-04-30 a319ca6
Update issue templates
We are busy but will consider all requests to add features.
Please describe how it will benefit you and other users 
in as much detail as you can so that we can consider it.

If this is relate to and Issue please mention it here.

Please see the CONTRUBUTING.md file for more details
on how to improve the package. Thank you.

**Is your feature request related to a problem? Please describe.**
A clear and concise description of what the problem is. Ex. I'm always frustrated when [...]

**Describe the solution you'd like**
A clear and concise description of what you want the
[…]
