+++ .github/ISSUE_TEMPLATE/feature_request.md
@@ -1,0 +1,30 @@
+---
+name: Feature request
+about: Suggest a new functionality for this project
+title: "[NEW FEATURE] Please add <functionality> because <reasons>"
+labels: enhancement
+assignees: ''
+
+---
+
+We are busy but will consider all requests to add features.
+Please describe how it will benefit you and other users 
+in as much detail as you can so that we can consider it.
+
+**Is your feature request related to a problem? Please describe.**
+A clear and concise description of what the problem is. Ex. I'm always frustrated when [...]
+
+**Describe the solution you'd like**
+A clear and concise description of what you want the package to be able to do.
+
+**Describe alternatives you've considered**
+A clear and concise description of any alternative solutions or features you've considered.
+
+**Additional context**
+Add any other context or screenshots about the feature request here.
+
+  Delete these instructions once you have read them.
+
+---
+
+Describe the feature you wish to be added
