+++ .github/ISSUE_TEMPLATE/report-error.md
@@ -1,0 +1,51 @@
+---
+name: Report Error
+about: Create a report to help us improve the package
+title: "[BUG REPORT] with <function> on <data>"
+labels: bug
+assignees: ''
+
+---
+
+We hope that you are finding the package useful.
+Thank you for reporting errors so that we can improve it.
+Please describe the problem you are having in as much
+ detail as you can so that we can fix it and test our solutions.
+
+**Describe the bug**
+Please briefly describe your problem and what output you expect. If you have a question,
+please check that it has not already been addressed by a previous issue or
+asked before. Please do an internet search for the error first and see if it is 
+similar to problems others have had.
+
+Please check if the problem is with the `graphsim` package or any of it's dependencies. 
+You can run `rlang::trace_back()` in R to check this.
+
+**To Reproduce**
+Please include a minimal reprex. The goal of a reprex is to make it as easy as possible
+for me to recreate your problem so that I can fix it. If you've never heard of a reprex
+before, start by reading <https://www.tidyverse.org/help/#reprex>, and follow the advice
+further down the page. Do NOT include session info unless it's explicitly asked for,
+or you've used `reprex::reprex(..., si = TRUE)` to hide it away.  
+
+**Expected behavior**
+Please describe what you expect to happen
+
+**Screenshots**
+If applicable, give examples of the code you've tried or screenshots
+of results  to help explain your problem.
+
+**Additional context**
+R is a cross-platform language so there is no need to specify which OS,
+version or packages are installed. We will ask for it if it is relevant
+such as problems installing the package.
+
+  Delete these instructions once you have read them.
+
+---
+
+Describe your problem
+
+```r
+# insert reprex here
+```
