### 2020-04-30 7d92711
clean up issue templates #10
