### 2020-04-30 TomKellyGenetics/graphsim#10 @TomKellyGenetics
Using GitHub "Setting" to [configure](https://help.github.com/en/github/building-a-strong-community/configuring-issue-templates-for-your-repository) the new issue format seems to have it up properly. I don't have strong preferences on how issues are submitted. I've removed anything specific to the tidyverse and I think it's sorted now.

![Screen Shot 2020-04-30 at 17 20 05](https://user-images.githubusercontent.com/5493325/80688633-144f6180-8b07-11ea-991a-b62f828cf154.png)
