### 2020-04-30 TomKellyGenetics/graphsim#6 @TomKellyGenetics
Thanks I'll come back to the examples once the vignettes are update as per #11 and #13.

> By the way, the existing vignettes rely on the animation package (among others), though the package itself does not; so it would be helpful to mention, with the installation instructions, that these packages must be installed in order to build the vignettes when installing graphsim

I'll bear this in mind. I may consider removing additional dependencies for the vignettes that aren't strictly needed. I think `gplots` for "heatmap.2" may fall into this category (looking into it this should really be under `Suggests:` in this case).
