### 2020-04-30 c610dd6
migrate to Rmarkdown #13 (figure 1)
