+++ paper/paper.md
@@ -248,7 +248,6 @@ library("gplots")
 
 
 ```r
-#, fig.subcap=c("Activating pathway structure", "Inhibiting pathway structure")}
 #generate graph structure
 
 graph_edges <- rbind(c("A", "C"), c("B", "C"), c("C", "D"), c("D", "E"),
@@ -274,22 +273,12 @@ plot_directed(graph, state=state, layout = layout.kamada.kawai,
 
 \begin{figure}
 
-{\centering \includegraphics[width=.455\linewidth,height=.455\linewidth]{Plotunnamed-chunk-4-1} \includegraphics[width=.455\linewidth,height=.455\linewidth]{Plotunnamed-chunk-4-2} 
+{\centering \includegraphics[width=.400\linewidth,height=.400\linewidth]{Plotunnamed-chunk-4-1} \includegraphics[width=.400\linewidth,height=.400\linewidth]{Plotunnamed-chunk-4-2} 
 
 }
 
-\caption{\textbf{Simulated graph structures.}. A constructed graph structure used as an example to demonstrate the simulation procedure inFigures 2 and 3. Activating links are denoted by black arrows and inhibiting links by red edges. Inhibiting edges have been highlighted in red.}\label{fig:unnamed-chunk-4}
+\caption{\textbf{Simulated graph structures}. A constructed graph structure used as an example to demonstrate the simulation procedure inFigures 2 and 3. Activating links are denoted by black arrows and inhibiting links by red edges. Inhibiting edges have been highlighted in red.}\label{fig:unnamed-chunk-4}
 \end{figure}
-
-![\small \textbf{\textbf{Simulating expression from a graph structure.}}
-An example of a graph structure (a) that has been used to derive a
-relationship matrix (b), $\Sigma$ matrix (c) and correlation structure
-(d) from the relative distances between the nodes. Non-negative values
-are coloured white to red from $0$ to $1$. This $\Sigma$ matrix has been
-used to generate a simulated expression dataset of 100 samples (coloured
-blue to red from low to high) via sampling from the multivariate
-normal distribution. Here genes with closer relationships in the pathway
-structure show a higher correlation between simulated values.](figure2.pdf){ width=95% #fig:simulation_activating label="simulation_activating"}
 
 Generating a Simulated Expression Dataset {#sec:graphsim_demo}
 -----------------------------------------
@@ -323,53 +312,87 @@ normalised on a log-scale.
 
 The simulated dataset can be generated using the following code:
 
+
 ```r
 # activating graph
-state <- rep(length(E(graph)), 1)
+state <- rep(1, length(E(graph)))
 plot_directed(graph, state=state, layout = layout.kamada.kawai,
-              cex.node=3, cex.arrow=5, arrow_clip = 0.2)
-              
+              cex.node=3, cex.arrow=5, arrow_clip = 0.2,
+              sub = "(a) Activating pathway structure", cex.sub = 1.5)
+
 #adjacency matrix
-
 adj_mat <- make_adjmatrix_graph(graph)
-
 #relationship matrix
-
 dist_mat <- make_distance_graph(graph, absolute = FALSE)
-
 #sigma matrix directly from graph
-
 sigma_mat <- make_sigma_mat_dist_graph(graph, 0.8, absolute = FALSE)
-
+#> Warning in make_state_matrix(graph, state): State inferred from non-
+#> integer weighted edges: Please give numeric states as integers: 0 or
+#> 1 for activating, -1 or 2 for inhibiting
 #show shortest paths of graph
-
 shortest_paths <- shortest.paths(graph)
-
 #generate expression data directly from graph
-
 expr <- generate_expression(100, graph, cor = 0.8, mean = 0, comm = FALSE,
[…]
