+++ paper/paper.Rmd
@@ -238,8 +238,7 @@ library("graphsim")
 library("gplots")
 ```
 
-```{r, fig.align='center', fig.show='hold', fig.width='1.0\\linewidth', fig.height='1.0\\linewidth', out.height='.455\\linewidth', out.width='.455\\linewidth', fig.retina=10, fig.margin = FALSE, fig.cap = '\\textbf{Simulated graph structures.}. A constructed graph structure used as an example to demonstrate the simulation procedure inFigures 2 and 3. Activating links are denoted by black arrows and inhibiting links by red edges. Inhibiting edges have been highlighted in red.'}
-#, fig.subcap=c("Activating pathway structure", "Inhibiting pathway structure")}
+```{r, fig.align='center', fig.show='hold', fig.width='1.0\\linewidth', fig.height='1.0\\linewidth', out.height='.400\\linewidth', out.width='.400\\linewidth', fig.retina=10, fig.margin = FALSE, fig.ncol = 2, fig.cap = '\\textbf{Simulated graph structures}. A constructed graph structure used as an example to demonstrate the simulation procedure inFigures 2 and 3. Activating links are denoted by black arrows and inhibiting links by red edges. Inhibiting edges have been highlighted in red.'}
 #generate graph structure
 
 graph_edges <- rbind(c("A", "C"), c("B", "C"), c("C", "D"), c("D", "E"),
@@ -262,16 +261,6 @@ plot_directed(graph, state=state, layout = layout.kamada.kawai,
               cex.node = 2, cex.arrow = 4, arrow_clip = 0.2,
               sub = "(b) Inhibiting pathway structure", cex.sub = 1.5)
 ```
-
-![\small \textbf{\textbf{Simulating expression from a graph structure.}}
-An example of a graph structure (a) that has been used to derive a
-relationship matrix (b), $\Sigma$ matrix (c) and correlation structure
-(d) from the relative distances between the nodes. Non-negative values
-are coloured white to red from $0$ to $1$. This $\Sigma$ matrix has been
-used to generate a simulated expression dataset of 100 samples (coloured
-blue to red from low to high) via sampling from the multivariate
-normal distribution. Here genes with closer relationships in the pathway
-structure show a higher correlation between simulated values.](figure2.pdf){ width=95% #fig:simulation_activating label="simulation_activating"}
 
 Generating a Simulated Expression Dataset {#sec:graphsim_demo}
 -----------------------------------------
@@ -305,53 +294,59 @@ normalised on a log-scale.
 
 The simulated dataset can be generated using the following code:
 
-```r
+```{r, fig.align='center', fig.show='hold', fig.width='1.0\\linewidth', fig.height='1.0\\linewidth', out.height='.400\\linewidth', out.width='.400\\linewidth', fig.retina=10, fig.margin = FALSE, fig.ncol = 2, fig.cap = '\\textbf{Simulating expression from a graph structure}. An example of a graph structure (a) that has been used to derive a relationship matrix (b), $\\Sigma$  matrix (c) and correlation structure (d) from the relative distances between the nodes. Non-negative values are coloured white to red from $0$ to $1$. This $\\Sigma$ matrix has been used to generate a simulated expression dataset of 100 samples (coloured blue to red from low to high) via sampling from the multivariate normal distribution. Here genes with closer relationships in the pathway structure show a higher correlation between simulated values.'}
 # activating graph
-state <- rep(length(E(graph)), 1)
+state <- rep(1, length(E(graph)))
[…]
