+++ paper/paper.md
@@ -273,11 +273,11 @@ plot_directed(graph, state=state, layout = layout.kamada.kawai,
 
 \begin{figure}
 
-{\centering \includegraphics[width=.400\linewidth,height=.400\linewidth]{Plotunnamed-chunk-4-1} \includegraphics[width=.400\linewidth,height=.400\linewidth]{Plotunnamed-chunk-4-2} 
+{\centering \includegraphics[width=.415\linewidth,height=.415\linewidth]{Plotsimple_graph-1} \includegraphics[width=.415\linewidth,height=.415\linewidth]{Plotsimple_graph-2} 
 
 }
 
-\caption{\textbf{Simulated graph structures}. A constructed graph structure used as an example to demonstrate the simulation procedure inFigures 2 and 3. Activating links are denoted by black arrows and inhibiting links by red edges. Inhibiting edges have been highlighted in red.}\label{fig:unnamed-chunk-4}
+\caption{\textbf{Simulated graph structures}. A constructed graph structure used as an example to demonstrate the simulation procedure in Figures 2 and 3. Activating links are denoted by black arrows and inhibiting links by red edges. Inhibiting edges have been highlighted in red.}\label{fig:simple_graph}
 \end{figure}
 
 Generating a Simulated Expression Dataset {#sec:graphsim_demo}
@@ -326,17 +326,11 @@ adj_mat <- make_adjmatrix_graph(graph)
 dist_mat <- make_distance_graph(graph, absolute = FALSE)
 #sigma matrix directly from graph
 sigma_mat <- make_sigma_mat_dist_graph(graph, 0.8, absolute = FALSE)
-#> Warning in make_state_matrix(graph, state): State inferred from non-
-#> integer weighted edges: Please give numeric states as integers: 0 or
-#> 1 for activating, -1 or 2 for inhibiting
 #show shortest paths of graph
 shortest_paths <- shortest.paths(graph)
 #generate expression data directly from graph
 expr <- generate_expression(100, graph, cor = 0.8, mean = 0, comm = FALSE,
                             dist = TRUE, absolute = FALSE, state = state)
-#> Warning in make_state_matrix(graph, state): State inferred from non-
-#> integer weighted edges: Please give numeric states as integers: 0 or
-#> 1 for activating, -1 or 2 for inhibiting
 #> Warning in generate_expression(100, graph, cor = 0.8, mean = 0,
 #> comm = FALSE, : sigma matrix was not positive definite, nearest
 #> approximation used.
@@ -355,21 +349,15 @@ heatmap.2(make_sigma_mat_dist_graph(graph, 0.8, absolute = FALSE),
 scale = "none", trace = "none", col = colorpanel(50, "white", "red"),
 colsep = 1:length(V(graph)), rowsep = 1:length(V(graph)),
               sub = expression(paste("(c) ", Sigma, " matrix")), cex.sub = 1.5)
-#> Warning in make_state_matrix(graph, state): State inferred from non-
-#> integer weighted edges: Please give numeric states as integers: 0 or
-#> 1 for activating, -1 or 2 for inhibiting
 expr <- generate_expression(100, graph, cor = 0.8, mean = 0,
 comm = FALSE, dist =TRUE, absolute = FALSE, state = state)
-#> Warning in make_state_matrix(graph, state): State inferred from non-
-#> integer weighted edges: Please give numeric states as integers: 0 or
-#> 1 for activating, -1 or 2 for inhibiting
-
-#> Warning in make_state_matrix(graph, state): sigma matrix was not
-#> positive definite, nearest approximation used.
+#> Warning in generate_expression(100, graph, cor = 0.8, mean = 0,
+#> comm = FALSE, : sigma matrix was not positive definite, nearest
+#> approximation used.
[…]
