### 2020-05-01 962ce21
correc figure formatting for graph structures
