+++ paper/paper.Rmd
@@ -248,8 +248,9 @@ graph <- graph.edgelist(graph_edges, directed = TRUE)
 #plot graph structure (Figure 1a)
 
 plot_directed(graph, state ="activating", layout = layout.kamada.kawai,
-              cex.node = 2, cex.arrow = 4, arrow_clip = 0.2,
-              sub = "(a) Activating pathway structure", cex.sub = 1.5)
+              cex.node = 2, cex.arrow = 4, arrow_clip = 0.2)
+mtext(text = "(a) Activating pathway structure", side=1, line=3, at=0.05, adj=0.5, cex=1.75)
+box()
 
 #generate parameters for inhibitions
 
@@ -258,8 +259,9 @@ state <- c(1, 1, -1, 1, 1, 1, 1, -1, 1)
 #plot graph structure with inhibitions (Figure 1b)
 
 plot_directed(graph, state=state, layout = layout.kamada.kawai,
-              cex.node = 2, cex.arrow = 4, arrow_clip = 0.2,
-              sub = "(b) Inhibiting pathway structure", cex.sub = 1.5)
+              cex.node = 2, cex.arrow = 4, arrow_clip = 0.2)
+mtext(text = "(b) Inhibiting pathway structure", side=1, line=2, at=0.075, adj=0.5, cex=1.75)
+box()
 ```
 
 Generating a Simulated Expression Dataset {#sec:graphsim_demo}
@@ -298,41 +300,51 @@ The simulated dataset can be generated using the following code:
 # activating graph
 state <- rep(1, length(E(graph)))
 plot_directed(graph, state=state, layout = layout.kamada.kawai,
-              cex.node=3, cex.arrow=5, arrow_clip = 0.2,
-              sub = "(a) Activating pathway structure", cex.sub = 1.5)
+              cex.node=2, cex.arrow=4, arrow_clip = 0.2,
+              sub = "(a) Activating pathway structure", cex.sub = 1.75)
+box()
 
 #adjacency matrix
 adj_mat <- make_adjmatrix_graph(graph)
+
 #relationship matrix
 dist_mat <- make_distance_graph(graph, absolute = FALSE)
+
 #sigma matrix directly from graph
 sigma_mat <- make_sigma_mat_dist_graph(graph, 0.8, absolute = FALSE)
+
 #show shortest paths of graph
 shortest_paths <- shortest.paths(graph)
+
 #generate expression data directly from graph
 expr <- generate_expression(100, graph, cor = 0.8, mean = 0, comm = FALSE,
                             dist = TRUE, absolute = FALSE, state = state)
+
 #plot adjacency matrix
 heatmap.2(make_adjmatrix_graph(graph), scale = "none", trace = "none",
           col = colorpanel(3, "grey75", "white", "blue"),
           colsep = 1:length(V(graph)), rowsep = 1:length(V(graph)),
-              sub = "Adjacency matrix", cex.sub = 1.5)
+              sub = "Adjacency matrix", cex.sub = 1.75)
+
 #plot relationship matrix
 heatmap.2(make_distance_graph(graph, absolute = FALSE),
           scale = "none", trace = "none", col = colorpanel(50, "white", "red"),
 colsep = 1:length(V(graph)), rowsep = 1:length(V(graph)),
-              sub = "(b) relationship matrix", cex.sub = 1.5)
+              sub = "(b) relationship matrix", cex.sub = 1.75)
+
 #plot sigma matrix
 heatmap.2(make_sigma_mat_dist_graph(graph, 0.8, absolute = FALSE),
 scale = "none", trace = "none", col = colorpanel(50, "white", "red"),
 colsep = 1:length(V(graph)), rowsep = 1:length(V(graph)),
[…]
