### 2020-05-01 fd59448
correct figure 2-4 formatting for graph structures
