+++ paper/paper.Rmd
@@ -260,7 +260,7 @@ state <- c(1, 1, -1, 1, 1, 1, 1, -1, 1)
 
 plot_directed(graph, state=state, layout = layout.kamada.kawai,
               cex.node = 2, cex.arrow = 4, arrow_clip = 0.2)
-mtext(text = "(b) Inhibiting pathway structure", side=1, line=2, at=0.075, adj=0.5, cex=1.75)
+mtext(text = "(b) Inhibiting pathway structure", side=1, line=3, at=0.075, adj=0.5, cex=1.75)
 box()
 ```
 
@@ -300,8 +300,8 @@ The simulated dataset can be generated using the following code:
 # activating graph
 state <- rep(1, length(E(graph)))
 plot_directed(graph, state=state, layout = layout.kamada.kawai,
-              cex.node=2, cex.arrow=4, arrow_clip = 0.2,
-              sub = "(a) Activating pathway structure", cex.sub = 1.75)
+              cex.node=2, cex.arrow=4, arrow_clip = 0.2)
+mtext(text = "(a) Activating pathway structure", side=1, line=3, at=0.075, adj=0.5, cex=1.75)
 box()
 
 #adjacency matrix
@@ -320,43 +320,65 @@ shortest_paths <- shortest.paths(graph)
 expr <- generate_expression(100, graph, cor = 0.8, mean = 0, comm = FALSE,
                             dist = TRUE, absolute = FALSE, state = state)
 
-#plot adjacency matrix
-heatmap.2(make_adjmatrix_graph(graph), scale = "none", trace = "none",
-          col = colorpanel(3, "grey75", "white", "blue"),
-          colsep = 1:length(V(graph)), rowsep = 1:length(V(graph)),
-              sub = "Adjacency matrix", cex.sub = 1.75)
-
 #plot relationship matrix
 heatmap.2(make_distance_graph(graph, absolute = FALSE),
           scale = "none", trace = "none", col = colorpanel(50, "white", "red"),
-colsep = 1:length(V(graph)), rowsep = 1:length(V(graph)),
-              sub = "(b) relationship matrix", cex.sub = 1.75)
+colsep = 1:length(V(graph)), rowsep = 1:length(V(graph)))
+mtext(text = "(b) relationship matrix", side=1, line=3, at=0.075, adj=0.5, cex=1.75)
 
 #plot sigma matrix
 heatmap.2(make_sigma_mat_dist_graph(graph, 0.8, absolute = FALSE),
 scale = "none", trace = "none", col = colorpanel(50, "white", "red"),
-colsep = 1:length(V(graph)), rowsep = 1:length(V(graph)),
-              sub = expression(paste("(c) ", Sigma, " matrix")), cex.sub = 1.75)
+colsep = 1:length(V(graph)), rowsep = 1:length(V(graph)))
+mtext(text = expression(paste("(c) ", Sigma, " matrix")), side=1, line=3, at=0.075, adj=0.5, cex=1.75)
 
-#plot simulated expression data
+#simulated data
 expr <- generate_expression(100, graph, cor = 0.8, mean = 0,
 comm = FALSE, dist =TRUE, absolute = FALSE, state = state)
-heatmap.2(expr, scale = "none", trace = "none", col = bluered(50),
-colsep = 1:length(V(graph)), rowsep = 1:length(V(graph)),
-              sub = "(e) Simulated expression data (log scale)", cex.sub = 1.75)
 
 #plot simulated correlations
 heatmap.2(cor(t(expr)), scale = "none", trace = "none", col = colorpanel(50, "white", "red"),
 colsep = 1:length(V(graph)), rowsep = 1:length(V(graph)))
[…]
