### 2020-05-01 ac500ac
correct figure 2-4 formatting for heatmaps
