### 2020-05-08 50ce830
correct figure widths (fig2-4, pane E)
