+++ paper/paper.Rmd
@@ -4,6 +4,7 @@ output:
   rmarkdown::pdf_document:
     fig_crop: no
     keep_md: TRUE
+    keep_tex: TRUE
     fig_caption: yes
 tags:
   - R
@@ -324,6 +325,7 @@ shortest_paths <- shortest.paths(graph)
 expr <- generate_expression(100, graph, cor = 0.8, mean = 0, comm = FALSE,
                             dist = TRUE, absolute = FALSE, state = state)
 
+op <- par(mfrow=c(2,2), mar=c(4,4,4,1), oma=c(0.5,0.5,0.5,0))
 #plot relationship matrix
 heatmap.2(make_distance_graph(graph, absolute = FALSE),
           scale = "none", trace = "none", col = colorpanel(50, "white", "red"),
@@ -345,11 +347,12 @@ heatmap.2(cor(t(expr)), scale = "none", trace = "none", col = colorpanel(50, "wh
 colsep = 1:length(V(graph)), rowsep = 1:length(V(graph)))
 mtext(text = "(d) Simulated correlation", side=1, line=3, at=0.075, adj=0.5, cex=1.75)
 
+op <- par(mfrow=c(1,1), mar=c(4,4,4,1), oma=c(0.5,0.5,0.5,0))
 #plot simulated expression data
 heatmap.2(expr, scale = "none", trace = "none", col = bluered(50),
-colsep = 1:length(V(graph)), rowsep = 1:length(V(graph)),
-sub = "(e) Simulated expression data (log scale)", cex.sub = 1.75,
-xlim = c(-0.5, 1.5))
+colsep = 1:length(V(graph)), rowsep = 1:length(V(graph)))
+mtext(text = "(e) Simulated expression data (log scale)", side=1, line=3, at=0.075, adj=0.5, cex=1.75)
+par(op)
 ```
 
 The simulation procedure
