### 2020-05-09 1d3f44d
correct figure formatting and labels  (fig2-4, pane B-D)
