+++ paper/paper.Rmd
@@ -39,7 +39,8 @@ header-includes:
 knitr::opts_chunk$set(collapse = TRUE, comment = "#>", width = 68)
 knitr::opts_chunk$set(fig.cap = "", fig.path = "Plot")
 options(width = 68, cli.unicode = FALSE, cli.width = 68)
-par(mai=c(1.02, 0.82, 0.82, 0.82)-0.82)
+#par(mai=c(2.82, 2.82, 0.82, 0.82)-0.82)
+par(mar=c(7, 10, 4, 2) + 0.1)
 par(bty="o")
 #captioner::captioner(prefix = "Fig.")
 ```
@@ -250,7 +251,7 @@ graph <- graph.edgelist(graph_edges, directed = TRUE)
 
 plot_directed(graph, state ="activating", layout = layout.kamada.kawai,
               cex.node = 2, cex.arrow = 4, arrow_clip = 0.2)
-mtext(text = "(a) Activating pathway structure", side=1, line=3, at=0.05, adj=0.5, cex=1.75)
+mtext(text = "(a) Activating pathway structure", side=1, line=3.5, at=0.05, adj=0.5, cex=1.75)
 box()
 
 #generate parameters for inhibitions
@@ -261,7 +262,7 @@ state <- c(1, 1, -1, 1, 1, 1, 1, -1, 1)
 
 plot_directed(graph, state=state, layout = layout.kamada.kawai,
               cex.node = 2, cex.arrow = 4, arrow_clip = 0.2)
-mtext(text = "(b) Inhibiting pathway structure", side=1, line=3, at=0.075, adj=0.5, cex=1.75)
+mtext(text = "(b) Inhibiting pathway structure", side=1, line=3.5, at=0.075, adj=0.5, cex=1.75)
 box()
 ```
 
@@ -301,12 +302,16 @@ The simulated dataset can be generated using the following code:
 graphics::layout(matrix(c(1:5, 5), nrow = 3, byrow = TRUE))
 ```
 
+```{r, include = FALSE}
+set.seed(9000)
+```
+
 ```{r simulation_activating, fig.align='center', fig.show='hold', fig.width='1.0\\linewidth', fig.height='1.0\\linewidth', out.height='.415\\linewidth', out.width='.415\\linewidth', fig.retina=10, fig.margin = FALSE, fig.ncol = 2, fig.cap = '\\textbf{Simulating expression from a graph structure}. An example of a graph structure (a) that has been used to derive a relationship matrix (b), $\\Sigma$  matrix (c) and correlation structure (d) from the relative distances between the nodes. Non-negative values are coloured white to red from $0$ to $1$. This $\\Sigma$ matrix has been used to generate a simulated expression dataset of 100 samples (coloured blue to red from low to high) via sampling from the multivariate normal distribution. Here genes with closer relationships in the pathway structure show a higher correlation between simulated values.'}
 # activating graph
 state <- rep(1, length(E(graph)))
 plot_directed(graph, state=state, layout = layout.kamada.kawai,
               cex.node=2, cex.arrow=4, arrow_clip = 0.2)
-mtext(text = "(a) Activating pathway structure", side=1, line=3, at=0.075, adj=0.5, cex=1.75)
+mtext(text = "(a) Activating pathway structure", side=1, line=3.5, at=0.075, adj=0.5, cex=1.75)
 box()
 
 #adjacency matrix
@@ -325,18 +330,17 @@ shortest_paths <- shortest.paths(graph)
 expr <- generate_expression(100, graph, cor = 0.8, mean = 0, comm = FALSE,
[…]
