+++ paper/paper.tex
@@ -340,7 +340,7 @@ commands in R:
 
 \KeywordTok{plot_directed}\NormalTok{(graph, }\DataTypeTok{state =}\StringTok{"activating"}\NormalTok{, }\DataTypeTok{layout =}\NormalTok{ layout.kamada.kawai,}
               \DataTypeTok{cex.node =} \DecValTok{2}\NormalTok{, }\DataTypeTok{cex.arrow =} \DecValTok{4}\NormalTok{, }\DataTypeTok{arrow_clip =} \FloatTok{0.2}\NormalTok{)}
-\KeywordTok{mtext}\NormalTok{(}\DataTypeTok{text =} \StringTok{"(a) Activating pathway structure"}\NormalTok{, }\DataTypeTok{side=}\DecValTok{1}\NormalTok{, }\DataTypeTok{line=}\DecValTok{3}\NormalTok{, }\DataTypeTok{at=}\FloatTok{0.05}\NormalTok{, }\DataTypeTok{adj=}\FloatTok{0.5}\NormalTok{, }\DataTypeTok{cex=}\FloatTok{1.75}\NormalTok{)}
+\KeywordTok{mtext}\NormalTok{(}\DataTypeTok{text =} \StringTok{"(a) Activating pathway structure"}\NormalTok{, }\DataTypeTok{side=}\DecValTok{1}\NormalTok{, }\DataTypeTok{line=}\FloatTok{3.5}\NormalTok{, }\DataTypeTok{at=}\FloatTok{0.05}\NormalTok{, }\DataTypeTok{adj=}\FloatTok{0.5}\NormalTok{, }\DataTypeTok{cex=}\FloatTok{1.75}\NormalTok{)}
 \KeywordTok{box}\NormalTok{()}
 
 \CommentTok{#generate parameters for inhibitions}
@@ -351,7 +351,7 @@ commands in R:
 
 \KeywordTok{plot_directed}\NormalTok{(graph, }\DataTypeTok{state=}\NormalTok{state, }\DataTypeTok{layout =}\NormalTok{ layout.kamada.kawai,}
               \DataTypeTok{cex.node =} \DecValTok{2}\NormalTok{, }\DataTypeTok{cex.arrow =} \DecValTok{4}\NormalTok{, }\DataTypeTok{arrow_clip =} \FloatTok{0.2}\NormalTok{)}
-\KeywordTok{mtext}\NormalTok{(}\DataTypeTok{text =} \StringTok{"(b) Inhibiting pathway structure"}\NormalTok{, }\DataTypeTok{side=}\DecValTok{1}\NormalTok{, }\DataTypeTok{line=}\DecValTok{3}\NormalTok{, }\DataTypeTok{at=}\FloatTok{0.075}\NormalTok{, }\DataTypeTok{adj=}\FloatTok{0.5}\NormalTok{, }\DataTypeTok{cex=}\FloatTok{1.75}\NormalTok{)}
+\KeywordTok{mtext}\NormalTok{(}\DataTypeTok{text =} \StringTok{"(b) Inhibiting pathway structure"}\NormalTok{, }\DataTypeTok{side=}\DecValTok{1}\NormalTok{, }\DataTypeTok{line=}\FloatTok{3.5}\NormalTok{, }\DataTypeTok{at=}\FloatTok{0.075}\NormalTok{, }\DataTypeTok{adj=}\FloatTok{0.5}\NormalTok{, }\DataTypeTok{cex=}\FloatTok{1.75}\NormalTok{)}
 \KeywordTok{box}\NormalTok{()}
 \end{Highlighting}
 \end{Shaded}
[…]

[…]
