### 2020-05-11 TomKellyGenetics/graphsim#13 @TomKellyGenetics
>Code in manuscript: Given that there is a one-to-one relation between the figures generated and the code blocks, would it be desirable to use an Rmd instead of an md file?

An Rmarkdown version of the paper has been generated. I've made some alterations to the output to allow different sized figures. I've made these changes to the TeX file to generate a preview of the PDF and the same changes to the markdown for @whedon. Subcaptions for figure panes are generated in R graphics to avoid depending on TeX packages for whedon. The Rmarkdown is a bit complicated to do this but it seems to work. I use `set.seed` and `pdf` (sizes in inches at 75 dpi) calls in R to make sure the larger figure (2E, 3E, and 4E) are the same 'pseudo-random' data as the smaller figure it replaces.

Thanks for the suggestion. I'm quite happy with the Rmarkdown format so I'll use something similar for the vignettes (as raised here and #11). Once I've got the documentation sorted, I'll update the text of the manuscript.
