### 2020-05-11 21eef61
update documentation and explain vignettes
