+++ vignettes/test_graph_divergent.Rmd
@@ -9,16 +9,20 @@ vignette: >
   %\VignetteEncoding{UTF-8}
 ---
 
-##Install dependencies for Demonstration
+## Install dependencies for Demonstration
+
+First we must load the packages required to use the pacakge functions with igraph objects and generate plots.
 
 ```{r, message=F, warning=F}
-library('mvtnorm')
 library("igraph")
 library("gplots")
 library("graphsim")
 ```
 
-##set up simulated graphs
+## Set up simulated graphs
+
+Here we set up a simple graph to demonstrate how connections in the graph structure lead to correlations in the final output. We create a simple pathway of 4 genes with 2 upstream genes activating a downstream pathway.
+
 
 ```{r, out.width = '50%', out.height  = '50%', fig.align='center', dpi=36}
 graph_test2_edges <- rbind(c("C", "E"), c("D", "E"), c("E", "F"))
@@ -26,27 +30,67 @@ graph_test2 <- graph.edgelist(graph_test2_edges, directed = T)
 plot_directed(graph_test2, layout = layout.kamada.kawai)
 ```
 
-#Generated simulated expression data from graph
+# Generated simulated expression data from graph
 
-##Adjacency matrix
+Here we show the data generated by this graph.
+
+## Adjacency matrix
+
+The data can be summarised by an "adjacency matrix" where a one (1) is given between a row `i` and column `j` if there is an edge between genes `i` and `j`. Otherise it is a zero (0) for genes that are not connected. For an undirected graph, edges are shown in a symmetical matrix.
 
 ```{r, out.width = '50%', out.height  = '50%', fig.align='center', dpi=36}
 adj_mat <- make_adjmatrix_graph(graph_test2)
 heatmap.2(make_adjmatrix_graph(graph_test2), scale = "none", trace = "none", col = colorpanel(3, "grey75", "white", "blue"), colsep = 1:4, rowsep = 1:4)
+```
+
+For a directed graph, the edges are not shown from gene (row) `i` and gene (column) `j` 
+
+```{r, out.width = '50%', out.height  = '50%', fig.align='center', dpi=36}
 heatmap.2(make_adjmatrix_graph(graph_test2, directed = T), scale = "none", trace = "none", col = colorpanel(3, "grey75", "white", "blue"), colsep = 1:4, rowsep = 1:4)
+```
+
+We can compute the common links between each pair of genes. This shows how many genes are connected to both genes `i` and `j`.
+
+```{r, out.width = '50%', out.height  = '50%', fig.align='center', dpi=36}
 comm_mat <- make_commonlink_graph(graph_test2)
 heatmap.2(make_commonlink_graph(graph_test2), scale = "none", trace = "none", col = bluered(50), colsep = 1:4, rowsep = 1:4)
 ```
 
+Note that this weights towards genes with a higher vertex degree (as does the Laplacian).
+
+```{r, out.width = '50%', out.height  = '50%', fig.align='center', dpi=36}
+laplacian_mat <- make_laplacian_graph(graph_test2)
+heatmap.2(make_laplacian_graph(graph_test2), scale = "none", trace = "none", col = bluered(50), colsep = 1:4, rowsep = 1:4)
+```
+
 ##Distance matrix
+
+To compute the relationships between each gene by "distance" we first compute the shortest paths between each pair of nodes by Dijkstra's algorithm. 
 
 ```{r, out.width = '50%', out.height  = '50%', fig.align='center', dpi=36}
 shortest.paths(graph_test2)
 heatmap.2(shortest.paths(graph_test2), scale = "none", trace = "none", col = bluered(50), colsep = 1:4, rowsep = 1:4)
+```
[…]
