### 2020-05-11 TomKellyGenetics/graphsim#13 @rcannood
Cool! I like the changes you've made; it's much easier to see what code is used to produce which figure.

One suggestion, in the "load required packages" session you might want to turn off messages (for example like this [6f979bb](https://github.com/rcannood/graphsim/commit/6f979bbb3513b95872b9e4227d7a30699e27fd14))). Perhaps you might also want to remove some of the warnings that are generated throughout the paper as well, but I'll leave that choice up to you.

↓ Warning such as these are usually a sign that something is amiss.
```r
#> Warning in eattrs[[name]][index] <- value: number of items to
#> replace is not a multiple of replacement length
#> Warning in state_path[jj] <- state[kk]: number of items to replace
#> is not a multiple of replacement length
```
