### 2020-05-13 ce18cc6
update toy examples in vignettes
