+++ vignettes/test_graph_convergent.Rmd
@@ -1,88 +1,188 @@
 ---
-title: "Simulating convergent graph structure in continuous data"
+title: "Simulating divergent graph structure in continuous data"
 author: "Tom Kelly"
 date: "`r Sys.Date()`"
 output: rmarkdown::html_document
 vignette: >
-  %\VignetteIndexEntry{Simulating convergent graph structure in continuous data}
+  %\VignetteIndexEntry{Simulating divergent graph structure in continuous data}
   %\VignetteEngine{knitr::knitr}
   %\VignetteEncoding{UTF-8}
 ---
 
-##Install dependencies for Demonstration
+# Getting Started
+
+## Install dependencies for Demonstration
+
+First we must load the packages required to use the pacakge functions with igraph objects and generate plots.
 
 ```{r, message=F, warning=F}
-library('mvtnorm')
 library("igraph")
 library("gplots")
 library("graphsim")
 ```
 
-##set up simulated graphs
+## Set up simulated graphs
+
+Here we set up a simple graph to demonstrate how connections in the graph structure lead to correlations in the final output. We create a simple pathway of 4 genes with 2 upstream genes activating a downstream pathway.
+
 
 ```{r, out.width = '50%', out.height  = '50%', fig.align='center', dpi=36}
-graph_test1_edges <- rbind(c("A", "B"), c("B", "C"), c("B", "D"))
-graph_test1 <- graph.edgelist(graph_test1_edges, directed = T)
-plot_directed(graph_test1, layout = layout.kamada.kawai)
+graph_test2_edges <- rbind(c("C", "E"), c("D", "E"), c("E", "F"))
+graph_test2 <- graph.edgelist(graph_test2_edges, directed = T)
+plot_directed(graph_test2, layout = layout.kamada.kawai)
 ```
 
-#Generated simulated expression data from graph
+# Generating  simulated expression data from graph
 
-##Adjacency matrix
+## Minimal example 
+
+A simulated dataset can be generated with a single command. This is all that is required to get started.
 
 ```{r, out.width = '50%', out.height  = '50%', fig.align='center', dpi=36}
-adj_mat <- make_adjmatrix_graph(graph_test1)
-heatmap.2(make_adjmatrix_graph(graph_test1), scale = "none", trace = "none", col = colorpanel(3, "grey75", "white", "blue"), colsep = 1:4, rowsep = 1:4)
-heatmap.2(make_adjmatrix_graph(graph_test1, directed = T), scale = "none", trace = "none", col = colorpanel(3, "grey75", "white", "blue"), colsep = 1:4, rowsep = 1:4)
-comm_mat <- make_commonlink_graph(graph_test1)
-heatmap.2(make_commonlink_graph(graph_test1), scale = "none", trace = "none", col = bluered(50), colsep = 1:4, rowsep = 1:4)
+expr<- generate_expression(100, graph_test2, cor = 0.8, mean = 0, comm = F, dist = T, absolute = F)
+heatmap.2(expr, scale = "none", trace = "none", col = bluered(50), colsep = 1:4, rowsep = 1:4)
 ```
 
-##Distance matrix
+Various arguments are supported to alter how the simulated datasets are computed. 
+
+# How it works step-by-step
+
+Here we show the data generated by for this graph structure. This demonstrates how several of the options available compute the necessary steps.
+
+## Adjacency matrix
+
[…]
