+++ vignettes/test_graph_network_inhibiting.Rmd
@@ -9,232 +9,182 @@ vignette: >
   %\VignetteEncoding{UTF-8}
 ---
 
-##Install dependencies for Demonstration
+# Getting Started
+  
+## Install dependencies for Demonstration
+  
+  First we must load the packages required to use the pacakge functions with igraph objects and generate plots.
 
 ```{r, message=F, warning=F}
-library('mvtnorm')
 library("igraph")
 library("gplots")
 library("graphsim")
-Rprof("test_graph_network_inhibiting.out", interval = 0.001, memory.profiling = F)
-#Rprof("test_graph_network_inhibiting.mem.out", interval = 0.001, memory.profiling = T)
 ```
 
-#one inhibiting link
+## Set up simulated graphs
 
-##set up simulated graphs
+Here we set up a simple graph to demonstrate how connections in the graph structure lead to correlations in the final output. We create a simple pathway of 9 genes with various branches.
 
 
-``````{r, out.width = '50%', out.height  = '50%', fig.align='center', dpi=36}
+```{r, out.width = '50%', out.height  = '50%', fig.align='center', dpi=36}
 graph_test4_edges <- rbind(c("A", "C"), c("B", "C"), c("C", "D"), c("D", "E"), c("D", "F"), c("F", "G"), c("F", "I"), c("H", "I"))
 graph_test4 <- graph.edgelist(graph_test4_edges, directed = T)
-state_vec <- c(1, 1, 1, 1, -1, 1, 1, 1)
-plot_directed(graph_test4, layout = layout.kamada.kawai, state = state_vec)
+state <- c(1, 1, -1, 1, 1, 1, 1, -1)
+E(graph_test4)$state <- state
+plot_directed(graph_test4, layout = layout.kamada.kawai)
 ```
 
-##Generated simulated expression data from graph
+# Generating  simulated expression data from graph
 
-###Adjacency matrix
+## Minimal example 
 
-``````{r, out.width = '50%', out.height  = '50%', fig.align='center', dpi=36}
+A simulated dataset can be generated with a single command. This is all that is required to get started.
+
+```{r, out.width = '50%', out.height  = '50%', fig.align='center', dpi=36}
+expr <- generate_expression(100, graph_test4, cor = 0.8, mean = 0, comm = F, dist = T, absolute = F)
+heatmap.2(expr, scale = "none", trace = "none", col = bluered(50), colsep = 1:4, rowsep = 1:4)
+```
+
+Various arguments are supported to alter how the simulated datasets are computed. 
+
+# How it works step-by-step
+
+Here we show the data generated by for this graph structure. This demonstrates how several of the options available compute the necessary steps.
+
+## Adjacency matrix
+
+The data can be summarised by an "adjacency matrix" where a one (1) is given between a row `i` and column `j` if there is an edge between genes `i` and `j`. Otherise it is a zero (0) for genes that are not connected. For an undirected graph, edges are shown in a symmetical matrix.
+
+```{r, out.width = '50%', out.height  = '50%', fig.align='center', dpi=36}
 adj_mat <- make_adjmatrix_graph(graph_test4)
-heatmap.2(make_adjmatrix_graph(graph_test4), scale = "none", trace = "none", col = colorpanel(3, "grey75", "white", "blue"), colsep = 1:length(V(graph_test4)), rowsep = 1:length(V(graph_test4)))
[…]
