+++ paper/paper.Rmd
@@ -439,7 +439,7 @@ colsep = 1:length(V(graph)), rowsep = 1:length(V(graph)))
 mtext(text = expression(paste("(c) ", Sigma, " matrix")), side=1, line=3.5, at=0, adj=0.5, cex=1.75)
 
 #simulated data
-expr <- generate_expression(100, graph,  state = state, cor = 0.8, mean = 0,
+expr <- generate_expression(100, graph, state = state, cor = 0.8, mean = 0,
 comm = FALSE, dist =TRUE, absolute = FALSE)
 
 #plot simulated correlations
@@ -543,7 +543,7 @@ colsep = 1:length(V(graph)), rowsep = 1:length(V(graph)), labCol = "")
 mtext(text = expression(paste("(c) ", Sigma, " matrix")), side=1, line=3.5, at=0, adj=0.5, cex=1.75)
 
 #simulated data
-expr <- generate_expression(100, graph,  state = state, cor = 0.8, mean = 0,
+expr <- generate_expression(100, graph, state = state, cor = 0.8, mean = 0,
 comm = FALSE, dist =TRUE, absolute = FALSE)
 
 #plot simulated correlations
