### 2020-05-13 ff1c88c
correct passing state for inhibiting graph attributes (migrate to state computing in sigma rather than generate)
