### 2020-05-21 #2161 @TomKellyGenetics
Thanks for following up, it's taken longer than I'd planned but you can see most of the recent progress here: https://github.com/TomKellyGenetics/graphsim/issues/13

The main update is that the paper has been migrated to Rmarkdown format.
https://github.com/TomKellyGenetics/graphsim/blob/master/paper/paper.Rmd

I've also made changes to the computation of the `state` parameter which addresses [this](https://github.com/openjournals/joss-reviews/issues/2161#issuecomment-601921101). Figures have been updated to reflect this change.
> I'm unable to exactly reproduce the heatmaps displayed in the paper. They're based on the graph with inhibitions, and i'm also not clear on how to incorporate the vector state into these calculations, after looking into the code only a slight bit. (For issues running examples, etc. in the package itself, i'll open issues on the GitHub repo.)

I've started on the vignettes  which was raised by both reviewers. This includes reformatting the vignettes in similar style to the Rmarkdown paper and including examples from the `data` directory as suggested.
https://github.com/TomKellyGenetics/graphsim/issues/11
https://github.com/TomKellyGenetics/graphsim/issues/6

As discussed on the [issue](https://github.com/TomKellyGenetics/graphsim/issues/13) I'll update the manuscript text after the documentation. I should have some time to finish the biggest changes to the vignettes soon. I'll update this thread when the revised package and manuscript and ready for further review.
