### 2020-05-21 #2161 @arfon
@TomKellyGenetics - a few comments on your paper:

The fact that `paper.md` and `paper.tex` exist in your repository confuses Whedon (hence the errors above).

The compiled paper [compiled here](https://github.com/openjournals/joss-reviews/files/4662696/10.21105.joss.02161.pdf) is very long (15 pages currently) and much longer than we normally expect at JOSS (~1000 words is the guidance on length). Could you take a look at reducing the length, for example, by moving some of the long code examples to other documentation?
