### 2020-05-25 #2161 @TomKellyGenetics
Thanks @arfon. I was unsure if Whedon would handle the new format. I'll look into that (I think adding the `paper.tex` to `.gitignore` and removing it from the repository) will avoid this.

I realise the manuscript is longer than recommended for the journal, I contacted @karthik about this before submitting. As suggested by reviewers, migrating to Rmarkdown format and including [examples of empirical networks](https://github.com/TomKellyGenetics/graphsim/issues/6#issue-587715947) aids readers to understand and reproduce the results. Unfortunately the Rmarkdown format does make the manuscript even longer (to include all code needed to produce the figures).

I'll consider removing some of it (e.g., Figures 2 and 3), when the vignettes have been updated to cover this material. As [suggested  here](https://github.com/TomKellyGenetics/graphsim/issues/13#issuecomment-626610045), removing warning messages in Rmarkdown would also make it shorter.

I'll check if panes of different sizes (e.g., Figure 2E) are compatible with Whedon. Since LaTeX many packages are not supported, it was difficult to include subcaptions on the figures, is it possible to add this functionality? The code to add captions in R also makes the manuscript longer than otherwise needed.
