### 2020-05-25 5ec3450
handling paper formatting
