### 2020-05-25 915b4c9
reformat code examples
