+++ paper/paper.Rmd
@@ -228,38 +228,28 @@ commands in R:
 
 ```{r, eval = FALSE}
 #install packages required (once per machine)
-
 install.packages("graphsim")
 ```
-
-```{r}
+```{r, warning=FALSE, results='hide'}
 #load required packages (once per R instance)
-
-library("igraph")
-library("graphsim")
-library("gplots")
+library("graphsim"); library("igraph"); library("gplots")
 ```
 
 ```{r simple_graph, fig.align='center', fig.show='hold', fig.width='1.0\\linewidth', fig.height='1.0\\linewidth', out.height='.415\\linewidth', out.width='.415\\linewidth', fig.retina=10, fig.margin = FALSE, fig.ncol = 2, fig.cap = '\\textbf{Simulated graph structures}. A constructed graph structure used as an example to demonstrate the simulation procedure in Figures 2 and 3. Activating links are denoted by black arrows and inhibiting links by red edges. Inhibiting edges have been highlighted in red.'}
 #generate graph structure
-
 graph_edges <- rbind(c("A", "C"), c("B", "C"), c("C", "D"), c("D", "E"),
                      c("D", "F"), c("F", "G"), c("F", "I"), c("H", "I"))
 graph <- graph.edgelist(graph_edges, directed = TRUE)
 
 #plot graph structure (Figure 1a)
-
 plot_directed(graph, state ="activating", layout = layout.kamada.kawai,
               cex.node = 2, cex.arrow = 4, arrow_clip = 0.2)
 mtext(text = "(a) Activating pathway structure", side=1, line=3.5, at=0.05, adj=0.5, cex=1.75)
 box()
 
 #generate parameters for inhibitions
-
 state <- c(1, 1, -1, 1, 1, 1, 1, -1, 1)
-
 #plot graph structure with inhibitions (Figure 1b)
-
 plot_directed(graph, state=state, layout = layout.kamada.kawai,
               cex.node = 2, cex.arrow = 4, arrow_clip = 0.2)
 mtext(text = "(b) Inhibiting pathway structure", side=1, line=3.5, at=0.075, adj=0.5, cex=1.75)
@@ -306,51 +296,45 @@ graphics::layout(matrix(c(1:5, 5), nrow = 3, byrow = TRUE))
 set.seed(9000)
 ```
 
-```{r simulation_activating, fig.align='center', fig.show='hold', fig.width='1.0\\linewidth', fig.height='1.0\\linewidth', out.height='.415\\linewidth', out.width='.415\\linewidth', fig.retina=10, fig.margin = FALSE, fig.ncol = 2, fig.cap = '\\textbf{Simulating expression from a graph structure}. An example of a graph structure (a) that has been used to derive a relationship matrix (b), $\\Sigma$  matrix (c) and correlation structure (d) from the relative distances between the nodes. Non-negative values are coloured white to red from $0$ to $1$. This $\\Sigma$ matrix has been used to generate a simulated expression dataset of 100 samples (coloured blue to red from low to high) via sampling from the multivariate normal distribution. Here genes with closer relationships in the pathway structure show a higher correlation between simulated values.'}
+```{r simulation_activating_setup, fig.align='center', fig.show='hold', fig.width='1.0\\linewidth', fig.height='1.0\\linewidth', out.height='.415\\linewidth', out.width='.415\\linewidth', fig.retina=10, fig.margin = FALSE, fig.ncol = 2, warning=FALSE}
 # activating graph
 state <- rep(1, length(E(graph)))
 plot_directed(graph, state=state, layout = layout.kamada.kawai,
[…]
