+++ paper/paper.md
@@ -221,15 +221,12 @@ commands in R:
 
 ```r
 #install packages required (once per machine)
-
 install.packages("graphsim")
 ```
 
-
 ```r
 #load required packages (once per R instance)
-
-library("igraph")
+library("graphsim"); library("igraph"); library("gplots")
 #> 
 #> Attaching package: 'igraph'
 #> The following objects are masked from 'package:stats':
@@ -238,8 +235,6 @@ library("igraph")
 #> The following object is masked from 'package:base':
 #> 
 #>     union
-library("graphsim")
-library("gplots")
 #> 
 #> Attaching package: 'gplots'
 #> The following object is masked from 'package:stats':
@@ -250,24 +245,19 @@ library("gplots")
 
 ```r
 #generate graph structure
-
 graph_edges <- rbind(c("A", "C"), c("B", "C"), c("C", "D"), c("D", "E"),
                      c("D", "F"), c("F", "G"), c("F", "I"), c("H", "I"))
 graph <- graph.edgelist(graph_edges, directed = TRUE)
 
 #plot graph structure (Figure 1a)
-
 plot_directed(graph, state ="activating", layout = layout.kamada.kawai,
               cex.node = 2, cex.arrow = 4, arrow_clip = 0.2)
 mtext(text = "(a) Activating pathway structure", side=1, line=3.5, at=0.05, adj=0.5, cex=1.75)
 box()
 
 #generate parameters for inhibitions
-
 state <- c(1, 1, -1, 1, 1, 1, 1, -1, 1)
-
 #plot graph structure with inhibitions (Figure 1b)
-
 plot_directed(graph, state=state, layout = layout.kamada.kawai,
               cex.node = 2, cex.arrow = 4, arrow_clip = 0.2)
 mtext(text = "(b) Inhibiting pathway structure", side=1, line=3.5, at=0.075, adj=0.5, cex=1.75)
@@ -327,50 +317,47 @@ plot_directed(graph, state=state, layout = layout.kamada.kawai,
               cex.node=2, cex.arrow=4, arrow_clip = 0.2)
 mtext(text = "(a) Activating pathway structure", side=1, line=3.5, at=0.075, adj=0.5, cex=1.75)
 box()
-
 #adjacency matrix
 adj_mat <- make_adjmatrix_graph(graph)
-
 #relationship matrix
 dist_mat <- make_distance_graph(graph, absolute = FALSE)
-
 #sigma matrix directly from graph
 sigma_mat <- make_sigma_mat_dist_graph(graph, 0.8, absolute = FALSE)
-
 #show shortest paths of graph
 shortest_paths <- shortest.paths(graph)
-
 #generate expression data directly from graph
 expr <- generate_expression(100, graph, cor = 0.8, mean = 0, comm = FALSE,
                             dist = TRUE, absolute = FALSE, state = state)
-#> Warning in generate_expression(100, graph, cor = 0.8, mean = 0,
-#> comm = FALSE, : sigma matrix was not positive definite, nearest
-#> approximation used.
+```
+
+\begin{figure}
+
+{\centering \includegraphics[width=.415\linewidth,height=.415\linewidth]{Plotsimulation_activating_setup-1} 
+
+}
+
+\end{figure}
+
+
+```r
 
 #plot relationship matrix
 heatmap.2(make_distance_graph(graph, absolute = FALSE),
           scale = "none", trace = "none", col = colorpanel(50, "white", "red"),
 colsep = 1:length(V(graph)), rowsep = 1:length(V(graph)))
 mtext(text = "(b) Relationship matrix", side=1, line=3.5, at=0, adj=0.5, cex=1.75)
-
 #plot sigma matrix
[…]
