### 2020-05-25 72bc867
update code examples
