+++ paper/paper.md
@@ -139,8 +139,8 @@ reference="fig:simple_graph:first"}. The graph can be defined by an
 adjacency matrix, **$A$** (with elements
 $A_{ij}$), where $$A_{ij} = 
 \begin{cases}
-   1                         & \mbox{if genes } i \mbox{ and } j \mbox{ are adjacent} \\
-   0                         & \mbox{otherwise}
+   1   & \mbox{if genes } i \mbox{ and } j \mbox{ are adjacent} \\
+   0   & \mbox{otherwise}
 \end{cases}$$
 
 A matrix, **$R$**, with elements
@@ -152,9 +152,9 @@ weighting is used to achieve this:
 
 [ $$R_{ij} = 
 \begin{cases}
-   1                            & \mbox{if genes } i \mbox{ and } j \mbox{ are adjacent} \\
-   (\frac{1}{2})^{d_{ij}}      & \mbox{if a path can be found between genes } i \mbox{ and } j \\
-   0                            & \mbox{if no path exists between genes } i \mbox{ and } j 
+   1  & \mbox{if genes } i \mbox{ and } j \mbox{ are adjacent} \\
+   (\frac{1}{2})^{d_{ij}}  & \mbox{if a path can be found between genes } i \mbox{ and } j \\
+   0  & \mbox{if no path exists between genes } i \mbox{ and } j 
 \end{cases}$$]{style="color: black"} where $d_{ij}$ is the length of
 the shortest path (i.e., minimum number of edges traversed) between
 genes (nodes) $i$ and $j$ in graph $G$. Each more distant node is thus
@@ -164,15 +164,15 @@ reference="fig:simulation_activating:second"}. An
 arithmetically-decreasing (absolute) distance weighting is also
 supported in the \texttt{graphsim} R package which implements this procedure: [ $$R_{ij} = 
 \begin{cases}
-   1                            & \mbox{if genes } i \mbox{ and } j \mbox{ are adjacent} \\
-   1-\frac{d_{ij}}{diam(G)}     & \mbox{if a path can be found between genes } i \mbox{ and } j \\
-   0                            & \mbox{if no path exists between genes } i \mbox{ and } j 
+   1  & \mbox{if genes } i \mbox{ and } j \mbox{ are adjacent} \\
+   1-\frac{d_{ij}}{diam(G)}   & \mbox{if a path can be found between genes } i \mbox{ and } j \\
+   0  & \mbox{if no path exists between genes } i \mbox{ and } j 
 \end{cases}$$ ]{style="color: black"}
 
 Assuming a unit variance for each gene, these values can be used to
 derive a $\Sigma$ matrix: $$\Sigma_{ij} = 
 \begin{cases}
-   1                            & \mbox{if } i=j \\
+   1  & \mbox{if } i=j \\
    \rho R_{ij}  & \mbox{otherwise}
 \end{cases}$$ where $\rho$ is the correlation between adjacent nodes.
 Thus covariances between adjacent nodes are assigned by a correlation
@@ -246,12 +246,12 @@ library("graphsim"); library("igraph"); library("gplots")
 ```r
 #generate graph structure
 graph_edges <- rbind(c("A", "C"), c("B", "C"), c("C", "D"), c("D", "E"),
-                     c("D", "F"), c("F", "G"), c("F", "I"), c("H", "I"))
+   c("D", "F"), c("F", "G"), c("F", "I"), c("H", "I"))
 graph <- graph.edgelist(graph_edges, directed = TRUE)
 
 #plot graph structure (Figure 1a)
[…]
