+++ paper/paper.Rmd
@@ -147,8 +147,8 @@ reference="fig:simple_graph:first"}. The graph can be defined by an
 adjacency matrix, **$A$** (with elements
 $A_{ij}$), where $$A_{ij} = 
 \begin{cases}
-   1                         & \mbox{if genes } i \mbox{ and } j \mbox{ are adjacent} \\
-   0                         & \mbox{otherwise}
+   1   & \mbox{if genes } i \mbox{ and } j \mbox{ are adjacent} \\
+   0   & \mbox{otherwise}
 \end{cases}$$
 
 A matrix, **$R$**, with elements
@@ -160,9 +160,9 @@ weighting is used to achieve this:
 
 [ $$R_{ij} = 
 \begin{cases}
-   1                            & \mbox{if genes } i \mbox{ and } j \mbox{ are adjacent} \\
-   (\frac{1}{2})^{d_{ij}}      & \mbox{if a path can be found between genes } i \mbox{ and } j \\
-   0                            & \mbox{if no path exists between genes } i \mbox{ and } j 
+   1  & \mbox{if genes } i \mbox{ and } j \mbox{ are adjacent} \\
+   (\frac{1}{2})^{d_{ij}}  & \mbox{if a path can be found between genes } i \mbox{ and } j \\
+   0  & \mbox{if no path exists between genes } i \mbox{ and } j 
 \end{cases}$$]{style="color: black"} where $d_{ij}$ is the length of
 the shortest path (i.e., minimum number of edges traversed) between
 genes (nodes) $i$ and $j$ in graph $G$. Each more distant node is thus
@@ -172,15 +172,15 @@ reference="fig:simulation_activating:second"}. An
 arithmetically-decreasing (absolute) distance weighting is also
 supported in the \texttt{graphsim} R package which implements this procedure: [ $$R_{ij} = 
 \begin{cases}
-   1                            & \mbox{if genes } i \mbox{ and } j \mbox{ are adjacent} \\
-   1-\frac{d_{ij}}{diam(G)}     & \mbox{if a path can be found between genes } i \mbox{ and } j \\
-   0                            & \mbox{if no path exists between genes } i \mbox{ and } j 
+   1  & \mbox{if genes } i \mbox{ and } j \mbox{ are adjacent} \\
+   1-\frac{d_{ij}}{diam(G)}   & \mbox{if a path can be found between genes } i \mbox{ and } j \\
+   0  & \mbox{if no path exists between genes } i \mbox{ and } j 
 \end{cases}$$ ]{style="color: black"}
 
 Assuming a unit variance for each gene, these values can be used to
 derive a $\Sigma$ matrix: $$\Sigma_{ij} = 
 \begin{cases}
-   1                            & \mbox{if } i=j \\
+   1  & \mbox{if } i=j \\
    \rho R_{ij}  & \mbox{otherwise}
 \end{cases}$$ where $\rho$ is the correlation between adjacent nodes.
 Thus covariances between adjacent nodes are assigned by a correlation
@@ -238,12 +238,12 @@ library("graphsim"); library("igraph"); library("gplots")
 ```{r simple_graph, fig.align='center', fig.show='hold', fig.width='1.0\\linewidth', fig.height='1.0\\linewidth', out.height='.415\\linewidth', out.width='.415\\linewidth', fig.retina=10, fig.margin = FALSE, fig.ncol = 2, fig.cap = '\\textbf{Simulated graph structures}. A constructed graph structure used as an example to demonstrate the simulation procedure in Figures 2 and 3. Activating links are denoted by black arrows and inhibiting links by red edges. Inhibiting edges have been highlighted in red.'}
 #generate graph structure
 graph_edges <- rbind(c("A", "C"), c("B", "C"), c("C", "D"), c("D", "E"),
[…]
