### 2020-05-25 #2161 @TomKellyGenetics
Here's a preview of the manuscript in Rmarkdown form with the code examples simplified (I'll add more details in the vignettes). This is a work-in-progress.

I've contacted my coauthor to discuss to reducing the length of the paper.
