### 2020-05-25 #2161 @arfon
> Here's a preview of the manuscript in Rmarkdown form with the code examples simplified (I'll add more details in the vignettes). This is a work-in-progress.

> I've contacted my coauthor to discuss to reducing the length of the paper.

Thanks for working on this @TomKellyGenetics.

> I'll check if panes of different sizes (e.g., Figure 2E) are compatible with Whedon. Since LaTeX many packages are not supported, it was difficult to include subcaptions on the figures, is it possible to add this functionality?

I don't believe a subcaption is possible with Pandoc and this template sorry.
