### 2020-05-25 #2161 @karthik
> Unfortunately the Rmarkdown format does make the manuscript even longer (to include all code needed to produce the figures).

Have you considered just hiding those blocks?
