+++ paper/paper.md
@@ -222,24 +222,14 @@ commands in R:
 ```r
 #install packages required (once per machine)
 install.packages("graphsim")
+
 ```
 
 ```r
 #load required packages (once per R instance)
-library("graphsim"); library("igraph"); library("gplots")
-#> 
-#> Attaching package: 'igraph'
-#> The following objects are masked from 'package:stats':
-#> 
-#>     decompose, spectrum
-#> The following object is masked from 'package:base':
-#> 
-#>     union
-#> 
-#> Attaching package: 'gplots'
-#> The following object is masked from 'package:stats':
-#> 
-#>     lowess
+library("graphsim")
+#load packages for examples
+library("igraph"); library("gplots")
 ```
 
 
@@ -252,16 +242,12 @@ graph <- graph.edgelist(graph_edges, directed = TRUE)
 #plot graph structure (Figure 1a)
 plot_directed(graph, state ="activating", layout = layout.kamada.kawai,
   cex.node = 2, cex.arrow = 4, arrow_clip = 0.2)
-mtext(text = "(a) Activating pathway structure", side=1, line=3.5, at=0.05, adj=0.5, cex=1.75)
-box()
 
-#generate parameters for inhibitions
+#generate parameters for inhibitions for each edge in E(graph)
 state <- c(1, 1, -1, 1, 1, 1, 1, -1, 1)
 #plot graph structure with inhibitions (Figure 1b)
 plot_directed(graph, state=state, layout = layout.kamada.kawai,
   cex.node = 2, cex.arrow = 4, arrow_clip = 0.2)
-mtext(text = "(b) Inhibiting pathway structure", side=1, line=3.5, at=0.075, adj=0.5, cex=1.75)
-box()
 ```
 
 \begin{figure}
@@ -310,36 +296,33 @@ The simulated dataset can be generated using the following code:
 
 
 
+
 ```r
 # activating graph
 state <- rep(1, length(E(graph)))
 plot_directed(graph, state=state, layout = layout.kamada.kawai,
   cex.node=2, cex.arrow=4, arrow_clip = 0.2)
-mtext(text = "(a) Activating pathway structure", side=1, line=3.5, at=0.075, adj=0.5, cex=1.75)
-box()
+
 #plot relationship matrix
 heatmap.2(make_distance_graph(graph, absolute = FALSE),
   scale = "none", trace = "none", col = colorpanel(50, "white", "red"),
-colsep = 1:length(V(graph)), rowsep = 1:length(V(graph)))
-mtext(text = "(b) Relationship matrix", side=1, line=3.5, at=0, adj=0.5, cex=1.75)
+  colsep = 1:length(V(graph)), rowsep = 1:length(V(graph)))
+
 #plot sigma matrix
 heatmap.2(make_sigma_mat_dist_graph(graph, 0.8, absolute = FALSE),
-scale = "none", trace = "none", col = colorpanel(50, "white", "red"),
-colsep = 1:length(V(graph)), rowsep = 1:length(V(graph)))
-mtext(text = expression(paste("(c) ", Sigma, " matrix")), side=1, line=3.5, at=0, adj=0.5, cex=1.75)
+  scale = "none", trace = "none", col = colorpanel(50, "white", "red"),
+  colsep = 1:length(V(graph)), rowsep = 1:length(V(graph)))
+
 #simulated data
 expr <- generate_expression(100, graph, cor = 0.8, mean = 0,
-comm = FALSE, dist =TRUE, absolute = FALSE, state = state)
+  comm = FALSE, dist =TRUE, absolute = FALSE, state = state)
[…]
