+++ paper/paper.Rmd
@@ -229,13 +229,33 @@ commands in R:
 ```{r, eval = FALSE}
 #install packages required (once per machine)
 install.packages("graphsim")
+
 ```
-```{r, warning=FALSE, results='hide'}
+```{r, warning=FALSE, results='hide', message=FALSE}
 #load required packages (once per R instance)
-library("graphsim"); library("igraph"); library("gplots")
+library("graphsim")
+#load packages for examples
+library("igraph"); library("gplots")
 ```
 
-```{r simple_graph, fig.align='center', fig.show='hold', fig.width='1.0\\linewidth', fig.height='1.0\\linewidth', out.height='.415\\linewidth', out.width='.415\\linewidth', fig.retina=10, fig.margin = FALSE, fig.ncol = 2, fig.cap = '\\textbf{Simulated graph structures}. A constructed graph structure used as an example to demonstrate the simulation procedure in Figures 2 and 3. Activating links are denoted by black arrows and inhibiting links by red edges. Inhibiting edges have been highlighted in red.'}
+```{r simple_graph_hide, fig.align='center', fig.show='hold', fig.width='1.0\\linewidth', fig.height='1.0\\linewidth', out.height='.415\\linewidth', out.width='.415\\linewidth', fig.retina=10, fig.margin = FALSE, fig.ncol = 2, eval=FALSE}
+#generate graph structure
+graph_edges <- rbind(c("A", "C"), c("B", "C"), c("C", "D"), c("D", "E"),
+   c("D", "F"), c("F", "G"), c("F", "I"), c("H", "I"))
+graph <- graph.edgelist(graph_edges, directed = TRUE)
+
+#plot graph structure (Figure 1a)
+plot_directed(graph, state ="activating", layout = layout.kamada.kawai,
+  cex.node = 2, cex.arrow = 4, arrow_clip = 0.2)
+
+#generate parameters for inhibitions for each edge in E(graph)
+state <- c(1, 1, -1, 1, 1, 1, 1, -1, 1)
+#plot graph structure with inhibitions (Figure 1b)
+plot_directed(graph, state=state, layout = layout.kamada.kawai,
+  cex.node = 2, cex.arrow = 4, arrow_clip = 0.2)
+```
+
+```{r simple_graph, fig.align='center', fig.show='hold', fig.width='1.0\\linewidth', fig.height='1.0\\linewidth', out.height='.415\\linewidth', out.width='.415\\linewidth', fig.retina=10, fig.margin = FALSE, fig.ncol = 2, fig.cap = '\\textbf{Simulated graph structures}. A constructed graph structure used as an example to demonstrate the simulation procedure in Figures 2 and 3. Activating links are denoted by black arrows and inhibiting links by red edges. Inhibiting edges have been highlighted in red.', echo=FALSE}
 #generate graph structure
 graph_edges <- rbind(c("A", "C"), c("B", "C"), c("C", "D"), c("D", "E"),
    c("D", "F"), c("F", "G"), c("F", "I"), c("H", "I"))
@@ -247,7 +267,7 @@ plot_directed(graph, state ="activating", layout = layout.kamada.kawai,
 mtext(text = "(a) Activating pathway structure", side=1, line=3.5, at=0.05, adj=0.5, cex=1.75)
 box()
 
-#generate parameters for inhibitions
+#generate parameters for inhibitions for each edge in E(graph)
 state <- c(1, 1, -1, 1, 1, 1, 1, -1, 1)
 #plot graph structure with inhibitions (Figure 1b)
 plot_directed(graph, state=state, layout = layout.kamada.kawai,
@@ -296,7 +316,36 @@ graphics::layout(matrix(c(1:5, 5), nrow = 3, byrow = TRUE))
 set.seed(9000)
 ```
 
[…]
