### 2020-05-28 d05be77
reduce size of figures
