+++ paper/images.sh
@@ -1,1 +1,1 @@
-sed -i '/Plot.*-5/s/\(.*\)width=[.]415/\1width=.830/g' paper.md paper.tex
+sed -i '/Plot.*-5/s/\(.*\)width=[.]375/\1width=.750/g' paper.md paper.tex
