### 2020-05-28 acdf4b1
correct code style in paper examples
