+++ paper/paper.Rmd
@@ -329,7 +329,7 @@ heatmap.2(make_distance_graph(graph, absolute = FALSE),
   colsep = 1:length(V(graph)), rowsep = 1:length(V(graph)))
 
 #plot sigma matrix
-heatmap.2(make_sigma_mat_dist_graph(graph, 0.8, absolute = FALSE),
+heatmap.2(make_sigma_mat_dist_graph(graph, cor = 0.8, absolute = FALSE),
   scale = "none", trace = "none", col = colorpanel(50, "white", "red"),
   colsep = 1:length(V(graph)), rowsep = 1:length(V(graph)))
 
@@ -358,7 +358,7 @@ heatmap.2(make_distance_graph(graph, absolute = FALSE),
 colsep = 1:length(V(graph)), rowsep = 1:length(V(graph)))
 mtext(text = "(b) Relationship matrix", side=1, line=3.5, at=0, adj=0.5, cex=1.75)
 #plot sigma matrix
-heatmap.2(make_sigma_mat_dist_graph(graph, 0.8, absolute = FALSE),
+heatmap.2(make_sigma_mat_dist_graph(graph, cor = 0.8, absolute = FALSE),
 scale = "none", trace = "none", col = colorpanel(50, "white", "red"),
 colsep = 1:length(V(graph)), rowsep = 1:length(V(graph)))
 mtext(text = expression(paste("(c) ", Sigma, " matrix")), side=1, line=3.5, at=0, adj=0.5, cex=1.75)
@@ -431,8 +431,8 @@ heatmap.2(make_distance_graph(graph, absolute = FALSE),
   scale = "none", trace = "none", col = colorpanel(50, "white", "red"),
   colsep = 1:length(V(graph)), rowsep = 1:length(V(graph)))
 
-# #plot sigma matrix
-heatmap.2(make_sigma_mat_dist_graph(graph, state = state, 0.8, absolute = FALSE),
+#plot sigma matrix
+heatmap.2(make_sigma_mat_dist_graph(graph, state = state, cor = 0.8, absolute = FALSE),
   scale = "none", trace = "none", col = colorpanel(50, "blue", "white", "red"),
   colsep = 1:length(V(graph)), rowsep = 1:length(V(graph)))
 
@@ -442,7 +442,6 @@ expr <- generate_expression(100, graph, state = state, cor = 0.8, mean = 0,
 #plot simulated correlations
 heatmap.2(cor(t(expr)), scale = "none", trace = "none", col = colorpanel(50, "blue", "white", "red"),
   colsep = 1:length(V(graph)), rowsep = 1:length(V(graph)))
-
 #plot simulated expression data
 heatmap.2(expr, scale = "none", trace = "none", col = bluered(50),
   colsep = 1:length(V(graph)), rowsep = 1:length(V(graph)), labCol = "")
@@ -461,8 +460,8 @@ heatmap.2(make_distance_graph(graph, absolute = FALSE),
   scale = "none", trace = "none", col = colorpanel(50, "white", "red"),
 colsep = 1:length(V(graph)), rowsep = 1:length(V(graph)))
 mtext(text = "(b) Relationship matrix", side=1, line=3.5, at=0, adj=0.5, cex=1.75)
-# #plot sigma matrix
-heatmap.2(make_sigma_mat_dist_graph(graph, state = state, 0.8, absolute = FALSE),
+#plot sigma matrix
+heatmap.2(make_sigma_mat_dist_graph(graph, state = state, cor = 0.8, absolute = FALSE),
 scale = "none", trace = "none", col = colorpanel(50, "blue", "white", "red"),
 colsep = 1:length(V(graph)), rowsep = 1:length(V(graph)))
[…]
