### 2020-05-28 a469b0e
update code examples in paper
