### 2020-05-28 e301f2d
correct code examples for paper margins
