### 2020-05-28 #2161 @TomKellyGenetics
Quick update here. I don't have as much time to work on this as I'd like. I will try to get back to vignettes sometime soon.

I've removed the code for the subcaptions from the examples (these run hidden). This makes the Rmarkdown messy but I think it's good to refer to if you wish to reproduce the figures. Regarding length, I get a [preview](https://github.com/openjournals/joss-reviews/files/4694683/paper.pdf) from pandoc/LaTeX of 11 pages but the Whedon version is longer due to different figure placement. I've altered the code examples to fit within the margins of the JOSS style. I'll see what I can do about the figures, making them smaller doesn't seem to help. Possibly moving them to different pages would.
