### 2020-05-29 #2161 @majensen
@TomKellyGenetics I really like this paper, and the material there is helpful and good. One way to reduce the size might be to take the example section and refer to it by reference from the paper to place where it is rendered -- could be a document in the repo itself, or simply the command to display the vignette in all its glory.
