### 2020-06-01 473cc89
update formatting of blog post vignette
