+++ vignettes/plots_viral_quantitate.Rmd
@@ -64,7 +64,7 @@ plot(G, layout = layout.fruchterman.reingold,
      edge.color="black")
 ```
 
-##Building a simulation
+## Building a simulation
 
 The ease and flexibility with which you can assign and update attributes of a graph's nodes and edges makes igraph a powerful tool for prototyping simulations with graphs. Let's consider a very simplified example. Let G be a graph whose nodes are people in a population. Two nodes in G will share an edge if they have daily in-person contact with each other. Suppose also that a virus has infected a few nodes of G. These infected nodes can spread the virus to other uninfected nodes with whom they have daily contact. Newly infected nodes will then spread the virus to other nodes, and so on. Let's use the simple probability model:
 
@@ -133,7 +133,7 @@ library("animation")
 L <- layout.fruchterman.reingold(Glist[[1]])
 ani.options(interval=1)
 saveGIF({
-  count =0
+  count = 0
   for(i in 1:length(Glist)){
     plot(Glist[[i]], layout = L,
          vertex.label = NA,
@@ -155,7 +155,7 @@ This is a very simplistic model and it does not, for example, take into account 
 
 We should note, however, that although this implementation is very easy to understand, it is not very efficient. The heavy use of loops will be quite slow when we start working with large graphs. Though we might gain some efficiency through vectorization or other optimizations, for running larger simulations we would probably want to port this routine to another platform like C++. Nevertheless, the strength of R's igraph for quickly building and testing a prototype should be apparent from this example.
 
-##More realistic graphs
+## More realistic graphs
 
 So far, the graphs we have worked with have been quite simple. How do we start building more complex graphs? There are many great sources for real graph data like Stanford's SNAP dataset collection. There are also many routines for generating random graphs in the igraph packages. The degree.sequence.game method in igraph is one such example. The degree of a node in a graph is the number of edges attached to that node. One way to study graphs is to look at the distribution of the degrees of the nodes in the graph. For many interesting graphs, the degree distribution often has the form of an exponential distribution:
 
