### 2020-06-01 5a87c1c
remove blog post vignette for igraph simulations
