+++ NAMESPACE
@@ -1,5 +1,23 @@
 # Generated by roxygen2: do not edit by hand
 
+export(generate_expression)
+export(generate_expression_mat)
+export(make_adjmatrix_graph)
+export(make_commonlink_adjmat)
+export(make_commonlink_graph)
+export(make_distance_adjmat)
+export(make_distance_graph)
+export(make_distance_laplacian)
+export(make_laplacian_adjmat)
+export(make_laplacian_graph)
+export(make_sigma_mat_adjmat)
+export(make_sigma_mat_comm)
+export(make_sigma_mat_dist_adjmat)
+export(make_sigma_mat_dist_graph)
+export(make_sigma_mat_graph)
+export(make_sigma_mat_laplacian)
+export(make_state_matrix)
+export(plot_directed)
 import(graphics)
 import(igraph)
 importClassesFrom(Matrix,dgCMatrix)
