### 2020-06-01 faea482
Expand explanation of plotting functions in Vignette and add empirical datasets
