+++ vignettes/plots_directed.Rmd
@@ -1,5 +1,5 @@
 ---
-title: "Directed Plots in iGraph"
+title: "Directed plots for igraph objects"
 author: "Tom Kelly"
 date: "`r Sys.Date()`"
 output: rmarkdown::html_vignette
@@ -9,59 +9,95 @@ vignette: >
   %\VignetteEncoding{UTF-8}
 ---
 
-## Install
+```{r, include = FALSE}
+knitr::opts_chunk$set(collapse = TRUE, comment = "#>", width = 68)
+knitr::opts_chunk$set(fig.cap = "", fig.path = "Plot")
+knitr::opts_chunk$set(fig.align = "center")
+options(width = 68, cli.unicode = FALSE, cli.width = 68)
+#par(mai=c(2.82, 2.82, 0.82, 0.82)-0.82)
+par(mar=c(7, 10, 4, 2) + 0.1)
+par(bty="o")
+#captioner::captioner(prefix = "Fig.")
+```
+
+## Motivations for directed graph plots
+
+Here we demonstrate the plotting functions that come built-in with `graphsim` and compare them them to the `plot.igraph` method provided by the `igraph` package. With all due respect to the efforts of the `igraph` developers, their plotting functions leave some things to be desired in terms of aesthetics. Here we provide additional functionality to plot directed graphs. This draws upon many functions provided by `igraph` including layout settings.
+
+In particular, graph and network represenations in biology often require displaying edge properties. Here we have the "state" parameter which can be used to differentiate these and plot activating and inhibiting or repressing relationships differently. We use different arrowheads as per convention in biochemistry and molecular genetics. 
+
+## Import packages
+
+To display these plots the following packages must be imported.
 
 ```{r, message=F}
 library("igraph")
 library("graphsim")
 ```
 
-##Set up simulated graph
+## Toy example
+
+Here we demonstrate the plot functions on a small toy graph.
+
+### Set up simulated graph
 
 ```{r}
 graph_edges <- rbind(c("A", "C"), c("B", "C"), c("C", "D"), c("D", "E"), c("D", "F"), c("F", "G"), c("F", "I"), c("H", "I"))
 graph <- graph.edgelist(graph_edges, directed = T)
 ```
 
-##Plotting
+### Plotting
 
 `plot_directed` with default settings uses the `layout.fruchterman.reingold` as does the default `igraph::graphsim`.
 
-```{r}
+```{r,  fig.align='center', out.width="80%", fig.height = 10, fig.width = 10}
 plot(graph)
+```
+```{r,  fig.align='center', out.width="80%", fig.height = 5, fig.width = 5}
 plot_directed(graph)
 ```
 
+Here you can see that the plotting functions display graphs in a similar layout with different aesthetic parameters. We suggest that you choose the function that suits your needs and datasets.
+
+#### Custom aesthetics
+
+We support various aesthetic parameters to control the colour and relative size of nodes and edges.
+
 `plot_directed` supports customised layouts and colours:
 
-```{r}
+```{r,  fig.align='center', out.width="80%", fig.height = 5, fig.width = 5}
 plot_directed(graph, layout = layout.kamada.kawai)
 plot_directed(graph, fill.node = "lightblue", border.node = "royalblue")
 ```
 
+#### Vectorisation
+
 Colours may also be entered as a vector for each node in `V(graph)`:
 
-```{r}
+```{r,  fig.align='center', out.width="80%", fig.height = 5, fig.width = 5}
 names(V(graph))
 colour_vector <- ifelse(names(V(graph)) %in% c("A", "D", "I"), 1, 2)
[…]
