### 2020-06-01 ef96c61
update vignette metadata for plotting
