### 2020-06-01 5ffa760
create centralised vignette for toy examples
