+++ vignettes/simulate_expression.Rmd
@@ -1,0 +1,203 @@
+---
+title: "Simulating continuous data based on network graph structures"
+author: "Tom Kelly"
+date: "`r Sys.Date()`"
+output: rmarkdown::html_document
+vignette: >
+  %\VignetteIndexEntry{Simulating network graph structure in continuous data}
+   %\VignetteEngine{knitr::knitr}
+  %\VignetteEncoding{UTF-8}
+---
+
+# Overview of graphsim
+
+This package is designed to balance user-friendliness (by providing sensible defaults and built-in functions) and flexbility (many options are available to use as needed). It is likely that you will disagree with the parameters that I've used to simulate datasets. That's a valid opinion and there are many options at your disposal to alter parameters to use these functions for your purposes.
+
+If you have issues or feedback, sumbmission to the the GitHub repository are welcome. See the DESCRIPTION and README.md for more details on how to suggest changes to the package.
+
+
+## Motivations
+
+Pathway and graph structures have a wide array of applications. Here will consider the simulation of (log-normalised) gene expression data from genomics experiments in a biological pathway. If you have another use for this software you are welcome to apply it to your problem, please bear in mind that it was designed with this application in mind however. In principle, normally-distributed continuous data can be generated based on any defined relationships. This uses the graph structure to define a ∑ covariance matrix and generate simulated data by sampling from a multivariate normal distribution.
+
+Crucially, this allows the simulation of negative correlations based on inhibitory or repressive relationships, as commonly occur in biology. A custom plotting function `plot_directed` is provided to visualise these relationships with the "state" parameter. This plotting function has a dedicated vignette on [plotting](plots_directed.Rmd).
+
+For more details on the background of this package, see the [paper](../paper/paper.Rmd) included with the package on GitHub. This vignette provides more detail on the code needed to reproduce the figures in this manuscript.
+  
+# Getting Started
+  
+## Install dependencies for Demonstration
+  
+  First we must load the packages required to use the pacakge functions with igraph objects and generate plots.
+
+```{r, message=F, warning=F}
+library("igraph")
+library("gplots")
+library("graphsim")
+```
+
+## Set up simulated graphs
+
+Here we set up a simple graph to demonstrate how connections in the graph structure lead to correlations in the final output. We create a simple pathway of 9 genes with various branches.
+
+
+```{r, out.width = '50%', out.height  = '50%', fig.align='center', dpi=36}
+graph_test4_edges <- rbind(c("A", "C"), c("B", "C"), c("C", "D"), c("D", "E"), c("D", "F"), c("F", "G"), c("F", "I"), c("H", "I"))
+graph_test4 <- graph.edgelist(graph_test4_edges, directed = T)
+plot_directed(graph_test4, layout = layout.kamada.kawai)
+```
+
+# Generating  simulated expression data from graph
+
+## Minimal example 
+
+A simulated dataset can be generated with a single command. This is all that is required to get started.
+
+```{r, out.width = '50%', out.height  = '50%', fig.align='center', dpi=36}
+expr <- generate_expression(100, graph_test4, cor = 0.8, mean = 0, comm = F, dist = T, absolute = F)
+heatmap.2(expr, scale = "none", trace = "none", col = bluered(50), colsep = 1:4, rowsep = 1:4)
[…]
