### 2020-06-01 6063aef
update network vignette with overview
