+++ vignettes/simulate_expression.Rmd
@@ -9,6 +9,16 @@ vignette: >
   %\VignetteEncoding{UTF-8}
 ---
 
+```{r, include = FALSE}
+#knitr::opts_chunk$set(collapse = TRUE, comment = "#>", width = 68)
+knitr::opts_chunk$set(fig.cap = "", fig.path = "Plot")
+options(width = 68, cli.unicode = FALSE, cli.width = 68)
+#par(mai=c(2.82, 2.82, 0.82, 0.82)-0.82)
+par(mar=c(7, 10, 4, 2) + 0.1)
+par(bty="o")
+#captioner::captioner(prefix = "Fig.")
+```
+
 # Overview of graphsim
 
 This package is designed to balance user-friendliness (by providing sensible defaults and built-in functions) and flexbility (many options are available to use as needed). It is likely that you will disagree with the parameters that I've used to simulate datasets. That's a valid opinion and there are many options at your disposal to alter parameters to use these functions for your purposes.
@@ -27,10 +37,26 @@ For more details on the background of this package, see the [paper](../paper/pap
 # Getting Started
   
 ## Install dependencies for Demonstration
-  
-  First we must load the packages required to use the pacakge functions with igraph objects and generate plots.
 
-```{r, message=F, warning=F}
+The package can be installed as follows. Run the following command to install the current release from CRAN (recommended).
+
+```{r, eval = FALSE}
+#install packages required (once per machine)
+install.packages("graphsim")
+```
+
+Run the following command to install the development version from GitHub (advanced users). This will import the latest changes to the package ahead of releasing updates, behaviour may be unstable.
+
+```{r, eval = FALSE}
+#install stable release
+remotes::install_github("TomKellyGenetics", ref = "master")
+#install development version
+remotes::install_github("TomKellyGenetics", ref = "dev")
+```
+
+Once the required packages are installed, we must load the packages required to use the package functions with igraph objects and generate plots. Here `igraph` is required to create `igraph` objects and `gplots` is required for plotting heatmaps.
+
+```{r, warning=FALSE, results='hide', message=FALSE}
 library("igraph")
 library("gplots")
 library("graphsim")
@@ -42,9 +68,9 @@ Here we set up a simple graph to demonstrate how connections in the graph struct
 
 
 ```{r, out.width = '50%', out.height  = '50%', fig.align='center', dpi=36}
-graph_test4_edges <- rbind(c("A", "C"), c("B", "C"), c("C", "D"), c("D", "E"), c("D", "F"), c("F", "G"), c("F", "I"), c("H", "I"))
-graph_test4 <- graph.edgelist(graph_test4_edges, directed = T)
-plot_directed(graph_test4, layout = layout.kamada.kawai)
+graph_structure_edges <- rbind(c("A", "C"), c("B", "C"), c("C", "D"), c("D", "E"), c("D", "F"), c("F", "G"), c("F", "I"), c("H", "I"))
+graph_structure <- graph.edgelist(graph_structure_edges, directed = T)
+plot_directed(graph_structure, layout = layout.kamada.kawai)
 ```
 
 # Generating  simulated expression data from graph
@@ -53,77 +79,79 @@ plot_directed(graph_test4, layout = layout.kamada.kawai)
 
 A simulated dataset can be generated with a single command. This is all that is required to get started.
 
-```{r, out.width = '50%', out.height  = '50%', fig.align='center', dpi=36}
[…]
