+++ vignettes/test_graph_network.Rmd
@@ -1,188 +1,0 @@
----
-title: "Simulating network graph structure in continuous data"
-author: "Tom Kelly"
-date: "`r Sys.Date()`"
-output: rmarkdown::html_document
-vignette: >
-  %\VignetteIndexEntry{Simulating network graph structure in continuous data}
-   %\VignetteEngine{knitr::knitr}
-  %\VignetteEncoding{UTF-8}
----
-  
-  # Getting Started
-  
-  ## Install dependencies for Demonstration
-  
-  First we must load the packages required to use the pacakge functions with igraph objects and generate plots.
-
-```{r, message=F, warning=F}
-library("igraph")
-library("gplots")
-library("graphsim")
-```
-
-## Set up simulated graphs
-
-Here we set up a simple graph to demonstrate how connections in the graph structure lead to correlations in the final output. We create a simple pathway of 9 genes with various branches.
-
-
-```{r, out.width = '50%', out.height  = '50%', fig.align='center', dpi=36}
-graph_test4_edges <- rbind(c("A", "C"), c("B", "C"), c("C", "D"), c("D", "E"), c("D", "F"), c("F", "G"), c("F", "I"), c("H", "I"))
-graph_test4 <- graph.edgelist(graph_test4_edges, directed = T)
-plot_directed(graph_test4, layout = layout.kamada.kawai)
-```
-
-# Generating  simulated expression data from graph
-
-## Minimal example 
-
-A simulated dataset can be generated with a single command. This is all that is required to get started.
-
-```{r, out.width = '50%', out.height  = '50%', fig.align='center', dpi=36}
-expr <- generate_expression(100, graph_test4, cor = 0.8, mean = 0, comm = F, dist = T, absolute = F)
-heatmap.2(expr, scale = "none", trace = "none", col = bluered(50), colsep = 1:4, rowsep = 1:4)
-```
-
-Various arguments are supported to alter how the simulated datasets are computed. 
-
-# How it works step-by-step
-
-Here we show the data generated by for this graph structure. This demonstrates how several of the options available compute the necessary steps.
-
-## Adjacency matrix
-
-The data can be summarised by an "adjacency matrix" where a one (1) is given between a row `i` and column `j` if there is an edge between genes `i` and `j`. Otherise it is a zero (0) for genes that are not connected. For an undirected graph, edges are shown in a symmetical matrix.
-
-```{r, out.width = '50%', out.height  = '50%', fig.align='center', dpi=36}
-adj_mat <- make_adjmatrix_graph(graph_test4)
-heatmap.2(make_adjmatrix_graph(graph_test4), scale = "none", trace = "none", col = colorpanel(3, "grey75", "white", "blue"), colsep = 1:4, rowsep = 1:4)
-```
-
-For a directed graph, the edges are not shown from gene (row) `i` and gene (column) `j` 
-
-```{r, out.width = '50%', out.height  = '50%', fig.align='center', dpi=36}
-heatmap.2(make_adjmatrix_graph(graph_test4, directed = T), scale = "none", trace = "none", col = colorpanel(3, "grey75", "white", "blue"), colsep = 1:4, rowsep = 1:4)
-```
-
-We can compute the common links between each pair of genes. This shows how many genes are connected to both genes `i` and `j`.
-
-```{r, out.width = '50%', out.height  = '50%', fig.align='center', dpi=36}
-comm_mat <- make_commonlink_graph(graph_test4)
-heatmap.2(make_commonlink_graph(graph_test4), scale = "none", trace = "none", col = colorpanel(50, "grey75", "red"), colsep = 1:4, rowsep = 1:4)
-```
-
[…]
