### 2020-06-01 84100b8
combine toy examples into full centralised vignette
