### 2020-06-01 #2161 @TomKellyGenetics
Thanks @majensen and all for the feedback, that is really encouraging. I've begun to revise the vignettes (see https://github.com/TomKellyGenetics/graphsim/commit/84100b88036287aee9a7db1497c2249107132da9 for details). The next step will be to add examples with the empirical datasets provided in `data/` (I've done this for the plotting function already https://github.com/TomKellyGenetics/graphsim/commit/faea482b8d9e3a3a0af2a9d791060595ded67051). Once the vignettes are revised, I think they would be suitable to refer (or link) to for detailed (reproducible) examples as you suggest.

Sorry for the delays on this. Now that the vignettes have been restructured, hopefully it won't take much longer to complete them and updates the examples in the documentation.
