### 2020-06-02 14fe93f
update handling of duplicate edges in state parameter
