+++ man/Pi3K_AKT_graph.Rd
@@ -4,12 +4,14 @@
 \name{Pi3K_AKT_graph}
 \alias{Pi3K_AKT_graph}
 \title{PI3K/AKT activation}
-\format{A graph object of 275 vertices and 21106 edges:
+\format{
+A graph object of 275 vertices and 21106 edges:
 \describe{
   \item{V}{gene symbol (human)}
   \item{E}{directed relationship for pathway}
   \item{state}{type of relationship (activating or inhibiting) as edge attribute}
-}}
+}
+}
 \source{
 PathwayCommons \url{https://reactome.org/content/detail/R-HSA-198203}
 }
