+++ man/Pi3K_graph.Rd
@@ -4,12 +4,14 @@
 \name{Pi3K_graph}
 \alias{Pi3K_graph}
 \title{PI3K Cascade}
-\format{A graph object of 35 vertices and 251 edges:
+\format{
+A graph object of 35 vertices and 251 edges:
 \describe{
   \item{V}{gene symbol (human)}
   \item{E}{directed relationship for pathway}
   \item{state}{type of relationship (activating or inhibiting) as edge attribute}
-}}
+}
+}
 \source{
 PathwayCommons \url{https://reactome.org/content/detail/R-HSA-109704}
 }
