+++ man/RAF_MAP_graph.Rd
@@ -4,11 +4,13 @@
 \name{RAF_MAP_graph}
 \alias{RAF_MAP_graph}
 \title{#' RAF/MAP kinase cascade}
-\format{A graph object of 17 vertices and 121 edges:
+\format{
+A graph object of 17 vertices and 121 edges:
 \describe{
   \item{V}{gene symbol (human)}
   \item{E}{directed relationship for pathway}
-}}
+}
+}
 \source{
 PathwayCommons \url{https://reactome.org/content/detail/R-HSA-5673001}
 }
