### 2020-06-02 0e1a168
consolidate vignettes #11 and expand descriptions of plots #13
