+++ vignettes/simulate_expression.Rmd
@@ -1,7 +1,7 @@
 ---
 title: "Simulating continuous data based on network graph structures"
 author: "Tom Kelly"
-date: "`r Sys.Date()`"
+date: "`r  format(Sys.time(), '%A %d %B %Y')`"
 output: rmarkdown::html_document
 vignette: >
   %\VignetteIndexEntry{Simulating network graph structure in continuous data}
@@ -60,6 +60,7 @@ Once the required packages are installed, we must load the packages required to 
 library("igraph")
 library("gplots")
 library("graphsim")
+library("scales")
 ```
 
 ## Set up simulated graphs
@@ -309,7 +310,7 @@ mtext(text = "(e) Simulated expression data (log scale)", side=1, line=3.5, at=0
 
 Here we give some toy examples to show how the simulations behave in simple cases. This serves to understand how modules within a larger graph will translate to correlations in the final simulated datasets.
 
-## Diverging branches
+### Diverging branches
 
 ```{r simulation_graph_diverging_hide, fig.show='hold', out.width = '50%', out.height  = '50%', fig.align='center', dpi=36, fig.retina=10, fig.margin = FALSE, fig.ncol = 2, warning=FALSE, message=FALSE}
 graph_diverging_edges <- rbind(c("A", "B"), c("B", "C"), c("B", "D"))
@@ -347,7 +348,7 @@ mtext(text = "(e) Simulated expression data (log scale)", side=1, line=3.5, at=0
 
 ```
 
-### Inhibiting relationships
+#### Inhibiting relationships
 
 Here we simulate the same graph structure with inhibiting edges but passing the `"state"` parameter. This takes one argument for each each to identify which are inhibitory (as documented).
 
@@ -384,7 +385,7 @@ mtext(text = "(e) Simulated expression data (log scale)", side=1, line=3.5, at=0
 
 ```
 
-## Converging branches
+### Converging branches
 
 
 ```{r simulation_graph_converging_hide, fig.show='hold', out.width = '50%', out.height  = '50%', fig.align='center', dpi=36, fig.retina=10, fig.margin = FALSE, fig.ncol = 2, warning=FALSE, message=FALSE}
@@ -423,7 +424,7 @@ mtext(text = "(e) Simulated expression data (log scale)", side=1, line=3.5, at=0
 
 ```
 
-### Inhibiting relationships
+#### Inhibiting relationships
 
 Here we simulate the same graph structure with inhibiting edges but passing the `"state"` parameter. This takes one argument for each each to identify which are inhibitory (as documented).
 
@@ -460,7 +461,7 @@ mtext(text = "(e) Simulated expression data (log scale)", side=1, line=3.5, at=0
 
 ```
 
-## Reconnecting paths
+### Reconnecting paths
 
 
 ```{r simulation_graph_reconnecting_hide, fig.show='hold', out.width = '50%', out.height  = '50%', fig.align='center', dpi=36, fig.retina=10, fig.margin = FALSE, fig.ncol = 2, warning=FALSE, message=FALSE}
@@ -499,7 +500,7 @@ mtext(text = "(e) Simulated expression data (log scale)", side=1, line=3.5, at=0
 
 ```
 
-### Inhibiting relationships
+#### Inhibiting relationships
 
 Here we simulate the same graph structure with inhibiting edges but passing the `"state"` parameter. This takes one argument for each each to identify which are inhibitory (as documented).
 
[…]
