### 2020-06-02 f296620
check state of edges in opposite directions
