### 2020-06-02 66ba7f8
correct passing state as edge attributes
