### 2020-06-02 fcb3b78
document laplacian
