### 2020-06-02 2c17903
correct plotting parameters for pathway heatmaps for vignette
