+++ vignettes/simulate_expression.Rmd
@@ -224,6 +224,11 @@ heatmap.2(cor(t(expr)), scale = "none", trace = "none", col = colorpanel(50, "gr
 
 Here we use absolute distance (relationships are arithmetrically weighted to the diameter).
 
+```{r, include = FALSE}
+set.seed(9000)
+```
+
+
 ```{r, out.width = '50%', out.height  = '50%', fig.align='center', dpi=36}
 #absolute dist
 expr <- generate_expression(100, graph_structure, cor = 0.8, mean = 0, comm = F, dist = T, absolute = T) # unable to generate from adj mat ## fixed PD
@@ -236,7 +241,12 @@ heatmap.2(cor(t(expr)), scale = "none", trace = "none", col = bluered(50), colse
 
 In summary, we compute the following expression dataset but on these underlying relationships in the graph structure. Here we use geometrically decreasing correlations between more distant nodes in the network.
 
-```{r simulation_activating_hide, fig.show='hold', out.width = '50%', out.height  = '50%', fig.align='center', dpi=36, fig.retina=10, fig.margin = FALSE, fig.ncol = 2, warning=FALSE, message=FALSE}
+```{r, include = FALSE}
+set.seed(9000)
+```
+
+
+```{r simulation_activating_hide, fig.show='hold', out.width = '50%', out.height  = '50%', fig.align='center', dpi=36, fig.retina=10, fig.margin = FALSE, fig.ncol = 2, warning=FALSE, message=FALSE, cache=TRUE}
 # activating graph
 state <- rep(1, length(E(graph_structure)))
 plot_directed(graph_structure, state=state, layout = layout.kamada.kawai,
@@ -273,7 +283,12 @@ mtext(text = "(e) Simulated expression data (log scale)", side=1, line=3.5, at=0
 
 Here we simulate the same graph structure with inhibiting edges but passing the `"state"` parameter. This takes one argument for each each to identify which are inhibitory (as documented).
 
-```{r simulation_inhibiting_hide, fig.show='hold', out.width = '50%', out.height  = '50%', fig.align='center', dpi=36, fig.retina=10, fig.margin = FALSE, fig.ncol = 2, warning=FALSE, message=FALSE}
+```{r, include = FALSE}
+set.seed(9000)
+```
+
+
+```{r simulation_inhibiting_hide, fig.show='hold', out.width = '50%', out.height  = '50%', fig.align='center', dpi=36, fig.retina=10, fig.margin = FALSE, fig.ncol = 2, warning=FALSE, message=FALSE, cache=TRUE}
 # activating graph
 state <- state <- c(1, 1, -1, 1, 1, 1, 1, -1)
 plot_directed(graph_structure, state=state, layout = layout.kamada.kawai,
@@ -312,7 +327,12 @@ Here we give some toy examples to show how the simulations behave in simple case
 
 ### Diverging branches
 
-```{r simulation_graph_diverging_hide, fig.show='hold', out.width = '50%', out.height  = '50%', fig.align='center', dpi=36, fig.retina=10, fig.margin = FALSE, fig.ncol = 2, warning=FALSE, message=FALSE}
+```{r, include = FALSE}
+set.seed(9000)
+```
+
+
+```{r simulation_graph_diverging_hide, fig.show='hold', out.width = '50%', out.height  = '50%', fig.align='center', dpi=36, fig.retina=10, fig.margin = FALSE, fig.ncol = 2, warning=FALSE, message=FALSE, cache=TRUE}
[…]
