### 2020-06-02 de2fcf6
update paper figures
