+++ paper/paper.Rmd
@@ -27,7 +27,7 @@ affiliations:
    index: 1
  - name: "RIKEN Center for Integrative Medical Sciences, Suehiro-cho-1-7-22, Tsurumi Ward, Yokohama, Kanagawa 230-0045, Japan"
    index: 2
-date: 30 April 2020
+date: "`r  format(Sys.time(), '%d %B %Y')`"
 bibliography: paper.bib
 header-includes:
   - \usepackage{caption}
