### 2020-06-02 TomKellyGenetics/graphsim#6 @TomKellyGenetics
Dependency on the `animation` package has been removed. Examples with empirical data are given in the vignettes. I'll relevant examples like this to the documentation `#' @usage` section if CRAN allows larger computations.
