### 2020-06-02 TomKellyGenetics/graphsim#11 @TomKellyGenetics
Vignettes have substantially revised. It's still a work in progress in terms of figures formatting but the code examples are working as expected. I'll update the documentation to address #6  and #13 using similar examples to this.
