### 2020-06-04 1315a7b
add citations to introduction
