+++ paper/paper.bib
@@ -235,6 +235,43 @@ eprint = {/oup/backfile/content_public/journal/bioinformatics/24/7/10.1093_bioin
    Month="Feb"
 }
 
+@article {Cannoodt2020,
+	author = {Cannoodt, Robrecht and Saelens, Wouter and Deconinck, Louise and Saeys, Yvan},
+	title = {dyngen: a multi-modal simulator for spearheading new single-cell omics analyses},
+	elocation-id = {2020.02.06.936971},
+	year = {2020},
+	doi = {10.1101/2020.02.06.936971},
+	publisher = {Cold Spring Harbor Laboratory},
+	abstract = {Purpose When developing new types of tools for single-cell analyses, there is often a lack of datasets on which to quantitatively assess the performance.Results We developed dyngen, a multi-modality simulator of single cells. In dyngen, the biomolecular state of an in silico changes over time according to a predefined gene regulatory network. We used dyngen to benchmark three emerging ways of analysing single-cell data: RNA velocity, cell-specific network inference and trajectory alignment methods.Conclusion dyngen lays the foundations for benchmarking a wide variety of computational single-cell tools and can be used to help kick-start the development of future types of analyses.},
+	URL = {https://www.biorxiv.org/content/early/2020/02/27/2020.02.06.936971},
+	eprint = {https://www.biorxiv.org/content/early/2020/02/27/2020.02.06.936971.full.pdf},
+	journal = {bioRxiv}
+}
+
+
+@Article{Schaffter2011,
+   Author="Schaffter, T.  and Marbach, D.  and Floreano, D. ",
+   Title="{{G}ene{N}et{W}eaver: in silico benchmark generation and performance profiling of network inference methods}",
+   Journal="Bioinformatics",
+   Year="2011",
+   Volume="27",
+   Number="16",
+   Pages="2263--2270",
+   Month="Aug"
+}
+
+@article {Dibaeinia2019,
+	author = {Dibaeinia, Payam and Sinha, Saurabh},
+	title = {A single-cell expression simulator guided by gene regulatory networks},
+	elocation-id = {716811},
+	year = {2019},
+	doi = {10.1101/716811},
+	publisher = {Cold Spring Harbor Laboratory},
+	abstract = {A common approach to benchmarking of single-cell transcriptomics tools is to generate synthetic data sets that resemble experimental data in their statistical properties. However, existing single-cell simulators do not incorporate known principles of transcription factor-gene regulatory interactions that underlie expression dynamics. Here we present SERGIO, a simulator of single-cell gene expression data that models the stochastic nature of transcription as well as linear and non-linear influences of multiple transcription factors on genes according to a user-provided gene regulatory network. SERGIO is capable of simulating any number of cell types in steady-state or cells differentiating to multiple fates according to a provided trajectory, reporting both unspliced and spliced transcript counts in single-cells. We show that data sets generated by SERGIO are comparable with experimental data in terms of multiple statistical measures. We also illustrate the use of SERGIO to benchmark several popular single-cell analysis tools, including GRN inference methods.},
+	URL = {https://www.biorxiv.org/content/early/2019/07/28/716811},
+	eprint = {https://www.biorxiv.org/content/early/2019/07/28/716811.full.pdf},
+	journal = {bioRxiv}
+}
 
 
 @Article{Shimamura2009,
@@ -279,3 +316,16 @@ url="https://doi.org/10.1186/1752-0509-3-41"
    Pages="37--46",
    Month="Jan"
 }
+
+@Article{Zappia2017,
+   Title="{S}platter: simulation of single-cell {RNA} sequencing data}",
[…]
