### 2020-06-10 4dd2e74
update intro and discussion
