+++ paper/paper.bib
@@ -273,6 +273,18 @@ eprint = {/oup/backfile/content_public/journal/bioinformatics/24/7/10.1093_bioin
 	journal = {bioRxiv}
 }
 
+@Article{Saelens2019,
+   Author="Saelens, W.  and Cannoodt, R.  and Todorov, H.  and Saeys, Y. ",
+   Title="{{A} comparison of single-cell trajectory inference methods}",
+   Journal="Nat. Biotechnol.",
+   Year="2019",
+   Volume="37",
+   Number="5",
+   Pages="547--554",
+   Month="05"
+}
+
+
 
 @Article{Shimamura2009,
 author="Shimamura, Teppei
