+++ paper/paper.Rmd
@@ -101,14 +101,37 @@ using gene regulatory modules in state-space models and recursive vector
 autoregressive models [@Hirose2008; @Shimamura2009]. Various approaches
 to gene regulation and networks at the genome-wide scale have led to
 novel biological insights [@Arner2015; @Komatsu2013]. However, inference
-of regulatory networks has thus far relied on experimental validation or
-resampling-based approaches to estimate the likelihood of specific
-network modules being predicted [@Markowetz2007; @Hawe2019].
+of regulatory networks has thus far primarily relied on experimental
+validation or resampling-based approaches to estimate the likelihood
+of specific network modules being predicted [@Markowetz2007; @Hawe2019].
 
+
+Simulated datasets with pathway structure are of particular interest for
+benchmarking regulatory network inference techniques and methods
+being developed for genomics data with complex biological 
+interactions [@Schaffter2011, @Saelens2019]. Dynamical models using differential equations have been employed,
+such as by GeneNetWeaver [@Schaffter2011], to generate simulated datasets
+specifically for benchmarking gene regulatory network inference techniques.
+There is renewed interest in modelling biological pathways and simulating
+data for benchmarking due to the emergence of single-cell genomics
+technologies and the plethora of bioinformatics techniques developed
+to use this data [@Zappia2017, @Saelens2019].
+
+Splatter [@Zappia2017] was developed to model single-cell data and 
+SERGIO [@Dibaeinia2019] and dyngen [@Cannoodt2020] build on this to
+include gene regulatory networks and multimodality respectively.
+
+These methods have been designed based on known or modelled reaction
+states and added stochasticity to determinisitic relationships.
+However, it is computationally-intensive to model these reactions
+at scale over iterations in benchmarking. In some cases, it is only
+necessary to model the statistical variability of RNA-Seq data, 
 
 GeneNetWeaver @Schaffter2011 - benchmarking of gene regulatory network inference methods
 Splatter @Zappia2017 - simulated single-cell RNA-Seq data (gamma-poisson)
+
 dyngen @Cannoodt2020 - single-cell multimodal -  Gillespie’s stochastic simulation at single-molecule level
+
 SERGIO @Dibaeinia2019 - single-cell gene expression (noise with poisson)
 
 
@@ -128,7 +151,7 @@ can be used to evaluate various bioinformatics methodologies, including
 statistical and network inference procedures.
 
 Methodology and software {#sec:methods}
-========================
+========================d
 
 Here we present a procedure to simulate gene expression data with
 correlation structure derived from a known graph structure. This
@@ -630,16 +653,20 @@ molecular biology. In order to translate findings from genomics studies
 into real-world applications such as improved healthcare, the roles of
 genes must be studied in the context of molecular pathways. Here we
 present a statistical framework to simulate gene expression from
-biological pathways, and provide the \texttt{graphsim} package in R to generate these
-simulated datasets. This approach is versatile and can be fine-tuned for
-modelling existing biological pathways or for testing whether
-constructed pathways can be detected by other means. In particular,
-methods to infer biological pathways and gene regulatory networks from
-gene expression data can be tested on simulated datasets using this
-framework. The package also enables simulation of complex gene
+biological pathways, and provide the \texttt{graphsim} package in R 
+to generate these simulated datasets. This approach is versatile and
+can be fine-tuned for modelling existing biological pathways or for
+testing whether constructed pathways can be detected by other means.
+In particular, methods to infer biological pathways and gene regulatory
+networks from gene expression data can be tested on simulated datasets
+using this framework. The package also enables simulation of complex gene
 expression datasets to test how these pathways impact on statistical
 analysis of gene expression data using existing methods or novel
[…]
