### 2020-06-10 e254b9d
vignette output
