### 2020-06-10 TomKellyGenetics/graphsim#13 @TomKellyGenetics
As suggested by the editors, removing some code examples from the paper would reduce the length. I like the Rmarkdown format though so I will keep it raw Rmd file to refer to reproduce the figures. I've also put similar examples in the [updated vignettes](https://github.com/TomKellyGenetics/graphsim/blob/66ba7f8e5e8e2f13784079338f44f9dc87eb58ff/vignettes/simulate_expression.Rmd). I hope this is acceptable to reduce the length of the paper while providing examples of the code. I've removed messages printed from the `library` calls as recommended.

>#> Warning in eattrs[[name]][index] <- value: number of items to
>#> replace is not a multiple of replacement length
>#> Warning in state_path[jj] <- state[kk]: number of items to replace
>#> is not a multiple of replacement length

I ran into this warning when computing the vignettes on empirical graphs from Reactome, it has been resolved [here](https://github.com/TomKellyGenetics/graphsim/blob/66ba7f8e5e8e2f13784079338f44f9dc87eb58ff/R/state.R#L164-L184).

I'm still updating the paper and documentation but hopefully it won't take much longer to have a revised version ready.
