### 2020-06-10 29f7973
update documentation with examples of biological networks #6
