+++ paper/paper.Rmd
@@ -272,7 +272,7 @@ plot_directed(graph, state ="activating", layout = layout.kamada.kawai,
   cex.node = 2, cex.arrow = 4, arrow_clip = 0.2)
 
 #generate parameters for inhibitions for each edge in E(graph)
-state <- c(1, 1, -1, 1, 1, 1, 1, -1, 1)
+state <- c(1, 1, -1, 1, 1, 1, 1, -1)
 #plot graph structure with inhibitions (Figure 1b)
 plot_directed(graph, state=state, layout = layout.kamada.kawai,
   cex.node = 2, cex.arrow = 4, arrow_clip = 0.2)
@@ -291,7 +291,7 @@ mtext(text = "(a) Activating pathway structure", side=1, line=3.5, at=0.05, adj=
 box()
 
 #generate parameters for inhibitions for each edge in E(graph)
-state <- c(1, 1, -1, 1, 1, 1, 1, -1, 1)
+state <- c(1, 1, -1, 1, 1, 1, 1, -1)
 #plot graph structure with inhibitions (Figure 1b)
 plot_directed(graph, state=state, layout = layout.kamada.kawai,
   cex.node = 2, cex.arrow = 4, arrow_clip = 0.2)
