+++ R/plot_directed.R
@@ -55,30 +55,31 @@
 ##'
 ##' #generate example graphs
 ##' library("igraph")
-##' graph_test4_edges <- rbind(c("A", "C"), c("B", "C"), c("C", "D"), c("D", "E"),
+##' graph_structure_edges <- rbind(c("A", "C"), c("B", "C"), c("C", "D"), c("D", "E"),
 ##'                            c("D", "F"), c("F", "G"), c("F", "I"), c("H", "I"))
-##' graph_test4 <- graph.edgelist(graph_test4_edges, directed = TRUE)
+##' graph_structure <- graph.edgelist(graph_structure_edges, directed = TRUE)
 ##'
 ##' #plots with igraph defaults
-##' plot(graph_test4, layout = layout.fruchterman.reingold)
-##' plot(graph_test4, layout = layout.kamada.kawai)
+##' plot(graph_structure, layout = layout.fruchterman.reingold)
+##' plot(graph_structure, layout = layout.kamada.kawai)
 ##'
 ##' #plots with scalar states
-##' plot_directed(graph_test4, state="activating")
-##' plot_directed(graph_test4, state="inhibiting")
+##' plot_directed(graph_structure, state="activating")
+##' plot_directed(graph_structure, state="inhibiting")
 ##'
 ##' #plots with vector states
-##' plot_directed(graph_test4, state=c(1, 1, 1, 1, -1, 1, 1, 1))
-##' plot_directed(graph_test4, state=c(1, 1, -1, 1, -1, 1, -1, 1))
+##' plot_directed(graph_structure, state = c(1, 1, 1, 1, -1, 1, 1, 1))
+##' plot_directed(graph_structure, state = c(1, 1, -1, 1, -1, 1, -1, 1))
+##' plot_directed(graph_structure, state = c(1, 1, -1, 1, 1, 1, 1, -1)
 ##' 
 ##' #plots states with graph attributes
-##' E(graph_test4)$state <- 1
-##' plot_directed(graph_test4)
-##' E(graph_test4)$state <- c(1, 1, -1, 1, -1, 1, -1, 1)
-##' plot_directed(graph_test4)
+##' E(graph_structure)$state <- 1
+##' plot_directed(graph_structure)
+##' E(graph_structure)$state <- c(1, 1, -1, 1, -1, 1, -1, 1)
+##' plot_directed(graph_structure)
 ##'
 ##' #plot layout customised
-##' plot_directed(graph_test4, state=c(1, 1, -1, 1, -1, 1, -1, 1), layout = layout.kamada.kawai)
+##' plot_directed(graph_structure, state=c(1, 1, -1, 1, -1, 1, -1, 1), layout = layout.kamada.kawai)
 ##' 
 ##' @return base R graphics
 ##' 
