+++ R/sigma.R
@@ -1,9 +1,9 @@
 ##' @name make_sigma
 ##' @rdname make_sigma
 ##'
-##' @title Generate Sigma Matrix
+##' @title Generate Sigma (Σ) Matrix
 ##'
-##' @description Compute the Sigma matrix from an \code{\link[igraph]{igraph}} structure 
+##' @description Compute the Sigma (Σ) matrix from an \code{\link[igraph]{igraph}} structure 
 ##' or pre-computed matrix. These are compatible with \code{\link[mvtnorm]{rmvnorm}} and
 ##' \code{\link[graphsim]{generate_expression}}.
 ##'
@@ -22,11 +22,106 @@
 ##' @import igraph
 ##' @examples 
 ##' 
+##' # construct a synthetic graph module
 ##' library("igraph")
 ##' graph_test_edges <- rbind(c("A", "B"), c("B", "C"), c("B", "D"))
 ##' graph_test <- graph.edgelist(graph_test_edges, directed = TRUE)
+##' # compute sigma (Σ) matrix for toy example
+##' sigma_matrix <- make_sigma_mat_graph(graph_test, cor = 0.8)
+##' sigma_matrix
+##' 
+##' # compute sigma (Σ) matrix  from adjacency matrix for toy example
 ##' adjacency_matrix <- make_adjmatrix_graph(graph_test)
 ##' sigma_matrix <- make_sigma_mat_adjmat(adjacency_matrix, cor = 0.8)
+##' sigma_matrix
+##' 
+##' # compute sigma (Σ) matrix from shared edges for toy example
+##' common_link_matrix <- make_commonlink_graph(graph_test)
+##' sigma_matrix <- make_sigma_mat_comm(common_link_matrix, cor = 0.8)
+##' sigma_matrix
+##' 
+##' # compute sigma (Σ) matrix from Laplacian for toy example
+##' laplacian_matrix <- make_laplacian_graph(graph_test)
+##' sigma_matrix <- make_sigma_mat_laplacian(laplacian_matrix, cor = 0.8)
+##' sigma_matrix
+##' 
+##' # compute sigma (Σ) matrix from distance matrix for toy example
+##' distance_matrix <- make_distance_graph(graph_test, absolute = FALSE)
+##' sigma_matrix <- make_sigma_mat_dist_adjmat(distance_matrix, cor = 0.8)
+##' sigma_matrix
+##' 
+##' # compute sigma (Σ) matrix from geometric distance directly from toy example graph
+##' sigma_matrix <- make_sigma_mat_dist_graph(graph_test, cor = 0.8)
+##' sigma_matrix
+##' 
+##' # compute sigma (Σ) matrix from absolute distance directly from toy example graph
+##' sigma_matrix <- make_sigma_mat_dist_graph(graph_test, cor = 0.8, absolute = TRUE)
+##' sigma_matrix
+##' 
+##' # construct a synthetic graph network
+##' graph_structure_edges <- rbind(c("A", "C"), c("B", "C"), c("C", "D"), c("D", "E"),
+##'                                c("D", "F"), c("F", "G"), c("F", "I"), c("H", "I"))
+##' graph_structure <- graph.edgelist(graph_structure_edges, directed = TRUE)
+##' 
+##' # compute sigma (Σ) matrix from geometric distance directly from synthetic graph network
+##' sigma_matrix_graph_structure <- make_sigma_mat_dist_graph(graph_structure, cor = 0.8, absolute = FALSE)
+##' sigma_matrix_graph_structure
+##' # visualise matrix
+##' library("gplots")
[…]

[…]
