### 2020-06-11 e6bd9fa
update documentation with examples for generating expression and migrate state handling to sigma #6
