+++ R/sigma.R
@@ -5,7 +5,12 @@
 ##'
 ##' @description Compute the Sigma (Σ) matrix from an \code{\link[igraph]{igraph}} structure 
 ##' or pre-computed matrix. These are compatible with \code{\link[mvtnorm]{rmvnorm}} and
-##' \code{\link[graphsim]{generate_expression}}.
+##'  \code{\link[graphsim]{generate_expression}}.
+##' By default data is generated with a mean of 0 and standard deviation of 1 for 
+##' each gene (with correlations between derived from the graph structure).
+##' Thus where the Sigma (Σ) matrix has diagonals of 1 (for the variance of each gene)
+##' then the symmetric non-diagonal terms (for covariance) determine the correlations
+##' between each gene in the output from \code{\link[graphsim]{generate_expression}}.
 ##'
 ##' @param mat precomputed adjacency, laplacian, commonlink, or scaled distance matrix.
 ##' @param graph An \code{\link[igraph]{igraph}} object. May be directed or weighted.
@@ -19,7 +24,7 @@
 ##' relative difference from the diameter for a geometric decay in distance.
 ##' @keywords graph network igraph mvtnorm
 ##' @importFrom igraph as_adjacency_matrix
-##' @import igraph
+##' @importFrom igraph graph_from_adjacency_matrix set_edge_attr E
 ##' @examples 
 ##' 
 ##' # construct a synthetic graph module
@@ -93,35 +98,35 @@
 ##' TGFBeta_Smad_graph <- identity(TGFBeta_Smad_graph)
 ##' 
 ##' # compute sigma (Σ) matrix from geometric distance directly from TGF-β pathway
-##' TFGB_Smad_state <- E(TGFBeta_Smad_graph)$state
-##' table(TFGB_Smad_state)
+##' TFGBeta_Smad_state <- E(TGFBeta_Smad_graph)$state
+##' table(TFGBeta_Smad_state)
 ##' # states are edge attributes
-##'  sigma_matrix_TFGB_Smad_inhib <- make_sigma_mat_dist_graph(TGFBeta_Smad_graph, cor = 0.8, absolute = FALSE)
+##'  sigma_matrix_TFGBeta_Smad_inhib <- make_sigma_mat_dist_graph(TGFBeta_Smad_graph, cor = 0.8, absolute = FALSE)
 ##' # visualise matrix
 ##' library("gplots")
-##' heatmap.2(sigma_matrix_TFGB_Smad_inhib, scale = "none", trace = "none", col = colorpanel(50, "blue", "white", "red"))
+##' heatmap.2(sigma_matrix_TFGBeta_Smad_inhib, scale = "none", trace = "none", col = colorpanel(50, "blue", "white", "red"))
 ##' 
 ##' # compute sigma (Σ) matrix from geometric distance directly from TGF-β pathway
 ##' TGFBeta_Smad_graph <- remove.edge.attribute(TGFBeta_Smad_graph, "state")
 ##' # compute with states removed (all negative)
-##' sigma_matrix_TFGB_Smad <- make_sigma_mat_dist_graph(TGFBeta_Smad_graph, state = -1, cor = 0.8, absolute = FALSE)
+##' sigma_matrix_TFGBeta_Smad <- make_sigma_mat_dist_graph(TGFBeta_Smad_graph, state = -1, cor = 0.8, absolute = FALSE)
 ##' # visualise matrix
 ##' library("gplots")
-##' heatmap.2(sigma_matrix_TFGB_Smad, scale = "none", trace = "none", col = colorpanel(50, "white", "red"))
[…]
