+++ R/generate.R
@@ -3,16 +3,38 @@
 ##'
 ##' @title Generate Simulated Expression
 ##'
-##' @description Compute simulated continuous expression data from a graph network structure. Requires an \code{\link[igraph]{igraph}} pathway structure and a matrix of states (1 for activating and -1 for inhibiting) for link signed correlations, from a vector of edge states to a signed adjacency matrix for use in \code{\link[graphsim]{generate_expression}}. Uses graph structure to pass a sigma covariance matrix from \code{\link[graphsim]{make_sigma_mat_dist_graph}} or \code{\link[graphsim]{make_sigma_mat_graph}} on to \code{\link[mvtnorm]{rmvnorm}}.
+##' @description Compute simulated continuous expression data from a graph 
+##' network structure. Requires an \code{\link[igraph]{igraph}} pathway 
+##' structure and a matrix of states (1 for activating and -1 for 
+##' inhibiting) for link signed correlations, from a vector of edge states 
+##' to a signed adjacency matrix for use in 
+##' \code{\link[graphsim]{generate_expression}}. 
+##' Uses graph structure to pass a sigma covariance matrix from 
+##' \code{\link[graphsim]{make_sigma_mat_dist_graph}} or 
+##' \code{\link[graphsim]{make_sigma_mat_graph}} on to 
+##' \code{\link[mvtnorm]{rmvnorm}}. By default data is generated with a mean of
+##'  0 and standard deviation of 1 for each gene (with correlations between 
+##'  derived from the graph structure).
 ##'
 ##' @param n number of observations (simulated samples).
-##' @param mat precomputed adjacency, laplacian, commonlink, or scaled distance matrix.
-##' @param graph An \code{\link[igraph]{igraph}} object. May must be directed if states are used.
-##' @param state numeric vector. Vector of length E(graph). Sign used to calculate state matrix, may be an integer state or inferred directly from expected correlations for each edge. May be applied a scalar across all edges or as a vector for each edge respectively. May also be entered as text for "activating" or "inhibiting" or as integers for activating (0,1) or inhibiting (-1,2). Compatible with inputs for \code{\link[graphsim]{plot_directed}}. Also takes a pre-computed state matrix from \code{\link[graphsim]{make_state_matrix}} if applied to the same graph multiple times.
-##' @param cor numeric. Simulated maximum correlation/covariance of two adjacent nodes. Default to 0.8.
-##' @param mean mean value of each simulated gene. Defaults to 0. May be entered as a scalar applying to all genes or a vector with a separate value for each.
-##' @param dist logical. Whether a graph distance (\code{\link[graphsim]{make_sigma_mat_dist_graph}}) or derived matrix (\code{\link[graphsim]{make_sigma_mat_graph}}) is used to compute the sigma matrix.
-##' @param comm,absolute,laplacian logical. Parameters for Sigma matrix generation. Passed on to \code{\link[graphsim]{make_sigma_mat_dist_graph}} or \code{\link[graphsim]{make_sigma_mat_graph}}.
+##' @param mat precomputed adjacency, laplacian, commonlink, or scaled 
+##' distance matrix.
+##' @param graph An \code{\link[igraph]{igraph}} object. May must be 
+##' directed if states are used.
[…]

[…]
