### 2020-06-11 abcf524
add standard deviation parameter
