+++ R/sigma.R
@@ -16,6 +16,7 @@
 ##' @param graph An \code{\link[igraph]{igraph}} object. May be directed or weighted.
 ##' @param state numeric vector. Vector of length E(graph). Sign used to calculate state matrix, may be an integer state or inferred directly from expected correlations for each edge. May be applied a scalar across all edges or as a vector for each edge respectively. May also be entered as text for "activating" or "inhibiting" or as integers for activating (0,1) or inhibiting (-1,2). Compatible with inputs for \code{\link[graphsim]{plot_directed}}. Also takes a pre-computed state matrix from \code{\link[graphsim]{make_state_matrix}} if applied to the same graph multiple times.
 ##' @param cor numeric. Simulated maximum correlation/covariance of two adjacent nodes. Default to 0.8.
+##' @param sd	 standard deviations of each gene. Defaults to 1. May be entered as a scalar applying to all genes or a vector with a separate value for each. 
 ##' @param directed logical. Whether directed information is passed to the distance matrix.
 ##' @param comm logical whether a common link matrix is used to compute sigma. Defaults to FALSE (adjacency matrix).
 ##' @param laplacian logical whether a Laplacian matrix is used to compute sigma. Defaults to FALSE (adjacency matrix).
@@ -24,7 +25,7 @@
 ##' relative difference from the diameter for a geometric decay in distance.
 ##' @keywords graph network igraph mvtnorm
 ##' @importFrom igraph as_adjacency_matrix
-##' @importFrom igraph graph_from_adjacency_matrix set_edge_attr E
+##' @importFrom igraph graph_from_adjacency_matrix set_edge_attr V E
 ##' @examples 
 ##' 
 ##' # construct a synthetic graph module
@@ -32,35 +33,35 @@
 ##' graph_test_edges <- rbind(c("A", "B"), c("B", "C"), c("B", "D"))
 ##' graph_test <- graph.edgelist(graph_test_edges, directed = TRUE)
 ##' # compute sigma (Σ) matrix for toy example
-##' sigma_matrix <- make_sigma_mat_graph(graph_test, cor = 0.8)
+##' sigma_matrix <- make_sigma_mat_graph(graph_test, cor = 0.8, sd = 1)
 ##' sigma_matrix
 ##' 
 ##' # compute sigma (Σ) matrix  from adjacency matrix for toy example
 ##' adjacency_matrix <- make_adjmatrix_graph(graph_test)
-##' sigma_matrix <- make_sigma_mat_adjmat(adjacency_matrix, cor = 0.8)
+##' sigma_matrix <- make_sigma_mat_adjmat(adjacency_matrix, cor = 0.8, sd = 1)
 ##' sigma_matrix
 ##' 
 ##' # compute sigma (Σ) matrix from shared edges for toy example
 ##' common_link_matrix <- make_commonlink_graph(graph_test)
-##' sigma_matrix <- make_sigma_mat_comm(common_link_matrix, cor = 0.8)
+##' sigma_matrix <- make_sigma_mat_comm(common_link_matrix, cor = 0.8, sd = 1)
 ##' sigma_matrix
 ##' 
 ##' # compute sigma (Σ) matrix from Laplacian for toy example
[…]
