+++ NAMESPACE
@@ -25,6 +25,7 @@ importFrom(Matrix,Matrix)
 importFrom(Matrix,nearPD)
 importFrom(gplots,heatmap.2)
 importFrom(igraph,E)
+importFrom(igraph,V)
 importFrom(igraph,as.undirected)
 importFrom(igraph,as_adjacency_matrix)
 importFrom(igraph,get.edge.attribute)

[…]
