### 2020-06-15 05362e8
update travis to test old versions (R 3.6.3 and R 4.0.1)
