+++ .travis.yml
@@ -4,6 +4,21 @@ language: R
 sudo: false
 cache: packages
 
+matrix:
+  include:
+    - os: linux
+      r: oldrel
+    - os: linux
+      r: release
+    - os: linux
+      r: devel
+    - os: osx
+      r: oldrel
+    - os: osx
+      r: release
+    - os: osx
+      r: devel
+      
 environment:
   # enable remotes
   USE_RTOOLS: true
