### 2020-06-17 #2161 @majensen
Hi @TomKellyGenetics - how are things progressing?
