+++ paper/paper.Rmd
@@ -105,7 +105,6 @@ of regulatory networks has thus far primarily relied on experimental
 validation or resampling-based approaches to estimate the likelihood
 of specific network modules being predicted [@Markowetz2007; @Hawe2019].
 
-
 Simulated datasets with pathway structure are of particular interest for
 benchmarking regulatory network inference techniques and methods
 being developed for genomics data with complex biological 
@@ -122,7 +121,7 @@ have been developed to model single-cell data.
 SERGIO [@Dibaeinia2019] and dyngen [@Cannoodt2020] build on
 this to include gene regulatory networks and multimodality
 respectively. These methods have been designed based on known
-or modelled reaction states and added stochasticity to determinisitic
+or modelled reaction states and added stochasticity to deterministic
 relationships. However, it is computationally-intensive to model
 these reactions at scale over iterations in benchmarking.
 In some cases, it is only necessary to model the statistical
