+++ .travis.yml
@@ -16,8 +16,6 @@ matrix:
       r: oldrel
     - os: osx
       r: release
-    - os: osx
-      r: devel
       
 environment:
   # enable remotes
