+++ man/make_state_matrix.Rd
@@ -77,16 +77,25 @@ table(TFGBeta_Smad_state)
 state_matrix_TFGBeta_Smad <- make_state_matrix(TGFBeta_Smad_graph)
 # visualise matrix
 library("gplots")
-heatmap.2(state_matrix_TFGBeta_Smad , scale = "none", trace = "none", dendrogram = "none", Rowv = FALSE, Colv = FALSE, col = colorpanel(50, "blue", "white", "red"))
+heatmap.2(state_matrix_TFGBeta_Smad , scale = "none", trace = "none",
+          dendrogram = "none", Rowv = FALSE, Colv = FALSE,
+          col = colorpanel(50, "blue", "white", "red"))
 
 # compare the states to the sign of expected correlations in the sigma matrix
-sigma_matrix_TFGBeta_Smad_inhib <- make_sigma_mat_dist_graph(TGFBeta_Smad_graph, cor = 0.8, absolute = FALSE)
+sigma_matrix_TFGBeta_Smad_inhib <- make_sigma_mat_dist_graph(TGFBeta_Smad_graph,
+                                                             cor = 0.8,
+                                                             absolute = FALSE)
 # visualise matrix
-heatmap.2(sigma_matrix_TFGBeta_Smad_inhib, scale = "none", trace = "none", dendrogram = "none", Rowv = FALSE, Colv = FALSE, col = colorpanel(50, "blue", "white", "red"))
+heatmap.2(sigma_matrix_TFGBeta_Smad_inhib,
+          scale = "none", trace = "none",
+          dendrogram = "none", Rowv = FALSE, Colv = FALSE,
+          col = colorpanel(50, "blue", "white", "red"))
 
 # compare the states to the sign of final correlations in the simulated matrix
 TFGBeta_Smad_data <- generate_expression(100, TGFBeta_Smad_graph, cor = 0.8)
-heatmap.2(cor(t(TFGBeta_Smad_data)), scale = "none", trace = "none", dendrogram = "none", Rowv = FALSE, Colv = FALSE, col = colorpanel(50, "blue", "white", "red"))
+heatmap.2(cor(t(TFGBeta_Smad_data)), scale = "none", trace = "none",
+          dendrogram = "none", Rowv = FALSE, Colv = FALSE,
+          col = colorpanel(50, "blue", "white", "red"))
 
 
 }

[…]
