### 2020-06-18 bbcae26
correct handling of sd as vector
