### 2020-06-18 35ec5e3
correct documentation for plotting function
