### 2020-06-18 #2161 @TomKellyGenetics
Hi @majensen, thanks for following up. I've been busy with other projects but I've got another round of revisions almost ready push to GitHub. This should address the remaining [issues](https://github.com/TomKellyGenetics/graphsim/issues/) on the [graphsim](https://github.com/TomKellyGenetics/graphsim) package repository.

I've reached out to my coauthor (former supervisor) for a second opinion before releasing the updated manuscript. At the moment, I'm making sure the package with updated documentation passes CRAN checks and formatting the manuscript so that it is shorter.

It should be ready for the reviewers to check soon (I think next week).
