### 2020-06-18 5caef59
reduce vignette figure resolution to meet CRAN checks
