+++ vignettes/simulate_expression.Rmd
@@ -248,7 +248,7 @@ set.seed(9000)
 ```
 
 
-```{r simulation_activating_hide, fig.show='hold', out.width = '50%', out.height  = '50%', fig.align='center', dpi=36, fig.retina=10, fig.margin = FALSE, fig.ncol = 2, warning=FALSE, message=FALSE, cache=TRUE}
+```{r simulation_activating_hide, fig.show='hold', out.width = '50%', out.height  = '50%', fig.align='center', dpi=36, fig.retina=2, fig.margin = FALSE, fig.ncol = 2, warning=FALSE, message=FALSE, cache=TRUE}
 # activating graph
 state <- rep(1, length(E(graph_structure)))
 plot_directed(graph_structure, state=state, layout = layout.kamada.kawai,
@@ -290,7 +290,7 @@ set.seed(9000)
 ```
 
 
-```{r simulation_inhibiting_hide, fig.show='hold', out.width = '50%', out.height  = '50%', fig.align='center', dpi=36, fig.retina=10, fig.margin = FALSE, fig.ncol = 2, warning=FALSE, message=FALSE, cache=TRUE}
+```{r simulation_inhibiting_hide, fig.show='hold', out.width = '50%', out.height  = '50%', fig.align='center', dpi=36, fig.retina=2, fig.margin = FALSE, fig.ncol = 2, warning=FALSE, message=FALSE, cache=TRUE}
 # activating graph
 state <- state <- c(1, 1, -1, 1, 1, 1, 1, -1)
 plot_directed(graph_structure, state=state, layout = layout.kamada.kawai,
@@ -334,7 +334,7 @@ set.seed(9000)
 ```
 
 
-```{r simulation_graph_diverging_hide, fig.show='hold', out.width = '50%', out.height  = '50%', fig.align='center', dpi=36, fig.retina=10, fig.margin = FALSE, fig.ncol = 2, warning=FALSE, message=FALSE, cache=TRUE}
+```{r simulation_graph_diverging_hide, fig.show='hold', out.width = '50%', out.height  = '50%', fig.align='center', dpi=36, fig.retina=2, fig.margin = FALSE, fig.ncol = 2, warning=FALSE, message=FALSE, cache=TRUE}
 graph_diverging_edges <- rbind(c("A", "B"), c("B", "C"), c("B", "D"))
 graph_diverging <- graph.edgelist(graph_diverging_edges, directed = T)
 
@@ -379,7 +379,7 @@ set.seed(9000)
 ```
 
 
-```{r simulation_graph_diverging_inhibiting_hide, fig.show='hold', out.width = '50%', out.height  = '50%', fig.align='center', dpi=36, fig.retina=10, fig.margin = FALSE, fig.ncol = 2, warning=FALSE, message=FALSE, cache=TRUE}
+```{r simulation_graph_diverging_inhibiting_hide, fig.show='hold', out.width = '50%', out.height  = '50%', fig.align='center', dpi=36, fig.retina=2, fig.margin = FALSE, fig.ncol = 2, warning=FALSE, message=FALSE, cache=TRUE}
 # activating graph
 state <- state <- c(1, 1, -1)
 plot_directed(graph_diverging, state=state, layout = layout.kamada.kawai,
@@ -419,7 +419,7 @@ set.seed(9000)
 ```
 
 
-```{r simulation_graph_converging_hide, fig.show='hold', out.width = '50%', out.height  = '50%', fig.align='center', dpi=36, fig.retina=10, fig.margin = FALSE, fig.ncol = 2, warning=FALSE, message=FALSE, cache=TRUE}
[…]
