### 2020-06-18 3cce084
test plotting functions
