### 2020-06-19 1463ede
update documentation to handle Greek letters as ASCII with pdflatex
