+++ R/distance.R
@@ -62,7 +62,7 @@
 ##'           col = colorpanel(50, "white", "red"))
 ##'           
 ##' # import graph from package for reactome pathway
-##' # TGF-β receptor signaling activates SMADs (R-HSA-2173789)
+##' # TGF-\eqn{\Beta} receptor signaling activates SMADs (R-HSA-2173789)
 ##' TGFBeta_Smad_graph <- identity(TGFBeta_Smad_graph)
 ##' # compute nodes with relationships between nodes (geometrically decreasing by default)
 ##' TGFBeta_Smad_adjacency_matrix <- make_adjmatrix_graph(TGFBeta_Smad_graph)
