+++ R/sigma.R
@@ -1,14 +1,14 @@
 ##' @name make_sigma
 ##' @rdname make_sigma
 ##'
-##' @title Generate Sigma (Σ) Matrix
+##' @title Generate Sigma (\eqn{\Sigma}) Matrix
 ##'
-##' @description Compute the Sigma (Σ) matrix from an \code{\link[igraph]{igraph}} structure 
+##' @description Compute the Sigma (\eqn{\Sigma}) matrix from an \code{\link[igraph]{igraph}} structure 
 ##' or pre-computed matrix. These are compatible with \code{\link[mvtnorm]{rmvnorm}} and
 ##'  \code{\link[graphsim]{generate_expression}}.
 ##' By default data is generated with a mean of 0 and standard deviation of 1 for 
 ##' each gene (with correlations between derived from the graph structure).
-##' Thus where the Sigma (Σ) matrix has diagonals of 1 (for the variance of each gene)
+##' Thus where the Sigma (\eqn{\Sigma}) matrix has diagonals of 1 (for the variance of each gene)
 ##' then the symmetric non-diagonal terms (for covariance) determine the correlations
 ##' between each gene in the output from \code{\link[graphsim]{generate_expression}}.
 ##'
@@ -42,39 +42,39 @@
 ##' library("igraph")
 ##' graph_test_edges <- rbind(c("A", "B"), c("B", "C"), c("B", "D"))
 ##' graph_test <- graph.edgelist(graph_test_edges, directed = TRUE)
-##' # compute sigma (Σ) matrix for toy example
+##' # compute sigma (\eqn{\Sigma}) matrix for toy example
 ##' sigma_matrix <- make_sigma_mat_graph(graph_test, cor = 0.8)
 ##' sigma_matrix
 ##' 
-##' # compute sigma (Σ) matrix  from adjacency matrix for toy example
+##' # compute sigma (\eqn{\Sigma}) matrix  from adjacency matrix for toy example
 ##' adjacency_matrix <- make_adjmatrix_graph(graph_test)
 ##' sigma_matrix <- make_sigma_mat_adjmat(adjacency_matrix, cor = 0.8)
 ##' sigma_matrix
 ##' 
-##' # compute sigma (Σ) matrix from shared edges for toy example
+##' # compute sigma (\eqn{\Sigma}) matrix from shared edges for toy example
 ##' common_link_matrix <- make_commonlink_graph(graph_test)
 ##' sigma_matrix <- make_sigma_mat_comm(common_link_matrix, cor = 0.8)
 ##' sigma_matrix
 ##' 
-##' # compute sigma (Σ) matrix from Laplacian for toy example
+##' # compute sigma (\eqn{\Sigma}) matrix from Laplacian for toy example
 ##' laplacian_matrix <- make_laplacian_graph(graph_test)
 ##' sigma_matrix <- make_sigma_mat_laplacian(laplacian_matrix, cor = 0.8)
 ##' sigma_matrix
 ##' 
-##' # compute sigma (Σ) matrix from distance matrix for toy example
+##' # compute sigma (\eqn{\Sigma}) matrix from distance matrix for toy example
 ##' distance_matrix <- make_distance_graph(graph_test, absolute = FALSE)
 ##' sigma_matrix <- make_sigma_mat_dist_adjmat(distance_matrix, cor = 0.8)
 ##' sigma_matrix
 ##' 
-##' # compute sigma (Σ) matrix from geometric distance directly from toy example graph
+##' # compute sigma (\eqn{\Sigma}) matrix from geometric distance directly from toy example graph
 ##' sigma_matrix <- make_sigma_mat_dist_graph(graph_test, cor = 0.8)
 ##' sigma_matrix
 ##' 
[…]
