+++ man/make_laplacian.Rd
@@ -48,10 +48,10 @@ graph_structure_laplacian_matrix <- make_laplacian_graph(graph_structure)
 graph_structure_laplacian_matrix
  
 # import graph from package for reactome pathway
-# TGF-β receptor signaling activates SMADs (R-HSA-2173789)
+# TGF-\eqn{\Beta} receptor signaling activates SMADs (R-HSA-2173789)
 TGFBeta_Smad_graph <- identity(TGFBeta_Smad_graph)
 
-# compute Laplacian matrix for TGF-β receptor signaling activates SMADs
+# compute Laplacian matrix for TGF-\eqn{\Beta} receptor signaling activates SMADs
 TGFBeta_Smad_laplacian_matrix <- make_laplacian_graph(TGFBeta_Smad_graph)
 dim(TGFBeta_Smad_laplacian_matrix)
 TGFBeta_Smad_laplacian_matrix[1:12, 1:12]
