### 2020-06-19 dbce383
test Mac OS on Travis
