+++ .travis.yml
@@ -27,20 +27,6 @@ before_install:
   - sudo apt-get install -y libmagick++-dev
   - sudo apt-get install tk
   
-addons:
-  apt:
-    sources:
-      - sourceline: 'ppa:opencpu/imagemagick'
-    packages:
-      - libmagick++-dev
-      - tk
-  homebrew:
-    packages:
-      - php
-      - imagemagick
-      - pkg-config
-    update: true  
-
 r_packages:
   - igraph
   - utils
