+++ .travis.yml
@@ -20,13 +20,21 @@ matrix:
 environment:
   # enable remotes
   USE_RTOOLS: true
-
-before_install:
-  - sudo add-apt-repository -y ppa:opencpu/imagemagick
-  - sudo apt-get update
-  - sudo apt-get install -y libmagick++-dev
-  - sudo apt-get install tk
   
+addons:
+  apt:
+    sources:
+      - sourceline: 'ppa:opencpu/imagemagick'
+    packages:
+      - libmagick++-dev
+      - tk
+  homebrew:
+    packages:
+      - php
+      - imagemagick
+      - pkg-config
+    update: true  
+
 r_packages:
   - igraph
   - utils
