### 2020-06-24 346bcf3
update paper introduction text (biorXiv version)
