### 2020-06-24 d0569b8
update paper formatting (biorXiv version)
