### 2020-06-24 e8f4b67
update figure formatting
