### 2020-06-24 20af833
update figure placement settings
