+++ paper/paper.md
@@ -23,7 +23,7 @@ authors:
     orcid: 0000-0003-1174-6054
     affiliation: "1"
 affiliations:
- - name: "Department of Biochemistry, University of Otago, PO Box 56, Dunedin 9054, New Zealand \n"
+ - name: "Department of Biochemistry, University of Otago, PO Box 56, Dunedin 9054, New Zealand"
    index: 1
  - name: "RIKEN Center for Integrative Medical Sciences, Suehiro-cho-1-7-22, Tsurumi Ward, Yokohama, Kanagawa 230-0045, Japan"
    index: 2
@@ -31,7 +31,7 @@ date: "24 June 2020"
 bibliography: paper.bib
 header-includes:
   - \usepackage{caption}
- # - \usepackage{subfig}
+  - \include{header.tex}
 ---
 
 
@@ -262,6 +262,10 @@ graph_edges <- rbind(c("A", "C"), c("B", "C"), c("C", "D"), c("D", "E"),
    c("D", "F"), c("F", "G"), c("F", "I"), c("H", "I"))
 graph <- graph.edgelist(graph_edges, directed = TRUE)
 
+#plot graph structure (Figure 1a)
+plot_directed(graph, state ="activating", layout = layout.kamada.kawai,
+  cex.node = 2, cex.arrow = 4, arrow_clip = 0.2)
+
 #generate parameters for inhibitions for each edge in E(graph)
 state <- c(1, 1, -1, 1, 1, 1, 1, -1)
 #plot graph structure with inhibitions (Figure 1b)
@@ -317,11 +321,6 @@ The simulated dataset can be generated using the following code:
 
 
 ```r
-# activating graph
-state <- rep(1, length(E(graph)))
-plot_directed(graph, state=state, layout = layout.kamada.kawai,
-  cex.node=2, cex.arrow=4, arrow_clip = 0.2)
-
 #plot relationship matrix
 heatmap.2(make_distance_graph(graph, absolute = FALSE),
   scale = "none", trace = "none", col = colorpanel(50, "white", "red"),
@@ -346,7 +345,7 @@ heatmap.2(expr, scale = "none", trace = "none", col = bluered(50),
 
 \begin{figure}[!htbp]
 
-{\centering \includegraphics[width=.375\linewidth,height=.375\linewidth]{Plotsimulation_activating-1} \includegraphics[width=.375\linewidth,height=.375\linewidth]{Plotsimulation_activating-2} \includegraphics[width=.375\linewidth,height=.375\linewidth]{Plotsimulation_activating-3} \includegraphics[width=.750\linewidth,height=.375\linewidth]{Plotsimulation_activating-4} \includegraphics[width=.750\linewidth,height=.375\linewidth]{Plotsimulation_activating-5} 
+{\centering \includegraphics[width=.375\linewidth,height=.375\linewidth]{Plotsimulation_activating-1} \includegraphics[width=.375\linewidth,height=.375\linewidth]{Plotsimulation_activating-2} \includegraphics[width=.375\linewidth,height=.375\linewidth]{Plotsimulation_activating-3} \includegraphics[width=.375\linewidth,height=.375\linewidth]{Plotsimulation_activating-4} \includegraphics[width=.750\linewidth,height=.375\linewidth]{Plotsimulation_activating-5} 
 
 }
 
@@ -414,7 +413,7 @@ heatmap.2(expr, scale = "none", trace = "none", col = bluered(50),
 
 \begin{figure}[!htbp]
 
[…]
