### 2020-06-24 b5c9757
update figure placement settings
