### 2020-06-24 d6918c6
update figure placement settings
