+++ paper/paper.md
@@ -32,6 +32,7 @@ bibliography: paper.bib
 header-includes:
   - \usepackage{caption}
   - \include{header.tex}
+  - \let\origfigure\figure \let\endorigfigure\endfigure \renewenvironment{figure}[1][2] { \expandafter\origfigure\expandafter[!htbp] } { \endorigfigure }
 ---
 
 
