+++ paper/paper.Rmd
@@ -32,13 +32,7 @@ bibliography: paper.bib
 header-includes:
   - \usepackage{caption}
   - \include{header.tex}
-  - \let\origfigure\figure |
-  \let\endorigfigure\endfigure |
-\renewenvironment{figure}[1][2] { |
-    \expandafter\origfigure\expandafter[!htbp]
-} { |
-    \endorigfigure |
-}
+  - \let\origfigure\figure \let\endorigfigure\endfigure \renewenvironment{figure}[1][2] { \expandafter\origfigure\expandafter[!htbp] } { \endorigfigure }
 ---
 
 
