### 2020-06-24 9b172db
update figure placement settings
