### 2020-06-24 2b6352b
update figure placement settings
