+++ paper/paper.md
@@ -31,9 +31,6 @@ date: "24 June 2020"
 bibliography: paper.bib
 header-includes:
   - \usepackage{caption}
-  - \let\origfigure\figure
-  - \let\endorigfigure\endfigure
-  - \renewenvironment{figure}[1][2] { \expandafter\origfigure\expandafter[!htbp] } { \endorigfigure }
 ---
 
 
@@ -219,6 +216,16 @@ definite matrix is computed using the `nearPD` function of the \texttt{Matrix} R
 variance-covariance matrices. The \texttt{graphsim} package gives a warning when this
 occurs.
 
+
+\begin{figure}
+
+{\centering \includegraphics[width=.375\linewidth,height=.375\linewidth]{Plotsimple_graph-1} \includegraphics[width=.375\linewidth,height=.375\linewidth]{Plotsimple_graph-2} 
+
+}
+
+\caption{\textbf{Simulated graph structures}. A constructed graph structure used as an example to demonstrate the simulation procedure in Figures 2 and 3. Activating links are denoted by black arrows and inhibiting links by red edges. Inhibiting edges have been highlighted in red.}\label{fig:simple_graph}
+\end{figure}
+
 Illustrations {#sec:illustrations}
 ===============================================================================
 
@@ -275,14 +282,6 @@ plot_directed(graph, state=state, layout = layout.kamada.kawai,
   cex.node = 2, cex.arrow = 4, arrow_clip = 0.2)
 ```
 
-\begin{figure}
-
-{\centering \includegraphics[width=.375\linewidth,height=.375\linewidth]{Plotsimple_graph-1} \includegraphics[width=.375\linewidth,height=.375\linewidth]{Plotsimple_graph-2} 
-
-}
-
-\caption{\textbf{Simulated graph structures}. A constructed graph structure used as an example to demonstrate the simulation procedure in Figures 2 and 3. Activating links are denoted by black arrows and inhibiting links by red edges. Inhibiting edges have been highlighted in red.}\label{fig:simple_graph}
-\end{figure}
 
 Generating a Simulated Expression Dataset {#sec:graphsim_demo}
 -----------------------------------------
@@ -300,7 +299,19 @@ normally-distributed) log-expression profile for each node
 (Figure [2e](#fig:simulation_activating:fourth){reference-type="ref"
 reference="fig:simulation_activating:fourth"}) with a corresponding
 correlation structure (Figure [2d](#fig:simulation_activating:fifth){reference-type="ref"
-reference="fig:simulation_activating:fifth"}). The simulated correlation
+reference="fig:simulation_activating:fifth"}).
+
+
+\begin{figure}
+
+{\centering \includegraphics[width=.375\linewidth,height=.375\linewidth]{Plotsimulation_activating-1} \includegraphics[width=.375\linewidth,height=.375\linewidth]{Plotsimulation_activating-2} \includegraphics[width=.375\linewidth,height=.375\linewidth]{Plotsimulation_activating-3} \includegraphics[width=.375\linewidth,height=.375\linewidth]{Plotsimulation_activating-4} \includegraphics[width=.375\linewidth,height=.375\linewidth]{Plotsimulation_activating-5} 
+
+}
+
[…]
