### 2020-06-24 17c1218
update figure placement settings
