+++ paper/paper.md
@@ -216,16 +216,6 @@ definite matrix is computed using the `nearPD` function of the \texttt{Matrix} R
 variance-covariance matrices. The \texttt{graphsim} package gives a warning when this
 occurs.
 
-
-\begin{figure}
-
-{\centering \includegraphics[width=.375\linewidth,height=.375\linewidth]{Plotsimple_graph-1} \includegraphics[width=.375\linewidth,height=.375\linewidth]{Plotsimple_graph-2} 
-
-}
-
-\caption{\textbf{Simulated graph structures}. A constructed graph structure used as an example to demonstrate the simulation procedure in Figures 2 and 3. Activating links are denoted by black arrows and inhibiting links by red edges. Inhibiting edges have been highlighted in red.}\label{fig:simple_graph}
-\end{figure}
-
 Illustrations {#sec:illustrations}
 ===============================================================================
 
@@ -247,6 +237,16 @@ activation or inhibition. These inhibition links (which occur frequently
 in biological pathways) are demonstrated in
 Figure [1b](#fig:simple_graph:second){reference-type="ref"
 reference="fig:simple_graph:second"}.
+
+\begin{figure}
+
+{\centering \includegraphics[width=.375\linewidth,height=.375\linewidth]{Plotsimple_graph-1} \includegraphics[width=.375\linewidth,height=.375\linewidth]{Plotsimple_graph-2} 
+
+}
+
+\caption{\textbf{Simulated graph structures}. A constructed graph structure used as an example to demonstrate the simulation procedure in Figures 2 and 3. Activating links are denoted by black arrows and inhibiting links by red edges. Inhibiting edges have been highlighted in red.}\label{fig:simple_graph}
+\end{figure}
+
 
 A graph structure can be generated and plotted using the following
 commands in R:
@@ -301,16 +301,6 @@ reference="fig:simulation_activating:fourth"}) with a corresponding
 correlation structure (Figure [2d](#fig:simulation_activating:fifth){reference-type="ref"
 reference="fig:simulation_activating:fifth"}).
 
-
-\begin{figure}
-
-{\centering \includegraphics[width=.375\linewidth,height=.375\linewidth]{Plotsimulation_activating-1} \includegraphics[width=.375\linewidth,height=.375\linewidth]{Plotsimulation_activating-2} \includegraphics[width=.375\linewidth,height=.375\linewidth]{Plotsimulation_activating-3} \includegraphics[width=.375\linewidth,height=.375\linewidth]{Plotsimulation_activating-4} \includegraphics[width=.375\linewidth,height=.375\linewidth]{Plotsimulation_activating-5} 
-
-}
-
-\caption{\textbf{Simulating expression from a graph structure}. An example of a graph structure (a) that has been used to derive a relationship matrix (b), $\Sigma$  matrix (c) and correlation structure (d) from the relative distances between the nodes. Non-negative values are coloured white to red from $0$ to $1$. This $\Sigma$ matrix has been used to generate a simulated expression dataset of 100 samples (coloured blue to red from low to high) via sampling from the multivariate normal distribution. Here genes with closer relationships in the pathway structure show a higher correlation between simulated values.}\label{fig:simulation_activating}
-\end{figure}
-
 The simulated correlation
 structure closely resembled the expected correlation structure ($\Sigma$
 in Figure [2c](#fig:simulation_activating:third){reference-type="ref"
@@ -324,6 +314,17 @@ reference="fig:simulation_activating:fourth"}), it can then be used to
 test procedures designed for analysis of empirical gene expression data
[…]
