### 2020-06-24 b707ce6
update figure placement settings
