+++ paper/paper.md
@@ -390,23 +390,15 @@ plot_directed(graph, state=state, layout = layout.kamada.kawai,
 ```
 
 ```r
-#plot relationship matrix
-heatmap.2(make_distance_graph(graph, absolute = FALSE),
-  scale = "none", trace = "none", col = colorpanel(50, "white", "red"),
-  colsep = 1:length(V(graph)), rowsep = 1:length(V(graph)))
-
-#plot sigma matrix
+#plot sigma matrix with inhibitions
 heatmap.2(make_sigma_mat_dist_graph(graph, state, cor = 0.8, absolute = FALSE),
   scale = "none", trace = "none", col = colorpanel(50, "blue", "white", "red"),
   colsep = 1:length(V(graph)), rowsep = 1:length(V(graph)))
 
-#simulated data
+#simulated data with inhibitions
 expr <- generate_expression(100, graph, state, cor = 0.8, mean = 0,
   comm = FALSE, dist =TRUE, absolute = FALSE)
-#plot simulated correlations
-heatmap.2(cor(t(expr)), scale = "none", trace = "none",
-          col = colorpanel(50, "blue", "white", "red"),
-  colsep = 1:length(V(graph)), rowsep = 1:length(V(graph)))
+
 ```
 
 \begin{figure}
@@ -417,14 +409,6 @@ heatmap.2(cor(t(expr)), scale = "none", trace = "none",
 
 \caption{\textbf{Simulating expression from graph structure with inhibitions}. An example of a graph structure (a), that has been used to derive a relationship matrix (b), $\Sigma$ matrix (c), and correlation structure (d), from the relative distances between the nodes. These values are coloured blue to red from $-1$ to $1$. This has been used to generate a simulated expression dataset of 100 samples (coloured blue to red from low to high) via sampling from the multivariate normal distribution. Here the inhibitory relationships between genes are reflected in negatively correlated simulated  values.}\label{fig:simulation_inhibiting}
 \end{figure}
-
-```r  
-#plot simulated expression data
-heatmap.2(expr, scale = "none", trace = "none", col = bluered(50),
-  colsep = 1:length(V(graph)), rowsep = 1:length(V(graph)), labCol = "")
-```
-
-
 
 
 The simulation procedure is also demonstrated here
@@ -463,28 +447,8 @@ plot_directed(graph, state = state, layout = layout.kamada.kawai,
   border.node=alpha("black", 0.75), fill.node="lightblue",
   col.arrow = c(alpha("navyblue", 0.25), alpha("red", 0.25))[state], 
   cex.node = 1.5, cex.label = 0.8, cex.arrow = 2)
-
-#plot relationship matrix
-heatmap.2(make_distance_graph(graph, absolute = FALSE),
-  scale = "none", trace = "none", col = colorpanel(50, "white", "red"),
-  colsep = 1:length(V(graph)), rowsep = 1:length(V(graph)), labCol = "")
-
-#plot sigma matrix
-heatmap.2(make_sigma_mat_dist_graph(graph, state, cor = 0.8, absolute = FALSE),
-  scale = "none", trace = "none", col = colorpanel(50, "blue", "white", "red"),
-  colsep = 1:length(V(graph)), rowsep = 1:length(V(graph)), labCol = "")
-
-#simulated data
-expr <- generate_expression(100, graph, state, cor = 0.8,
-  mean = 0,comm = FALSE, dist =TRUE, absolute = FALSE)
-#plot simulated correlations
-heatmap.2(cor(t(expr)), scale = "none", trace = "none", 
-          col = colorpanel(50, "blue", "white", "red"),
[…]
