### 2020-06-24 01cc3ae
update figure placement settings
