+++ paper/paper.md
@@ -4,7 +4,7 @@ output:
   rmarkdown::pdf_document:
     fig_crop: no
     keep_md: TRUE
-    keep_tex: TRUE
+    #keep_tex: TRUE
     fig_caption: yes
 tags:
   - R
@@ -238,16 +238,6 @@ in biological pathways) are demonstrated in
 Figure [1b](#fig:simple_graph:second){reference-type="ref"
 reference="fig:simple_graph:second"}.
 
-\begin{figure}
-
-{\centering \includegraphics[width=.375\linewidth,height=.375\linewidth]{Plotsimple_graph-1} \includegraphics[width=.375\linewidth,height=.375\linewidth]{Plotsimple_graph-2} 
-
-}
-
-\caption{\textbf{Simulated graph structures}. A constructed graph structure used as an example to demonstrate the simulation procedure in Figures 2 and 3. Activating links are denoted by black arrows and inhibiting links by red edges. Inhibiting edges have been highlighted in red.}\label{fig:simple_graph}
-\end{figure}
-
-
 A graph structure can be generated and plotted using the following
 commands in R:
 
@@ -282,6 +272,14 @@ plot_directed(graph, state=state, layout = layout.kamada.kawai,
   cex.node = 2, cex.arrow = 4, arrow_clip = 0.2)
 ```
 
+\begin{figure}[!htbp]
+
+{\centering \includegraphics[width=.375\linewidth,height=.375\linewidth]{Plotsimple_graph-1} \includegraphics[width=.375\linewidth,height=.375\linewidth]{Plotsimple_graph-2} 
+
+}
+
+\caption{\textbf{Simulated graph structures}. A constructed graph structure used as an example to demonstrate the simulation procedure in Figures 2 and 3. Activating links are denoted by black arrows and inhibiting links by red edges. Inhibiting edges have been highlighted in red.}\label{fig:simple_graph}
+\end{figure}
 
 Generating a Simulated Expression Dataset {#sec:graphsim_demo}
 -----------------------------------------
@@ -299,9 +297,7 @@ normally-distributed) log-expression profile for each node
 (Figure [2e](#fig:simulation_activating:fourth){reference-type="ref"
 reference="fig:simulation_activating:fourth"}) with a corresponding
 correlation structure (Figure [2d](#fig:simulation_activating:fifth){reference-type="ref"
-reference="fig:simulation_activating:fifth"}).
-
-The simulated correlation
+reference="fig:simulation_activating:fifth"}). The simulated correlation
 structure closely resembled the expected correlation structure ($\Sigma$
 in Figure [2c](#fig:simulation_activating:third){reference-type="ref"
 reference="fig:simulation_activating:third"}) even for the relatively
@@ -315,18 +311,12 @@ test procedures designed for analysis of empirical gene expression data
 (such as those generated by microarrays or RNA-Seq) that have been
 normalised on a log-scale.
 
-\begin{figure}
-
-{\centering \includegraphics[width=.375\linewidth,height=.375\linewidth]{Plotsimulation_activating-1} \includegraphics[width=.375\linewidth,height=.375\linewidth]{Plotsimulation_activating-2} \includegraphics[width=.375\linewidth,height=.375\linewidth]{Plotsimulation_activating-3} \includegraphics[width=.375\linewidth,height=.375\linewidth]{Plotsimulation_activating-4} \includegraphics[width=.750\linewidth,height=.375\linewidth]{Plotsimulation_activating-5} 
-
-}
-
[…]
