### 2020-06-24 06bee52
update figure placement settings
