+++ paper/paper.md
@@ -216,6 +216,7 @@ definite matrix is computed using the `nearPD` function of the \texttt{Matrix} R
 variance-covariance matrices. The \texttt{graphsim} package gives a warning when this
 occurs.
 
+
 Illustrations {#sec:illustrations}
 ===============================================================================
 
@@ -237,6 +238,15 @@ activation or inhibition. These inhibition links (which occur frequently
 in biological pathways) are demonstrated in
 Figure [1b](#fig:simple_graph:second){reference-type="ref"
 reference="fig:simple_graph:second"}.
+
+\begin{figure}[!htbp]
+
+{\centering \includegraphics[width=.375\linewidth,height=.375\linewidth]{Plotsimple_graph-1} \includegraphics[width=.375\linewidth,height=.375\linewidth]{Plotsimple_graph-2} 
+
+}
+
+\caption{\textbf{Simulated graph structures}. A constructed graph structure used as an example to demonstrate the simulation procedure in Figures 2 and 3. Activating links are denoted by black arrows and inhibiting links by red edges. Inhibiting edges have been highlighted in red.}\label{fig:simple_graph}
+\end{figure}
 
 A graph structure can be generated and plotted using the following
 commands in R:
@@ -261,25 +271,12 @@ graph_edges <- rbind(c("A", "C"), c("B", "C"), c("C", "D"), c("D", "E"),
    c("D", "F"), c("F", "G"), c("F", "I"), c("H", "I"))
 graph <- graph.edgelist(graph_edges, directed = TRUE)
 
-#plot graph structure (Figure 1a)
-plot_directed(graph, state ="activating", layout = layout.kamada.kawai,
-  cex.node = 2, cex.arrow = 4, arrow_clip = 0.2)
-
 #generate parameters for inhibitions for each edge in E(graph)
 state <- c(1, 1, -1, 1, 1, 1, 1, -1)
-#plot graph structure with inhibitions (Figure 1b)
+#plot graph structure with inhibitions
 plot_directed(graph, state=state, layout = layout.kamada.kawai,
   cex.node = 2, cex.arrow = 4, arrow_clip = 0.2)
 ```
-
-\begin{figure}[!htbp]
-
-{\centering \includegraphics[width=.375\linewidth,height=.375\linewidth]{Plotsimple_graph-1} \includegraphics[width=.375\linewidth,height=.375\linewidth]{Plotsimple_graph-2} 
-
-}
-
-\caption{\textbf{Simulated graph structures}. A constructed graph structure used as an example to demonstrate the simulation procedure in Figures 2 and 3. Activating links are denoted by black arrows and inhibiting links by red edges. Inhibiting edges have been highlighted in red.}\label{fig:simple_graph}
-\end{figure}
 
 Generating a Simulated Expression Dataset {#sec:graphsim_demo}
 -----------------------------------------
@@ -311,12 +308,16 @@ test procedures designed for analysis of empirical gene expression data
 (such as those generated by microarrays or RNA-Seq) that have been
 normalised on a log-scale.
 
+\begin{figure}[!htbp]
+
+{\centering \includegraphics[width=.375\linewidth,height=.375\linewidth]{Plotsimulation_activating-1} \includegraphics[width=.375\linewidth,height=.375\linewidth]{Plotsimulation_activating-2} \includegraphics[width=.750\linewidth,height=.375\linewidth]{Plotsimulation_activating-3} \includegraphics[width=.750\linewidth,height=.375\linewidth]{Plotsimulation_activating-4} \includegraphics[width=.750\linewidth,height=.375\linewidth]{Plotsimulation_activating-5} 
+
+}
+
[…]
