### 2020-06-24 7bda8ff
update figure placement settings
