### 2020-06-24 7334521
update figure placement settings
