### 2020-06-28 cc78bde
update plot vignette formatting
