+++ vignettes/plots_directed.Rmd
@@ -1,8 +1,18 @@
 ---
 title: "Directed plots for igraph objects"
-author: "Tom Kelly"
+author: "S. Thomas Kelly^1,2^, Michael A. Black^2^ <br> ^1^ Department of Biochemistry, University of Otago, PO Box 56, Dunedin 9054, New Zealand <br> ^2^ RIKEN Center for Integrative Medical Sciences, Suehiro-cho-1-7-22, Tsurumi Ward, Yokohama"
 date: "`r  format(Sys.time(), '%A %d %B %Y')`"
-output: rmarkdown::html_vignette
+output:
+  prettydoc::html_pretty:
+       theme: cayman
+  #html_document:
+       #theme: united
+       number_sections: true
+       toc: true
+       toc_depth: 4
+       #toc_float: true
+       #code_folding: show
+       keep_html: true
 vignette: >
   %\VignetteIndexEntry{Directed plots for igraph objects}
   %\VignetteEngine{knitr::knitr}
@@ -20,13 +30,36 @@ par(bty="o")
 #captioner::captioner(prefix = "Fig.")
 ```
 
-## Motivations for directed graph plots
+
+<!-- <h1 class="title toc-ignore">Using the graphsim package:  Directed plots for igraph objects</h1> -->
+
+<!-- <h4 class="author">S. Thomas Kelly^1,2^, Michael A. Black^2^</h4> **** -->
+
+<!-- ^1^ Department of Biochemistry, University of Otago, PO Box 56, Dunedin 9054, New Zealand -->
+
+<!-- ^2^ RIKEN Center for Integrative Medical Sciences, Suehiro-cho-1-7-22, Tsurumi Ward, Yokohama -->
+
+<!-- <h4 class="date">`r  format(Sys.time(), '%A %d %B %Y')`</h4> -->
+
+<!-- ---------------------------------------------------------- -->
+
+**Abstract**
+
+Guide on how to plot directed graphs from `igraph` objects to display activating and repressive states.
+
+**Package**
+
+graphsim 0.1.2
+
+# Directed graph plots
+
+## Motivations
 
 Here we demonstrate the plotting functions that come built-in with `graphsim` and compare them them to the `plot.igraph` method provided by the `igraph` package. With all due respect to the efforts of the `igraph` developers, their plotting functions leave some things to be desired in terms of aesthetics. Here we provide additional functionality to plot directed graphs. This draws upon many functions provided by `igraph` including layout settings.
 
 In particular, graph and network represenations in biology often require displaying edge properties. Here we have the "state" parameter which can be used to differentiate these and plot activating and inhibiting or repressing relationships differently. We use different arrowheads as per convention in biochemistry and molecular genetics. 
 
-## Import packages
+## Getting started
 
 To display these plots the following packages must be imported.
 
@@ -35,18 +68,22 @@ library("igraph")
 library("graphsim")
 ```
 
-## Toy example
+# Toy example
 
 Here we demonstrate the plot functions on a small toy graph.
 
-### Set up simulated graph
+## Set up simulated graph
+
+First we generate a graph object using the `igraph` package.
 
 ```{r}
 graph_edges <- rbind(c("A", "C"), c("B", "C"), c("C", "D"), c("D", "E"), c("D", "F"), c("F", "G"), c("F", "I"), c("H", "I"))
 graph <- graph.edgelist(graph_edges, directed = T)
 ```
 
-### Plotting
+## Plotting
+
+Then we compare plotting functions to the built-in `plot.igraph` S3 method provided for igraph objects by the `igraph` package.
 
[…]
