+++ DESCRIPTION
@@ -21,11 +21,12 @@ Imports:
 Suggests:
     devtools,
     knitr (>= 1.5),
+    prettydoc,
     rmarkdown,
     testthat,
     scales,
     vdiffr
 LazyData: TRUE
-RoxygenNote: 7.1.0.9000
+RoxygenNote: 7.1.1
 Encoding: UTF-8
 VignetteBuilder: knitr
