### 2020-06-28 b4ae4b1
update simulation vignette formatting
