### 2020-06-30 00e33ec
update plotting vignettes for clarity
