+++ vignettes/plots_directed.Rmd
@@ -1,5 +1,5 @@
 ---
-title: "Directed plots for igraph objects"
+title: "graphsim: directed plots for igraph objects"
 author: "S. Thomas Kelly^1,2^, Michael A. Black^2^ <br> ^1^ Department of Biochemistry, University of Otago, PO Box 56, Dunedin 9054, New Zealand <br> ^2^ RIKEN Center for Integrative Medical Sciences, Suehiro-cho-1-7-22, Tsurumi Ward, Yokohama"
 date: "`r  format(Sys.time(), '%A %d %B %Y')`"
 output:
@@ -43,28 +43,29 @@ par(bty="o")
 
 <!-- ---------------------------------------------------------- -->
 
-**Abstract**
+**Summary**
 
 Guide on how to plot directed graphs from `igraph` objects to display activating and repressive states.
 
 **Package**
 
-graphsim 0.1.2
+graphsim 1.0.0
 
 # Directed graph plots
 
 ## Motivations
 
-Here we demonstrate the plotting functions that come built-in with `graphsim` and compare them them to the `plot.igraph` method provided by the `igraph` package. With all due respect to the efforts of the `igraph` developers, their plotting functions leave some things to be desired in terms of aesthetics. Here we provide additional functionality to plot directed graphs. This draws upon many functions provided by `igraph` including layout settings.
+Here we demonstrate the plotting functions that come built-in with `graphsim`, an alternative to the `plot.igraph` provided by the `igraph` package. Here we provide additional functionality for plotting directed graphs. This draws upon many functions provided by `igraph`, including layout settings <span class="citation">(Csardi and Nepusz 2006)</span>.
 
-In particular, graph and network represenations in biology often require displaying edge properties. Here we have the "state" parameter which can be used to differentiate these and plot activating and inhibiting or repressing relationships differently. We use different arrowheads as per convention in biochemistry and molecular genetics. 
+In particular, graph and network representations in biology often require displaying edge properties <span class="citation">(Barabási and Oltvai 2004)</span>. Here we have the "state" parameter which can be used to differentiate these, and allow us to represent activating and inhibiting or repressing relationships differently. We use different arrowheads as per convention in molecular biology. 
 
 ## Getting started
 
-To display these plots the following packages must be imported.
+To generate these plots, the following packages must be imported.
 
 ```{r, message=F}
 library("igraph")
+library("scales")
 library("graphsim")
 ```
 
@@ -77,96 +78,114 @@ Here we demonstrate the plot functions on a small toy graph.
 First we generate a graph object using the `igraph` package.
 
 ```{r}
-graph_edges <- rbind(c("A", "C"), c("B", "C"), c("C", "D"), c("D", "E"), c("D", "F"), c("F", "G"), c("F", "I"), c("H", "I"))
-graph <- graph.edgelist(graph_edges, directed = T)
+graph_edges <- rbind(c("A", "C"), c("B", "C"), c("C", "D"), c("D", "E"),
+                     c("D", "F"), c("F", "G"), c("F", "I"), c("H", "I"))
+graph <- graph.edgelist(graph_edges, directed = TRUE)
 ```
 
 ## Plotting
 
-Then we compare plotting functions to the built-in `plot.igraph` S3 method provided for igraph objects by the `igraph` package.
[…]
