### 2020-06-30 9dfc2a8
update simulation vignettes for clarity
