+++ vignettes/simulate_expression.Rmd
@@ -1,5 +1,5 @@
 ---
-title: "Simulating continuous data based on network graph structures"
+title: "graphsim: simulating continuous data based on network graph structures"
 author: "S. Thomas Kelly^1,2^, Michael A. Black^2^ <br> ^1^ Department of Biochemistry, University of Otago, PO Box 56, Dunedin 9054, New Zealand <br> ^2^ RIKEN Center for Integrative Medical Sciences, Suehiro-cho-1-7-22, Tsurumi Ward, Yokohama"
 date: "`r  format(Sys.time(), '%A %d %B %Y')`"
 output:
@@ -29,41 +29,41 @@ par(bty="o")
 #captioner::captioner(prefix = "Fig.")
 ```
 
-**Abstract**
+**Summary**
 
-Guide on how to simulate gene expression data from pathway strucutres defined by graphs from `igraph` objects.
+Guide on how to simulate gene expression data from pathway structures defined by graphs from `igraph` objects.
 
 **Package**
 
-graphsim 0.1.2
+graphsim 1.0.0
 
 # Overview of graphsim
 
-This package is designed to balance user-friendliness (by providing sensible defaults and built-in functions) and flexbility (many options are available to use as needed). It is likely that you will disagree with the parameters that I've used to simulate datasets. That's a valid opinion and there are many options at your disposal to alter parameters to use these functions for your purposes.
+This package is designed to balance user-friendliness (by providing sensible defaults and built-in functions) and flexbility (many options are available to use as needed). 
 
-If you have issues or feedback, sumbmission to the the GitHub repository are welcome. See the DESCRIPTION and README.md for more details on how to suggest changes to the package.
+If you have problems or feedback, sumbmitting an issue to the the GitHub repository is encouraged. See the DESCRIPTION and README.md for more details on how to suggest changes to the package.
 
 
 ## Motivations
 
-Pathway and graph structures have a wide array of applications. Here will consider the simulation of (log-normalised) gene expression data from genomics experiments in a biological pathway. If you have another use for this software you are welcome to apply it to your problem, please bear in mind that it was designed with this application in mind however. In principle, normally-distributed continuous data can be generated based on any defined relationships. This uses the graph structure to define a ∑ covariance matrix and generate simulated data by sampling from a multivariate normal distribution.
+Pathway and graph structures have a wide array of applications. Here we consider the simulation of (log-normalised) gene expression data from a biological pathway. If you have another use for this software you are welcome to apply it to your problem but please bear in mind that it was designed with this application in mind. In principle, normally-distributed continuous data can be generated based on any defined relationships. This package uses the graph structure to define a ∑ covariance matrix and generate simulated data by sampling from a multivariate normal distribution.
 
-Crucially, this allows the simulation of negative correlations based on inhibitory or repressive relationships, as commonly occur in biology. A custom plotting function `plot_directed` is provided to visualise these relationships with the "state" parameter. This plotting function has a dedicated vignette on [plotting](plots_directed.Rmd).
[…]
