### 2020-06-30 dc1e947
update manscript with feedback
