+++ paper/paper.Rmd
@@ -3,8 +3,8 @@ title: 'graphsim: An R package for simulating gene expression data from graph st
 output:
   rmarkdown::pdf_document:
     fig_crop: no
-    keep_md: TRUE
-    #keep_tex: TRUE
+    #keep_md: TRUE
+    keep_tex: TRUE
     fig_caption: yes
 tags:
   - R
@@ -52,13 +52,14 @@ identifying alterations in activity at the level of individual genes,
 understanding changes in the gene networks that regulate fundamental
 biological mechanisms is also an important objective of molecular 
 analysis. As a result, databases that describe biological pathways 
-are increasingly relied on to assist with the interpretation of results
+are increasingly uesad to assist with the interpretation of results
 from large-scale genomics studies. Incorporating information from 
 biological pathways and gene regulatory networks into a genomic data
 analysis is a popular strategy, and there are many methods that provide
 this functionality for gene expression data. When developing or comparing
 such methods, it is important to gain an accurate assessment of their 
-performance, with simulation-based validation studies a popular choice.
+performance. Simulation-based validation studies are frequently used
+for this.
 This necessitates the use of simulated data that correctly accounts for
 pathway relationships and correlations. Here we present a versatile
 statistical framework to simulate correlated gene expression data from
@@ -84,49 +85,50 @@ of genes involved in particular functions, despite the fact that the
 relationships defined by pathway structure could further inform gene
 expression analyses. In many cases, the pathway relationships are
 well-defined, experimentally-validated, and are available in public
-databases [@Reactome]. As a result, network analysis techniques could
+databases [@Reactome]. As a result, network analysis techniques can
 play an important role in furthering our understanding of biological
 pathways and aiding in the interpretation of genomics studies.
 
 Gene networks provide insights into how cells are regulated, by mapping
 regulatory interactions between target genes and transcription factors,
 enhancers, and sites of epigenetic marks or chromatin structures
-[@Barabasi2004; @Yamaguchi2007]. Inference of these regulatory
-interactions for genomics investigations has the potential to radically
+[@Barabasi2004; @Yamaguchi2007]. Inference using these regulatory
+interactions genomic analysis has the potential to radically
 expand the range of candidate biological pathways to be further
 explored, or to improve the accuracy of bioinformatics and functional
-genomic analysis. A number of methods have already been developed to
-utilise timecourse gene expression data [@Arner2015; @Yamaguchi2007]
+genomic analysis. A number of methods have been developed to
+utilise timecourse gene expression data [@Yamaguchi2007; @Arner2015]
 using gene regulatory modules in state-space models and recursive vector
 autoregressive models [@Hirose2008; @Shimamura2009]. Various approaches
 to gene regulation and networks at the genome-wide scale have led to
-novel biological insights [@Arner2015; @Komatsu2013]. However, inference
+novel biological insights [@Arner2015; @Komatsu2013], however, inference
 of regulatory networks has thus far primarily relied on experimental
 validation or resampling-based approaches to estimate the likelihood
 of specific network modules being predicted [@Markowetz2007; @Hawe2019].
 
-Simulated datasets with pathway structure are of particular interest for
-benchmarking regulatory network inference techniques and methods
-being developed for genomics data with complex biological 
[…]
