+++ NAMESPACE
@@ -6,6 +6,7 @@ export(make_adjmatrix_graph)
 export(make_commonlink_adjmat)
 export(make_commonlink_graph)
 export(make_distance_adjmat)
+export(make_distance_comm)
 export(make_distance_graph)
 export(make_distance_laplacian)
 export(make_laplacian_adjmat)
