+++ paper/paper.bib
@@ -1,5 +1,5 @@
 @Article{Arner2015,
-  Author="Arner, E.  and Daub, C. O.  and Vitting-Seerup, K.  and Andersson, R.  and Lilje, B.  and Drabl{\o}s, F.  and Lennartsson, A.  and R{\"o}nnerblad, M.  and Hrydziuszko, O.  and Vitezic, M.  and Freeman, T. C.  and Alhendi, A. M.  and Arner, P.  and Axton, R.  and Baillie, J. K.  and Beckhouse, A.  and Bodega, B.  and Briggs, J.  and Brombacher, F.  and Davis, M.  and Detmar, M.  and Ehrlund, A.  and Endoh, M.  and Eslami, A.  and Fagiolini, M.  and Fairbairn, L.  and Faulkner, G. J.  and Ferrai, C.  and Fisher, M. E.  and Forrester, L.  and Goldowitz, D.  and Guler, R.  and Ha, T.  and Hara, M.  and Herlyn, M.  and Ikawa, T.  and Kai, C.  and Kawamoto, H.  and Khachigian, L. M.  and Klinken, S. P.  and Kojima, S.  and Koseki, H.  and Klein, S.  and Mejhert, N.  and Miyaguchi, K.  and Mizuno, Y.  and Morimoto, M.  and Morris, K. J.  and Mummery, C.  and Nakachi, Y.  and Ogishima, S.  and Okada-Hatakeyama, M.  and Okazaki, Y.  and Orlando, V.  and Ovchinnikov, D.  and Passier, R.  and Patrikakis, M.  and Pombo, A.  and Qin, X. Y.  and Roy, S.  and Sato, H.  and Savvi, S.  and Saxena, A.  and Schwegmann, A.  and Sugiyama, D.  and Swoboda, R.  and Tanaka, H.  and Tomoiu, A.  and Winteringham, L. N.  and Wolvetang, E.  and Yanagi-Mizuochi, C.  and Yoneda, M.  and Zabierowski, S.  and Zhang, P.  and Abugessaisa, I.  and Bertin, N.  and Diehl, A. D.  and Fukuda, S.  and Furuno, M.  and Harshbarger, J.  and Hasegawa, A.  and Hori, F.  and Ishikawa-Kato, S.  and Ishizu, Y.  and Itoh, M.  and Kawashima, T.  and Kojima, M.  and Kondo, N.  and Lizio, M.  and Meehan, T. F.  and Mungall, C. J.  and Murata, M.  and Nishiyori-Sueki, H.  and Sahin, S.  and Nagao-Sato, S.  and Severin, J.  and de Hoon, M. J.  and Kawai, J.  and Kasukawa, T.  and Lassmann, T.  and Suzuki, H.  and Kawaji, H.  and Summers, K. M.  and Wells, C.  and Hume, D. A.  and Forrest, A. R.  and Sandelin, A.  and Carninci, P.  and Hayashizaki, Y. ",
[…]
