### 2020-07-01 f487390
update reference formatting
