+++ paper/paper.bib
@@ -214,7 +214,7 @@ eprint = {/oup/backfile/content_public/journal/bioinformatics/24/7/10.1093_bioin
 }
 
 @Article{Svensson2018,
-   Author={Svensson, V.  and Vento-Tormo, R.  and Teichmann, S. A., {},
+   Author={Svensson, V.  and Vento-Tormo, R.  and Teichmann, S. A., },
    Title="{{E}xponential scaling of single-cell {R}{N}{A}-seq in the past decade}",
    Journal="Nat Protoc",
    Year="2018",
@@ -225,7 +225,7 @@ eprint = {/oup/backfile/content_public/journal/bioinformatics/24/7/10.1093_bioin
 }
 
 @Article{Ozsolak2011,
-   Author={Ozsolak, F.,  and Milos, P. M., },
+   Author={Ozsolak, F., and Milos, P. M.,},
    Title="{{R}{N}{A} sequencing: advances, challenges and opportunities}",
    Journal="Nat. Rev. Genet.",
    Year="2011",
@@ -330,7 +330,7 @@ url="https://doi.org/10.1186/1752-0509-3-41"
 }
 
 @Article{Zappia2017,
-   Title="{S}platter: simulation of single-cell {RNA} sequencing data}",
+   Title="{S}platter: simulation of single-cell {RNA} sequencing data",
    Author={Zappia, L.,  and Phipson, B.,  and Oshlack, A., },
    Journal="Genome Biol.",
    Year="2017",
