### 2020-07-01 #2161 @TomKellyGenetics
@majensen @rcannood @corybrunson

Hi everyone,

I'm pleased to update that I've pushed a substantial revision to the package repository. The main changes are to the documentation (examples and vignettes) and the manuscript. Some examples in the manuscript have been removed (but are available in the Rmarkdown version and vignettes.  Figure placement by Whedon is hardcoded with `\begin{figure}[H]` which explains this issues we had before. This makes the PDF several pages shorter as suggested.

I think this addresses all remaining [issues](https://github.com/TomKellyGenetics/graphsim/issues) raised by the reviewers on the [graphsim](https://github.com/TomKellyGenetics/graphsim) repository. It is now ready for the reviewers to check again. I understand that this may take some time. Sorry that it has taken me longer than I expected to get to this point. I think the package and paper have improved as a result so thank you all for your feedback.

I've submitted updates to [CRAN](https://CRAN.R-project.org/package=graphsim) and [biorXiv](https://www.biorxiv.org/content/10.1101/2020.03.02.972471v1) to reflect these revisions. I look forward to your comments.

Thank you!
