### 2020-07-01 TomKellyGenetics/graphsim#13 @rcannood
Good job @TomKellyGenetics +1 ! I went through the new versions of the article and the documentation on the graphsim GitHub repo. I'm pleased to see that the article is looking really nice. The changes you made adequately address my previous concerns. I will be ticking off the remaining boxes in openjournals/joss-reviews#2161.
