### 2020-07-01 TomKellyGenetics/graphsim#13 @TomKellyGenetics
That's great to hear. Thanks for the quick response!
