### 2020-07-01 TomKellyGenetics/graphsim#11 @TomKellyGenetics
Here is a preview of the rendered [vignettes](https://github.com/TomKellyGenetics/graphsim/files/4856451/vignettes.zip), I'll try get this through CRAN checks/build soon.
