### 2020-07-01 5d6bb00
update for building vignettes and CRAN release
