+++ DESCRIPTION
@@ -21,6 +21,7 @@ Imports:
 Suggests:
     devtools,
     knitr (>= 1.5),
+    markdown,
     prettydoc,
     rmarkdown,
     testthat,
@@ -29,4 +30,4 @@ Suggests:
 LazyData: TRUE
 RoxygenNote: 7.1.1
 Encoding: UTF-8
-VignetteBuilder: knitr
+VignetteBuilder: knitr, rmarkdown
