+++ DESCRIPTION
@@ -1,7 +1,7 @@
 Package: graphsim
 Type: Package
 Title: Simulate Expression Data from 'igraph' Networks 
-Version: 1.0.0
+Version: 1.0.0.9001
 Date: 2020-06-15
 Authors@R: c(person("S. Thomas", "Kelly",email = "tom.kelly@postgrad.otago.ac.nz", role = c("ctb")),
              person("S. Thomas", "Kelly",email = "tom.kelly@riken.jp", role = c("aut", "cre")),
