### 2020-07-01 60e09a1
update for CRAN submission
