### 2020-07-01 a7f1965
pass CRAN checks
