### 2020-07-02 fbf4ea8
update build settings
