### 2020-07-03 TomKellyGenetics/graphsim#11 @corybrunson
The previews look good! The only suggestions i have are

* to consider arranging the figures in grids, or at least putting some pairs on the same line, so that more figures can be viewed on the screen at once; and
* to wrap the recurring analysis steps in a single function (not a package function, just a function defined at the beginning of the vignette) that can be called to produce each set of figures without a large code chunk.

But those are, again, suggestions. I'll keep the issue open until they're installable.
