### 2020-07-03 TomKellyGenetics/graphsim#6 @corybrunson
The documentation is much improved in the current version! Again i have a couple of suggestions, but these need not be met for me to recommend acceptance of the paper.

* While several objects (including functions) include links to related objects in their descriptions, others (for example, the `make_distance_*()` functions) don't include links to other documentation. Two roxygen2 fields that help cross-reference related objects are `@seealso` and `@family`.
* The package itself does not have a documentation page that can be accessed by `help(`graphsim-package`)` or just `help(graphsim)`. This isn't strictly necessary, but i find it very helpful. For an example, see [the package documentation for igraph](https://github.com/igraph/rigraph/blob/dev/R/igraph-package.R)—which is unusually long (in my experience) but showcases the kinds of elements that often make up such a page.
