### 2020-07-03 #2161 @corybrunson
@TomKellyGenetics most everything looks good. I've commented on the remaining open issues i raised, and i the package and paper don't suggest to me that you're making performance claims, so my only remaining issue is that the last link in the paper (in the Computational details section) is broken. Instead of linking to the paper folder in the repo, i'd suggest just linking to the repo.
