### 2020-07-04 903a16d
update computational details https://github.com/openjournals/joss-reviews/issues/2161
