### 2020-07-04 1393de4
update computational details https://github.com/openjournals/joss-reviews/issues/2161
