### 2020-07-04 TomKellyGenetics/graphsim#6 @TomKellyGenetics
Thanks, I had some trouble with links recently actually (see 5bf900f6). There has been some updates to CRAN checks (as discussed here: https://github.com/r-lib/roxygen2/issues/707). Co-incidentally, this was caused by linking to the `help(igraph-package)` because you now have to link to the file name: [aaa-igraph-package](https://github.com/igraph/rigraph/blob/a14fc50abf9bf3ae17b870ccbc21f88d0eefed6e/R/igraph-package.R#L41).

I hadn't considered doing a "help" page for the overall package but I agree this may be helpful to describe the different functions. I'm planning to release a minor update to CRAN soon anyway so I look into including these suggestions (#6) or (#11) for that. This would also update citation information if the paper is accepted.
