### 2020-07-04 #2161 @TomKellyGenetics
Thanks everyone for the quick and helpful responses. I've updated the manuscript to link to the GitHub repo correctly.

I agree to leaving the remaining issues on the repo open until I've had more time to look at the suggestions for documentation in more detail. I will consider these changes for the next release on CRAN.
