### 2020-07-05 #2161 @corybrunson
@TomKellyGenetics that's reasonable! Also, the package installs for me from CRAN, including the vignettes.

@majensen my concerns have been addressed and i have completed my review. : )
