### 2020-07-06 #2161 @majensen
@TomKellyGenetics - we're on the final leg of this journey.
* Can you review the "missing dois" noted in https://github.com/openjournals/joss-reviews/issues/2161#issuecomment-653989928 and add them to the MS?
* Can you create an open archive of the repo, ideally with the same name as that of the paper, and with the same authors, and provide the DOI of that archive here? [Zenodo](https://zenodo.org) is highly suitable for this. There is a brief tutorial at https://github.com/openjournals/joss-reviews/issues/1839#issuecomment-560048729
Thanks for all your work on this paper.
