### 2020-07-06 7b6077b
add DOIs to bibliography https://github.com/openjournals/joss-reviews/issues/2161
