+++ paper/paper.bib
@@ -16,6 +16,7 @@
   Volume = {5},
   Number = {2},
   Pages = {101-13},
+  doi = "{10.1038/nrg1272},
   Keywords = {Animals
     Cell Communication
     Cell Compartmentation
@@ -78,6 +79,7 @@
    Address = "Heidelberg",
    ISBN = "978-3-642-01688-2", 
    Year = "2009",
+   doi = {10.1007/978-3-642-01689-9},
    }
 
 @Manual{mvtnorm,
@@ -94,7 +96,8 @@
    Journal="Front Genet",
    Year="2019",
    Volume="10",
-   Pages="535"
+   Pages="535",
+   doi = {10.3389/fgene.2019.00535}.
 }
 
 @article{Higham2002,
@@ -132,7 +135,8 @@ eprint = {/oup/backfile/content_public/journal/bioinformatics/24/7/10.1093_bioin
    Volume="17",
    Number="10",
    Pages="615--629",
-   Month="10"
+   Month="10",
+   doi = {10.1038/nrg.2016.87},
 }
 
 
@@ -144,7 +148,8 @@ eprint = {/oup/backfile/content_public/journal/bioinformatics/24/7/10.1093_bioin
    Volume="42",
    Number="2",
    Pages="478--506",
-   Month="Feb"
+   Month="Feb",
+   doi = {10.3892/ijo.2012.1744},
 }
 
 
@@ -168,7 +173,8 @@ eprint = {/oup/backfile/content_public/journal/bioinformatics/24/7/10.1093_bioin
    Year="2015",
    Volume="16",
    Pages="347",
-   Month="Oct"
+   Month="Oct",
+   doi = {10.1186/s12859-015-0778-7},
 }
 
 @Article{Markowetz2017,
@@ -189,7 +195,8 @@ eprint = {/oup/backfile/content_public/journal/bioinformatics/24/7/10.1093_bioin
    Year="2007",
    Volume="8 Suppl 6",
    Pages="S5",
-   Month="Sep"
+   Month="Sep",
+   doi = {10.1186/1471-2105-8-s6-s5},
 }
 
 @Article{Perou2000,
@@ -211,6 +218,7 @@ eprint = {/oup/backfile/content_public/journal/bioinformatics/24/7/10.1093_bioin
   volume = {43},
   number = {7},
   pages = {e47},
+  doi = {10.1093/nar/gkv007},
 }
 
 @Article{Svensson2018,
@@ -221,7 +229,8 @@ eprint = {/oup/backfile/content_public/journal/bioinformatics/24/7/10.1093_bioin
    Volume="13",
    Number="4",
    Pages="599--604",
-   Month="04"
+   Month="04",
+   doi = {10.1038/nprot.2017.149},
 }
 
 @Article{Ozsolak2011,
@@ -232,7 +241,8 @@ eprint = {/oup/backfile/content_public/journal/bioinformatics/24/7/10.1093_bioin
    Volume="12",
    Number="2",
    Pages="87--98",
-   Month="Feb"
+   Month="Feb",
+   doi = {10.1038/nrg2934},
 }
 
 @article {Cannoodt2020,
@@ -315,7 +325,8 @@ url="https://doi.org/10.1186/1752-0509-3-41"
    Volume="115",
    Number="28",
    Pages="E6437-E6446",
-   Month="07"
+   Month="07",
+   doi = {10.1101/227033},
 }
 
 @Article{Yamaguchi2007,
@@ -326,7 +337,8 @@ url="https://doi.org/10.1186/1752-0509-3-41"
    Volume="24",
    Number="1",
    Pages="37--46",
-   Month="Jan"
+   Month="Jan",
[…]
