### 2020-07-06 cddd1f2
add DOIs to bibliography https://github.com/openjournals/joss-reviews/issues/2161
