### 2020-07-06 #2161 @TomKellyGenetics
@majensen Thanks, that's great to hear! I've updated `paper.bib` to include the DOIs.

For the "archive" version, is a GitHub "release" okay? I've already tagged [version 1.0.0](https://github.com/TomKellyGenetics/graphsim/releases/tag/v1.0.0) which was released on CRAN. I've integrated Zenodo with package so a DOI: [10.5281/zenodo.3926914](https://zenodo.org/record/3926913#.XwLUOZMzbBU) has already been generated automatically. Is this appropriate or is a dedicated tar.gz with the paper name (as done [here](https://zenodo.org/record/3558829#.XwLVY5MzZTY)) preferable?
