### 2020-07-06 #2161 @TomKellyGenetics
A new version has been released on Zenodo DOI [105281/zenodo.3931288](https://zenodo.org/record/3931288#.XwLfT5MzaCU) to correspond to the manuscript. This version has `zip` and `tar.gz` formats available for all files relevant to the paper and software release. This version is tagged with `1.0..0-joss`, the software is the same as the CRAN release with minor tweaks to the paper discussed above. Hope that does the trick.
