### 2020-07-06 #2161 @arfon
Looks like there's a typo in the BibTeX here: https://github.com/TomKellyGenetics/graphsim/blob/master/paper/paper.bib#L100 (the period needs removing).

I'm going to look at adding better error handling for the BibTeX parser in Whedon.
