### 2020-07-06 #2161 @majensen
@arfon @danielskatz thanks for having a look - @TomKellyGenetics, can you have a look at the bibtex file per https://github.com/openjournals/joss-reviews/issues/2161#issuecomment-654430778 ?
