### 2020-07-06 #2161 @danielskatz
the author field appears to be wrong - it should be
"Hawe, J. S. and Theis, F. J. and Heinig, M."
