### 2020-07-06 #2161 @arfon
Also, there's an issue with ` doi = {10.3389/fgene.2019.00535}.` <- the period needs removing on this line.
