+++ paper/paper.bib
@@ -97,7 +97,7 @@
    Year="2019",
    Volume="10",
    Pages="535",
-   doi = {10.3389/fgene.2019.00535}.
+   doi = {10.3389/fgene.2019.00535},
 }
 
 @article{Higham2002,
