+++ paper/paper.bib
@@ -16,7 +16,7 @@
   Volume = {5},
   Number = {2},
   Pages = {101-13},
-  doi = "{10.1038/nrg1272},
+  doi = {10.1038/nrg1272},
   Keywords = {Animals
     Cell Communication
     Cell Compartmentation
