### 2020-07-07 #2161 @TomKellyGenetics
Typos in references corrected. Just pushed updated PDF of figures. @majensen PRs welcome :-)
