### 2020-07-07 #2161 @majensen
Hi @TomKellyGenetics - well, it's a very clearly written paper, and I enjoyed going through it again. The only tiny thing I would have done is remove the second comma in this snippet at (raw) line 137 in paper.md:

`studies, as well as, bulk`

i.e., "studies, as well as bulk...".

Not worth a PR, IMO. I will see if I can cause the editor-in-chief to manifest him/herself.
