### 2020-07-07 #2161 @TomKellyGenetics
@majensen Thanks, I agree to this change so I've pushed a correction.

> i.e., "studies, as well as bulk...".

I think I've got this in time to be compiled in the final submission.
