### 2020-07-08 bbcd8b7
disable processing large graph in vignettes (slow CRAN checks)
