### 2020-07-08 #2161 @kyleniemeyer
Hi @TomKellyGenetics, before accepting I found just a few things in the paper to fix:

- typo in third sentence: "uesad" -> "used"
- the `limma` reference has a a typo in the bib file, I think there is a space missing between a comma and `and` causing a problem:
```
{Matthew E R., and Phipson, B.,and Wu, D.,  and Hu, Y.,  and Law, C.W.,  and Shi, W., and Smyth, G.K.,}
```
should be
```
{Matthew E R., and Phipson, B., and Wu, D., and Hu, Y., and Law, C.W., and Shi, W., and Smyth, G.K.,}
```
