### 2020-07-09 cf8479d
copyediting paper https://github.com/openjournals/joss-reviews/issues/2161
