+++ paper/paper.bib
@@ -211,7 +211,7 @@ eprint = {/oup/backfile/content_public/journal/bioinformatics/24/7/10.1093_bioin
 }
 
 @article{limma,
-  author = {Matthew E R., and Phipson, B.,and Wu, D.,  and Hu, Y.,  and Law, C.W.,  and Shi, W., and Smyth, G.K.,},
+  author = {Matthew E R., and Phipson, B., and Wu, D.,  and Hu, Y.,  and Law, C.W.,  and Shi, W., and Smyth, G.K.,},
   title = {{limma} powers differential expression analyses for {RNA}-sequencing and microarray studies},
   journal = {Nucleic Acids Research},
   year = {2015},
