### 2020-07-09 #2161 @majensen
Really nice paper @TomKellyGenetics - @corybrunson and @rcannood I appreciate your huge contribution during these strange days. Hope you will all think of JOSS next time you publish!
--@majensen
