### 2020-07-09 #2161 @TomKellyGenetics
Thanks again, @majensen @corybrunson and @rcannood, I'm very impressed by the JOSS review process. This a great initiative for the open-source community. I have put my name in the hat to review. 😊
